### 2025-05-20 6e0adc9
feat(pin_color): add pin color feature to visualize nearest neighbor and helpers and unit test files individually
