+++ src/spac/visualization.py
@@ -26,6 +26,9 @@ import time
 import json
 import re
 from typing import Dict, List, Union
+import matplotlib.colors as mcolors
+import matplotlib.patches as mpatch
+from functools import partial
 
 
 # Configure logging
@@ -1613,28 +1616,17 @@ def boxplot_interactive(
         DPI (dots per inch) for the figure. Default is 200.
 
     defined_color_map : str, optional
-        Predefined color mapping stored in adata.uns for specific labels.
-        Default is None, which will generate the color mapping automatically.
-
-    annotation_colorscale : str, default='viridis'
-        Name of the color scale to use for the dots when annotation
-        is used.
-
-    feature_colorscale: str, default='seismic'
-        Name of the color scale to use for the dots when feature
-        is used.
-
-    figure_type : {"interactive", "static", "png"}, default = "interactive"
-        If "interactive", the plot is interactive, allowing for zooming
-        and panning.
-        If "static", the plot is static.
-        If "png", the plot is returned as a PNG image.
-
-    return_metrics: bool, default = False
-        If True, the function also returns the computed boxplot metrics.
-
-    **kwargs : additional keyword arguments
-        Any other keyword arguments passed to the underlying plotting function.
+        Key in 'adata.uns' holding a pre-computed color dictionary.
+        Falls back to automatic generation from 'annotation' values.
+    ax : matplotlib.axes.Axes, optional
+        A Matplotlib Axes object. Currently, this parameter is not used by the
+        underlying plotting functions (Seaborn's `catplot`/`displot`), which
+        will always generate a new figure and axes. The `ax` key in the
+        returned dictionary will contain the Axes from these new plots.
+        This parameter is maintained for API consistency and potential
+        future enhancements. Default is None.
+    **kwargs : dict
+        Additional arguments for seaborn figure-level functions.
 
     Returns
     -------
@@ -3210,190 +3202,242 @@ def _plot_spatial_distance_dispatch(
     plot_type,
     stratify_by=None,
     facet_plot=False,
-    **kwargs
+    distance_col="distance",
+    hue_axis="group",
+    palette=None,
+    **kwargs,
 ):
     """
-    Decides the figure layout based on 'stratify_by' and 'facet_plot'
-    and dispatches actual plotting calls.
+    Dispatch a seaborn call to visualise nearest-neighbor distances.
+    Returns Axes object(s) for further customization.
 
-    Logic:
-      1) If stratify_by and facet_plot => single figure with subplots (faceted)
-      2) If stratify_by and not facet_plot => multiple figures, one per group
-      3) If stratify_by is None => single figure (no subplots)
-
-    This function calls seaborn figure-level functions (catplot or displot).
+    Layout logic
+    ------------
+    1. ``stratify_by`` & ``facet_plot`` → Faceted plot, returns ``Axes``
+       or ``List[Axes]`` for the "ax" key.
+    2. ``stratify_by`` & not ``facet_plot`` → List of plots, returns
+       ``List[Axes]`` for the "ax" key.
+    3. ``stratify_by`` is None → Single plot, returns ``Axes`` or
+       ``List[Axes]`` (if plot_type creates facets) for the "ax" key.
 
     Parameters
     ----------
     df_long : pd.DataFrame
-        Tidy DataFrame with columns ['cellid', 'group', 'distance',
+        Tidy DataFrame returned by `_prepare_spatial_distance_data` with
+        a long layout and with  columns ['cellid', 'group', 'distance',
         'phenotype', 'stratify_by'].
     method : {'numeric', 'distribution'}
[…]
