### 2025-05-20 22e047a
feat(visualize_nn): add pin-color feature and enhance layout of fig and ax
