+++ tests/test_visualization/test_visualize_nearest_neighbor.py
@@ -1,6 +1,3 @@
-import os
-import sys
-sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../../src")
 import unittest
 import pandas as pd
 import numpy as np
