+++ CHANGELOG.md
@@ -1,42 +1,19 @@
 # CHANGELOG
 
 
-## v0.8.9 (2025-05-13)
-
-### Bug Fixes
-
-- **knn clustering**: Fixed conflict in imports
-  ([`e6d5f1e`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/e6d5f1eb4668d8e25ed09469525e128abcabbb6a))
-
-- **knn_clustering**: Fixed format of the error message
-  ([`ddf92e6`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/ddf92e63c760c56e035071c0edc2ce245aa3fb7c))
-
-- **select_values**: Added support when observation are numerical
-  ([`4bcbaa2`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/4bcbaa249e70c51860ae55356c5b6ab2bf8961bc))
-
-- **summarize_dataframe**: Remove duplicated missing index
-  ([`2c0d907`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/2c0d9077ea9414172090d8c10488f75360430e10))
-
-### Code Style
-
-- **knn_clustering**: Adjusted to code style standar
-  ([`2b02538`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/2b02538d4d15a748ce399eb3837dc51eb99108f6))
+## v0.8.10 (2025-05-20)
 
 ### Features
 
-- **visualize_nearest_neighbor**: Add pin-color and corresponding unit test
-  ([`015692f`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/015692f87e75e5c4581638dedd4d277c63c6ddec))
+- **pin_color**: Add pin color feature to visualize nearest neighbor and helpers and unit test files
+  individually
+  ([`6e0adc9`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/6e0adc9aaf63c443c615e27dd3d6a6ded4a4b391))
 
-### Testing
-
-- **comments**: Add extensive comments for complex data set generation in utag tests
-  ([`ef95276`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/ef952769f8d20310903f8a9269772a67cb1057d4))
-
-- **comments**: Add extensive comments for complex data set generation in utag tests
-  ([`cd8cb25`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/cd8cb25764a3a5f4718692dea3585bae17f88105))
+- **visualize_nn**: Add pin-color feature and enhance layout of fig and ax
+  ([`22e047a`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/22e047aa3d765469c8d0acd32f8ea3d8be5b27bd))
 
 
-## v0.8.8 (2025-04-06)
+## v0.8.9 (2025-05-13)
 
 ### Bug Fixes
 
@@ -58,11 +35,23 @@
 - **interactive_spatial_plot**: Fixed typos in api name and arguments
   ([`b833123`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/b833123dc9f124abf34e539be553a8388048ee1b))
 
+- **knn clustering**: Fixed conflict in imports
+  ([`e6d5f1e`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/e6d5f1eb4668d8e25ed09469525e128abcabbb6a))
+
+- **knn_clustering**: Fixed format of the error message
+  ([`ddf92e6`](https://github.com/FNLCR-DMAP/SCSAWorkflow/commit/ddf92e63c760c56e035071c0edc2ce245aa3fb7c))
+
[…]
