+++ README.md
@@ -309,3 +309,7 @@ Datetime: 2025-04-06 16:13:04, Version:  spac-dev-0.8.8-py39_0.tar.bz2
 Datetime: 2025-05-13 12:33:10, Version:  spac-dev-0.8.9-py39_0.tar.bz2
 <br>[View Commit Log](Commit_Log_spac-dev-0.8.9-py39_0.log)
 <br>Deployment Status: Success
+
+Datetime: 2025-05-20 06:09:36, Version:  spac-dev-0.8.10-py39_0.tar.bz2
+<br>[View Commit Log](Commit_Log_spac-dev-0.8.10-py39_0.log)
+<br>Deployment Status: Success
