### 2025-05-22 4b49fac
fix(visualize_nearest_neighbor): rewrite unit tests and move up  library in function
