+++ tests/test_visualization/test_visualize_nearest_neighbor.py
@@ -3,6 +3,7 @@ import pandas as pd
 import numpy as np
 import anndata
 import matplotlib
+import matplotlib.collections as mcoll
 matplotlib.use('Agg')  # Uses a non-interactive backend for tests
 import matplotlib.pyplot as plt
 from spac.visualization import visualize_nearest_neighbor
@@ -12,7 +13,9 @@ class TestVisualizeNearestNeighbor(unittest.TestCase):
 
     @staticmethod
     def _create_test_adata():
-        """Creates a common AnnData object for testing various scenarios."""
+        """
+        Creates a common AnnData object for testing various scenarios.
+        """
         data = np.array([[1.0, 2.0], [3.0, 4.0], [5.0, 6.0], [7.0, 8.0]])
         obs = pd.DataFrame({
             'cell_type': ['typeA', 'typeB', 'typeA', 'typeC'],
@@ -37,10 +40,37 @@ class TestVisualizeNearestNeighbor(unittest.TestCase):
         """Close all Matplotlib figures after each test."""
         plt.close('all')
 
+    def test_axis_labels_across_plot_types(self):
+        """
+        X/Y labels are correct for each plotting mode (no duplication).
+        """
+        scenarios = [
+            # explicit plot_type
+            ('numeric', 'box', 'group'),
+            ('distribution', 'kde', 'Density'),
+            # default plot_type (None)
+            ('numeric', None, 'group'),
+            ('distribution', None, 'Density'),
+        ]
+        for method, plot_type, expected_ylabel in scenarios:
+            with self.subTest(method=method, plot_type=plot_type):
+                result = visualize_nearest_neighbor(
+                    adata=self.adata,
+                    annotation='cell_type',
+                    distance_from='typeA',
+                    distance_to='typeB',
+                    method=method,
+                    plot_type=plot_type,
+                )
+                ax = (result['ax'] if method == 'numeric'
+                      else result['fig'].axes[0])
+                self.assertEqual(
+                    ax.get_xlabel(), "Nearest Neighbor Distance"
+                )
+                self.assertEqual(ax.get_ylabel(), expected_ylabel)
+
     def test_output_structure_and_types_single_plot(self):
-        """
-        Tests the basic output structure and types for a single plot.
-        """
+        """Dict keys and object types for a simple numeric plot."""
         result = visualize_nearest_neighbor(
             adata=self.adata,
             annotation='cell_type',
@@ -49,44 +79,19 @@ class TestVisualizeNearestNeighbor(unittest.TestCase):
             method='numeric',
             plot_type='box'
         )
-
-        self.assertIsInstance(result, dict, "Result should be a dict.")
         expected_keys = ['data', 'fig', 'ax', 'palette']
         for key in expected_keys:
-            self.assertIn(key, result, f"Key '{key}' missing in result.")
+            self.assertIn(key, result)
+        self.assertIsInstance(result['data'], pd.DataFrame)
+        self.assertIsInstance(result['fig'], matplotlib.figure.Figure)
+        self.assertIsInstance(result['ax'], matplotlib.axes.Axes)
+        self.assertIsInstance(result['palette'], dict)
+        self.assertEqual(len(result['fig'].axes), 1)
+        self.assertIs(result['ax'], result['fig'].axes[0])
+        self.assertIs(result['ax'].figure, result['fig'])
 
-        self.assertIsInstance(
-            result['data'], pd.DataFrame, "'data' should be a DataFrame."
-        )
-        self.assertIsInstance(
-            result['fig'], matplotlib.figure.Figure,
-            "'fig' should be a Figure."
-        )
-        self.assertIsInstance(
-            result['ax'], matplotlib.axes.Axes,
-            "'ax' should be an Axes object for a single plot."
-        )
-        self.assertIsInstance(
[…]
