### 2025-05-23 fa20694
fix(visualize_nearest_neighbor): add comments on function and refactor unit tests
