+++ tests/test_visualization/test_visualize_nearest_neighbor.py
@@ -1,3 +1,6 @@
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../../src")
 import unittest
 import pandas as pd
 import numpy as np
@@ -40,37 +43,10 @@ class TestVisualizeNearestNeighbor(unittest.TestCase):
         """Close all Matplotlib figures after each test."""
         plt.close('all')
 
-    def test_axis_labels_across_plot_types(self):
-        """
-        X/Y labels are correct for each plotting mode (no duplication).
-        """
-        scenarios = [
-            # explicit plot_type
-            ('numeric', 'box', 'group'),
-            ('distribution', 'kde', 'Density'),
-            # default plot_type (None)
-            ('numeric', None, 'group'),
-            ('distribution', None, 'Density'),
-        ]
-        for method, plot_type, expected_ylabel in scenarios:
-            with self.subTest(method=method, plot_type=plot_type):
-                result = visualize_nearest_neighbor(
-                    adata=self.adata,
-                    annotation='cell_type',
-                    distance_from='typeA',
-                    distance_to='typeB',
-                    method=method,
-                    plot_type=plot_type,
-                )
-                ax = (result['ax'] if method == 'numeric'
-                      else result['fig'].axes[0])
-                self.assertEqual(
-                    ax.get_xlabel(), "Nearest Neighbor Distance"
-                )
-                self.assertEqual(ax.get_ylabel(), expected_ylabel)
-
     def test_output_structure_and_types_single_plot(self):
-        """Dict keys and object types for a simple numeric plot."""
+        """
+        Tests the basic output structure and types for a single plot.
+        """
         result = visualize_nearest_neighbor(
             adata=self.adata,
             annotation='cell_type',
@@ -79,19 +55,44 @@ class TestVisualizeNearestNeighbor(unittest.TestCase):
             method='numeric',
             plot_type='box'
         )
+
+        self.assertIsInstance(result, dict, "Result should be a dict.")
         expected_keys = ['data', 'fig', 'ax', 'palette']
         for key in expected_keys:
-            self.assertIn(key, result)
-        self.assertIsInstance(result['data'], pd.DataFrame)
-        self.assertIsInstance(result['fig'], matplotlib.figure.Figure)
-        self.assertIsInstance(result['ax'], matplotlib.axes.Axes)
-        self.assertIsInstance(result['palette'], dict)
-        self.assertEqual(len(result['fig'].axes), 1)
-        self.assertIs(result['ax'], result['fig'].axes[0])
-        self.assertIs(result['ax'].figure, result['fig'])
+            self.assertIn(key, result, f"Key '{key}' missing in result.")
 
-    def test_minimal_numeric_palette(self):
-        """ Palette filtered to target groups."""
+        self.assertIsInstance(
+            result['data'], pd.DataFrame, "'data' should be a DataFrame."
+        )
+        self.assertIsInstance(
+            result['fig'], matplotlib.figure.Figure,
+            "'fig' should be a Figure."
+        )
+        self.assertIsInstance(
+            result['ax'], matplotlib.axes.Axes,
+            "'ax' should be an Axes object for a single plot."
+        )
+        self.assertIsInstance(
+            result['palette'], dict, "'palette' should be a dictionary."
+        )
+
+        self.assertEqual(
+            len(result['fig'].axes), 1,
+            "Single plot figure should contain one axis."
+        )
+        self.assertIs(
+            result['ax'], result['fig'].axes[0],
+            "Returned 'ax' should be the axis in 'fig'."
+        )
+        self.assertIs(
+            result['ax'].figure, result['fig'],
+            "Returned 'ax' should belong to returned 'fig'."
+        )
+
[…]
