+++ README.md
@@ -317,3 +317,7 @@ Datetime: 2025-05-20 06:09:36, Version:  spac-dev-0.8.10-py39_0.tar.bz2
 Datetime: 2025-05-23 18:41:57, Version:  spac-dev-0.8.11-py39_0.tar.bz2
 <br>[View Commit Log](Commit_Log_spac-dev-0.8.11-py39_0.log)
 <br>Deployment Status: Success
+
+Datetime: 2025-05-23 19:58:14, Version:  spac-dev-0.9.0-py39_0.tar.bz2
+<br>[View Commit Log](Commit_Log_spac-dev-0.9.0-py39_0.log)
+<br>Deployment Status: Success
