+++ docs/source/modules/transformations._validate_transformation_inputs.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.transformations._validate_transformation_inputs
