+++ docs/source/modules/utag_functions.low_variance_filter.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.utag_functions.low_variance_filter
