+++ docs/source/modules/spatial_analysis._process_matrixs.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.spatial_analysis._process_matrixs
