+++ docs/source/modules/utag_functions.rst
@@ -1,0 +1,13 @@
+.. automodule:: spac.utag_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :autosummary:
+
+spac.utag_functions.z_score
+spac.utag_functions.sparse_matrix_dstack
+spac.utag_functions._parallel_message_pass
+spac.utag_functions.custom_message_passing
+spac.utag_functions.low_variance_filter
+spac.utag_functions.add_probabilities_to_centroid
+spac.utag_functions.utag
