+++ docs/source/index.rst
@@ -1,0 +1,16 @@
+.. mdinclude:: ../../README.md
+
+.. toctree::
+   :maxdepth: 3
+
+   self
+   modules
+
+   modules/visualization
+   modules/transformations
+   modules/_ripley
+   modules/data_utils
+   modules/spatial_analysis
+   modules/phenotyping
+   modules/utag_functions
+   modules/utils
