+++ docs/source/modules/visualization._prepare_spatial_distance_data.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization._prepare_spatial_distance_data
