+++ docs/source/modules/utag_functions.z_score.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.utag_functions.z_score
