+++ docs/source/modules/data_utils.combine_dfs.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.data_utils.combine_dfs
