+++ docs/source/modules/visualization.prepare_spatial_dataframe.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.prepare_spatial_dataframe
