+++ docs/source/modules/visualization._make_axes_object.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization._make_axes_object
