+++ docs/_sources/index.rst.txt
@@ -1,0 +1,16 @@
+.. mdinclude:: ../../README.md
+
+.. toctree::
+   :maxdepth: 3
+
+   self
+   modules
+
+   modules/phenotyping
+   modules/data_utils
+   modules/visualization
+   modules/spatial_analysis
+   modules/transformations
+   modules/_ripley
+   modules/utag_functions
+   modules/utils
