+++ docs/source/modules/utils.check_distances.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.utils.check_distances
