+++ docs/source/modules/spatial_analysis.rst
@@ -1,0 +1,19 @@
+.. automodule:: spac.spatial_analysis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :autosummary:
+
+spac.spatial_analysis.spatial_interaction
+spac.spatial_analysis._Neighborhood_Enrichment_Analysis
+spac.spatial_analysis._Cluster_Interaction_Matrix_Analysis
+spac.spatial_analysis._perform_analysis
+spac.spatial_analysis._get_labels
+spac.spatial_analysis._process_matrixs
+spac.spatial_analysis._processes_function_return
+spac.spatial_analysis.ripley_l
+spac.spatial_analysis.process_region
+spac.spatial_analysis.append_results
+spac.spatial_analysis.neighborhood_profile
+spac.spatial_analysis._neighborhood_profile_core
+spac.spatial_analysis.calculate_nearest_neighbor
