+++ docs/_sources/modules/spatial_analysis.process_region.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.spatial_analysis.process_region
