+++ docs/source/modules/visualization.calculate_histogram.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.calculate_histogram
