+++ docs/_sources/modules.rst.txt
@@ -1,0 +1,7 @@
+spac
+====
+
+.. toctree::
+   :maxdepth: 4
+
+   spac
