+++ docs/source/modules/visualization.interactive_spatial_plot.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.interactive_spatial_plot
