+++ docs/_sources/modules/visualization.main_figure_generation.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.main_figure_generation
