+++ docs/source/modules/visualization.cal_bin_num.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.cal_bin_num
