+++ docs/source/spac.rst
@@ -1,0 +1,69 @@
+spac package
+============
+
+Submodules
+----------
+
+spac.data\_utils module
+-----------------------
+
+.. automodule:: spac.data_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+spac.phenotyping module
+-----------------------
+
+.. automodule:: spac.phenotyping
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+spac.spatial\_analysis module
+-----------------------------
+
+.. automodule:: spac.spatial_analysis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+spac.transformations module
+---------------------------
+
+.. automodule:: spac.transformations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+spac.utag\_functions module
+---------------------------
+
+.. automodule:: spac.utag_functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+spac.utils module
+-----------------
+
+.. automodule:: spac.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+spac.visualization module
+-------------------------
+
+.. automodule:: spac.visualization
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: spac
+   :members:
+   :undoc-members:
+   :show-inheritance:
