+++ docs/_sources/modules/utils.annotation_category_relations.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.utils.annotation_category_relations
