+++ docs/_sources/modules/_ripley._ppp.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac._ripley._ppp
