+++ docs/_sources/modules/phenotyping.assign_manual_phenotypes.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.phenotyping.assign_manual_phenotypes
