+++ docs/_sources/modules/_ripley._l_function.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac._ripley._l_function
