+++ docs/_sources/modules/transformations.tsne.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.transformations.tsne
