+++ docs/_sources/modules/spatial_analysis._perform_analysis.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.spatial_analysis._perform_analysis
