+++ docs/_sources/modules/data_utils._select_values_dataframe.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.data_utils._select_values_dataframe
