+++ docs/_sources/modules/_ripley.rst.txt
@@ -1,0 +1,13 @@
+.. automodule:: spac._ripley
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :autosummary:
+
+spac._ripley.ripley
+spac._ripley._reshape_res
+spac._ripley._f_g_function
+spac._ripley._l_function
+spac._ripley._l_multiple_function
+spac._ripley._l_multiple_function
+spac._ripley._ppp
