### 2025-05-29 520934e
feat(downsample):anndata.AnnData input for downsample_cells
Modified downsample_cells function to accept anndata.AnnData objects as input. When an AnnData object is provided, .X` and .obs data are combined into a pandas DataFrame, before applying the rest of the downsampling function.
