+++ tests/test_data_utils/test_downsample_cells.py
@@ -185,6 +185,61 @@ class TestDownsampleCells(unittest.TestCase):
         )
         self.assertDictEqual(expected_counts, actual_counts)
 
+    def test_anndata_input(self):
+        """
+        Test that downsample_cells accepts anndata objects as input.
+        Check that it merges anndata .X and .obs into a dataframe 
+        and performs downsampling correctly.
+        """
+        
+        # create anndata object
+        X_data = pd.DataFrame({
+            'feature1': [1, 3, 5, 7, 9, 12, 14, 16],
+            'feature2': [2, 4, 6, 8, 10, 13, 15, 18],
+            'feature3': [3, 5, 7, 9, 11, 14, 16, 19]
+        })
+
+        obs_data = pd.DataFrame({
+            'phenotype': [
+                'phenotype1',
+                'phenotype1',
+                'phenotype2',
+                'phenotype2',
+                'phenotype3',
+                'phenotype3',
+                'phenotype4',
+                'phenotype4'
+            ]
+        })
+
+        anndata_obj = anndata.AnnData(X = X_data, obs = obs_data)
+            
+        # call downsample on the anndata object
+        downsampled_df = downsample_cells(
+            input_data = anndata_obj,
+            annotations = 'phenotype',
+            n_samples = 1,
+            stratify = False,
+            rand = True,
+            combined_col_name= '_combined_',
+            min_threshold= 5
+        )
+        
+        # confirm the downsampled_df is a pandas dataframe
+        self.assertTrue(isinstance(downsampled_df, pd.DataFrame))
+
+        # confirm number of samples after downsampling is correct 
+        # (four groups with one sample each is four rows total)
+        self.assertTrue(len(downsampled_df) == 4)		
+
+        # confirm the number of groups (phenotypes) is still four 
+        self.assertTrue(downsampled_df['phenotype'].nunique() == 4)
+
+        # confirm original annotation column & feature columns are present
+        expected_feature_columns = X_data.columns.tolist()
+        for col in expected_feature_columns:
+            self.assertIn(col, downsampled_df.columns)
+        self.assertIn('phenotype', downsampled_df.columns)
 
 if __name__ == '__main__':
     unittest.main()
