### 2025-05-29 ecb3940
addittional formatting of histogram function
