+++ tests/test_visualization/test_datashader.py
@@ -6,6 +6,7 @@ import matplotlib
 
 matplotlib.use('Agg')  # Set the backend to 'Agg' to suppress plot window
 
+
 class TestDataShaderHeatMap(unittest.TestCase):
     def setUp(self):
         """Prepare data for testing."""
@@ -19,24 +20,27 @@ class TestDataShaderHeatMap(unittest.TestCase):
         """Test handling of invalid input types."""
         with self.assertRaises(ValueError) as context_manager:
             heatmap_datashader(1, self.y, labels=self.labels_categorical)
-        self.assertIn("x and y must be array-like", str(context_manager.exception))
+        self.assertIn("x and y must be array-like", 
+                      str(context_manager.exception))
 
     def test_labels_length_mismatch(self):
         """Test handling of mismatched lengths between data and labels."""
         wrong_labels = pd.Series(['A'] * 9)  # Shorter than x and y
         with self.assertRaises(ValueError) as context_manager:
             heatmap_datashader(self.x, self.y, labels=wrong_labels)
-        self.assertIn("Labels length should match x and y length", str(context_manager.exception))
+        self.assertIn("Labels length should match x and y length", 
+                      str(context_manager.exception))
 
     def test_valid_input_returns_figure_basic(self):
         """Test that valid input returns a matplotlib figure with expected subplots."""
-        fig = heatmap_datashader(self.x, self.y, labels=self.labels_categorical)
+        fig = heatmap_datashader(self.x, self.y, 
+                                 labels=self.labels_categorical)
         self.assertIsInstance(fig, matplotlib.figure.Figure)
 
         num_axes = len(fig.axes)
         expected_axes = self.labels_categorical.nunique()
         self.assertEqual(num_axes, expected_axes)
-        
+
     def test_labels_not_multiple_of_three(self):
         """Test heatmap generation when the number of labels is not a multiple of 3."""
         x = np.random.rand(7)
@@ -51,12 +55,15 @@ class TestDataShaderHeatMap(unittest.TestCase):
         self.assertEqual(num_axes, expected_axes)
 
         for ax in fig.axes:
-            images = [child for child in ax.get_children() if isinstance(child, matplotlib.image.AxesImage)]
-            self.assertGreater(len(images), 0, "Expected at least one image in each subplot.")
+            images = [child for child in ax.get_children() 
+                      if isinstance(child, matplotlib.image.AxesImage)]
+            self.assertGreater(len(images), 0, 
+                               "Expected at least one image in each subplot.")
 
     def test_valid_input_returns_figure(self):
         """Test that valid input returns a matplotlib figure with expected subplots and images."""
-        fig = heatmap_datashader(self.x, self.y, labels=self.labels_categorical)
+        fig = heatmap_datashader(self.x, self.y, 
+                                 labels=self.labels_categorical)
         self.assertIsInstance(fig, matplotlib.figure.Figure)
 
         # Check number of axes matches number of unique labels
@@ -66,10 +73,10 @@ class TestDataShaderHeatMap(unittest.TestCase):
 
         # Check that each axis has an image plotted
         for ax in fig.axes:
-            images = [child for child in ax.get_children() if isinstance(child, matplotlib.image.AxesImage)]
-            self.assertGreater(len(images), 0, "Expected at least one image in each subplot.")
-
-
+            images = [child for child in ax.get_children() 
+                      if isinstance(child, matplotlib.image.AxesImage)]
+            self.assertGreater(len(images), 0, 
+                               "Expected at least one image in each subplot.")
 
 if __name__ == "__main__":
     unittest.main()
