+++ docs/source/conf.py
@@ -75,3 +75,118 @@ html_theme_options = {
     'titles_only': False,
     'style_external_links': True,
 }
+import os
+import os
+import sys
+path = os.path.abspath('../../src')
+sys.path.insert(0,path)
+extensions = [
+	'sphinx.ext.napoleon',
+	'sphinx.ext.autodoc',
+	'sphinx.ext.autosectionlabel',
+	'sphinx.ext.todo',
+	'sphinx.ext.viewcode',
+	'sphinx.ext.githubpages',
+	'sphinx.ext.autosummary',
+	'm2r']
+autosummary_generate = True
+source_suffix = ['.rst', '.md']
+html_theme_options = {
+    'collapse_navigation': False,
+    'navigation_depth': 3,
+    'sticky_navigation': True,
+    'titles_only': False,
+    'style_external_links': True,
+}
+import os
+import os
+import sys
+path = os.path.abspath('../../src')
+sys.path.insert(0,path)
+extensions = [
+	'sphinx.ext.napoleon',
+	'sphinx.ext.autodoc',
+	'sphinx.ext.autosectionlabel',
+	'sphinx.ext.todo',
+	'sphinx.ext.viewcode',
+	'sphinx.ext.githubpages',
+	'sphinx.ext.autosummary',
+	'm2r']
+autosummary_generate = True
+source_suffix = ['.rst', '.md']
+html_theme_options = {
+    'collapse_navigation': False,
+    'navigation_depth': 3,
+    'sticky_navigation': True,
+    'titles_only': False,
+    'style_external_links': True,
+}
+import os
+import os
+import sys
+path = os.path.abspath('../../src')
+sys.path.insert(0,path)
+extensions = [
+	'sphinx.ext.napoleon',
+	'sphinx.ext.autodoc',
+	'sphinx.ext.autosectionlabel',
+	'sphinx.ext.todo',
+	'sphinx.ext.viewcode',
+	'sphinx.ext.githubpages',
+	'sphinx.ext.autosummary',
+	'm2r']
+autosummary_generate = True
+source_suffix = ['.rst', '.md']
+html_theme_options = {
+    'collapse_navigation': False,
+    'navigation_depth': 3,
+    'sticky_navigation': True,
+    'titles_only': False,
+    'style_external_links': True,
+}
+import os
+import os
+import sys
+path = os.path.abspath('../../src')
+sys.path.insert(0,path)
+extensions = [
+	'sphinx.ext.napoleon',
+	'sphinx.ext.autodoc',
+	'sphinx.ext.autosectionlabel',
+	'sphinx.ext.todo',
+	'sphinx.ext.viewcode',
+	'sphinx.ext.githubpages',
+	'sphinx.ext.autosummary',
+	'm2r']
+autosummary_generate = True
+source_suffix = ['.rst', '.md']
+html_theme_options = {
+    'collapse_navigation': False,
+    'navigation_depth': 3,
+    'sticky_navigation': True,
+    'titles_only': False,
+    'style_external_links': True,
+}
+import os
+import os
+import sys
+path = os.path.abspath('../../src')
+sys.path.insert(0,path)
+extensions = [
+	'sphinx.ext.napoleon',
+	'sphinx.ext.autodoc',
+	'sphinx.ext.autosectionlabel',
+	'sphinx.ext.todo',
+	'sphinx.ext.viewcode',
+	'sphinx.ext.githubpages',
+	'sphinx.ext.autosummary',
+	'm2r']
+autosummary_generate = True
+source_suffix = ['.rst', '.md']
+html_theme_options = {
+    'collapse_navigation': False,
+    'navigation_depth': 3,
+    'sticky_navigation': True,
+    'titles_only': False,
+    'style_external_links': True,
+}
