+++ docs/_sources/modules/transformations.z_score_normalization.rst.txt
@@ -1,1 +1,0 @@
-.. autofunction:: spac.transformations.z_score_normalization
