+++ docs/_sources/modules/transformations.rename_annotations.rst.txt
@@ -1,1 +1,0 @@
-.. autofunction:: spac.transformations.rename_annotations
