+++ docs/source/modules/spatial_analysis._get_labels.rst
@@ -1,1 +1,0 @@
-.. autofunction:: spac.spatial_analysis._get_labels
