+++ docs/source/modules/visualization._plot_spatial_distance_dispatch.rst
@@ -1,1 +1,0 @@
-.. autofunction:: spac.visualization._plot_spatial_distance_dispatch
