+++ docs/_sources/modules/data_utils.combine_dfs.rst.txt
@@ -1,1 +1,0 @@
-.. autofunction:: spac.data_utils.combine_dfs
