+++ docs/_sources/modules/spatial_analysis._process_matrixs.rst.txt
@@ -1,1 +1,0 @@
-.. autofunction:: spac.spatial_analysis._process_matrixs
