+++ docs/source/modules/visualization.interactive_spatial_plot.rst
@@ -1,1 +1,0 @@
-.. autofunction:: spac.visualization.interactive_spatial_plot
