+++ docs/_sources/modules/utils.rst.txt
@@ -1,23 +1,0 @@
-.. automodule:: spac.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :autosummary:
-
-spac.utils.regex_search_list
-spac.utils.regex_search
-spac.utils.check_list_in_list
-spac.utils.check_table
-spac.utils.check_annotation
-spac.utils.check_feature
-spac.utils.check_column_name
-spac.utils.check_distances
-spac.utils.text_to_others
-spac.utils.check_same
-spac.utils.annotation_category_relations
-spac.utils.color_mapping
-spac.utils.check_label
-spac.utils.spell_out_special_characters
-spac.utils.get_defined_color_map
-spac.utils.compute_boxplot_metrics
-spac.utils.compute_metrics
