+++ docs/_sources/modules/utag_functions.utag.rst.txt
@@ -1,1 +1,0 @@
-.. autofunction:: spac.utag_functions.utag
