+++ docs/source/modules/spatial_analysis.ripley_l.rst
@@ -1,1 +1,0 @@
-.. autofunction:: spac.spatial_analysis.ripley_l
