+++ docs/_sources/modules/visualization.calculate_histogram.rst.txt
@@ -1,1 +1,0 @@
-.. autofunction:: spac.visualization.calculate_histogram
