### 2025-05-30 b0e26d7
fix: updating documentation method
