+++ docs/source/modules/spac.utag_functions.rst
@@ -5,3 +5,17 @@ spac.utag\_functions module
    :members:
    :undoc-members:
    :show-inheritance:
+
+Functions
+---------
+
+.. toctree::
+   :maxdepth: 1
+
+   utag_functions.z_score
+   utag_functions.sparse_matrix_dstack
+   utag_functions._parallel_message_pass
+   utag_functions.custom_message_passing
+   utag_functions.low_variance_filter
+   utag_functions.add_probabilities_to_centroid
+   utag_functions.utag
