+++ docs/source/modules/spac.transformations.rst
@@ -5,3 +5,26 @@ spac.transformations module
    :members:
    :undoc-members:
    :show-inheritance:
+
+Functions
+---------
+
+.. toctree::
+   :maxdepth: 1
+
+   transformations.phenograph_clustering
+   transformations.knn_clustering
+   transformations.get_cluster_info
+   transformations.tsne
+   transformations.run_umap
+   transformations._validate_transformation_inputs
+   transformations._select_input_features
+   transformations.batch_normalize
+   transformations.rename_annotations
+   transformations.normalize_features
+   transformations.normalize_features_core
+   transformations.arcsinh_transformation
+   transformations.arcsinh_transformation_core
+   transformations.z_score_normalization
+   transformations.apply_per_batch
+   transformations.run_utag_clustering
