+++ docs/_sources/modules/phenotyping.is_binary_0_1.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.phenotyping.is_binary_0_1
