+++ docs/source/modules/visualization.present_summary_as_figure.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.present_summary_as_figure
