+++ docs/source/modules/transformations.knn_clustering.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.transformations.knn_clustering
