+++ docs/_sources/modules/spatial_analysis.neighborhood_profile.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.spatial_analysis.neighborhood_profile
