+++ docs/_sources/modules/spac.data_utils.rst.txt
@@ -5,3 +5,28 @@ spac.data\_utils module
    :members:
    :undoc-members:
    :show-inheritance:
+
+Functions
+---------
+
+.. toctree::
+   :maxdepth: 1
+
+   data_utils.append_annotation
+   data_utils.ingest_cells
+   data_utils.concatinate_regions
+   data_utils.rescale_features
+   data_utils.add_rescaled_features
+   data_utils.subtract_min_per_region
+   data_utils.subtract_min_quantile
+   data_utils.load_csv_files
+   data_utils.select_values
+   data_utils._select_values_dataframe
+   data_utils._select_values_anndata
+   data_utils.downsample_cells
+   data_utils.calculate_centroid
+   data_utils.bin2cat
+   data_utils.combine_dfs
+   data_utils.add_pin_color_rules
+   data_utils.combine_annotations
+   data_utils.summarize_dataframe
