+++ docs/source/modules/utag_functions.add_probabilities_to_centroid.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.utag_functions.add_probabilities_to_centroid
