+++ docs/_sources/modules/spac.phenotyping.rst.txt
@@ -5,3 +5,16 @@ spac.phenotyping module
    :members:
    :undoc-members:
    :show-inheritance:
+
+Functions
+---------
+
+.. toctree::
+   :maxdepth: 1
+
+   phenotyping.is_binary_0_1
+   phenotyping.decode_phenotype
+   phenotyping.generate_phenotypes_dict
+   phenotyping.apply_phenotypes
+   phenotyping.combine_phenotypes
+   phenotyping.assign_manual_phenotypes
