+++ docs/_sources/modules/utils.color_mapping.rst.txt
@@ -1,0 +1,1 @@
+.. autofunction:: spac.utils.color_mapping
