+++ docs/source/modules/visualization.boxplot.rst
@@ -1,0 +1,1 @@
+.. autofunction:: spac.visualization.boxplot
