+++ docs/_sources/modules/transformations.apply_per_batch.rst.txt
@@ -1,1 +1,4 @@
+apply_per_batch
+---------------
+
 .. autofunction:: spac.transformations.apply_per_batch
