+++ docs/source/modules/utils.check_table.rst
@@ -1,1 +1,4 @@
+check_table
+-----------
+
 .. autofunction:: spac.utils.check_table
