+++ docs/_sources/modules/phenotyping.decode_phenotype.rst.txt
@@ -1,1 +1,4 @@
+decode_phenotype
+----------------
+
 .. autofunction:: spac.phenotyping.decode_phenotype
