+++ docs/source/modules/data_utils.downsample_cells.rst
@@ -1,1 +1,4 @@
+downsample_cells
+----------------
+
 .. autofunction:: spac.data_utils.downsample_cells
