+++ docs/source/modules/spatial_analysis.calculate_nearest_neighbor.rst
@@ -1,1 +1,4 @@
+calculate_nearest_neighbor
+--------------------------
+
 .. autofunction:: spac.spatial_analysis.calculate_nearest_neighbor
