+++ docs/_sources/modules/utils.color_mapping.rst.txt
@@ -1,1 +1,4 @@
+color_mapping
+-------------
+
 .. autofunction:: spac.utils.color_mapping
