+++ docs/source/modules/visualization.interactive_spatial_plot.rst
@@ -1,1 +1,4 @@
+interactive_spatial_plot
+------------------------
+
 .. autofunction:: spac.visualization.interactive_spatial_plot
