+++ docs/source/modules/transformations.batch_normalize.rst
@@ -1,1 +1,4 @@
+batch_normalize
+---------------
+
 .. autofunction:: spac.transformations.batch_normalize
