+++ docs/source/modules/utag_functions._parallel_message_pass.rst
@@ -1,1 +1,4 @@
+_parallel_message_pass
+----------------------
+
 .. autofunction:: spac.utag_functions._parallel_message_pass
