+++ docs/_sources/modules/data_utils.rescale_features.rst.txt
@@ -1,1 +1,4 @@
+rescale_features
+----------------
+
 .. autofunction:: spac.data_utils.rescale_features
