### 2025-05-30 6a76768
fix: updating documentation method
