+++ docs/source/spac.rst
@@ -1,0 +1,19 @@
+spac package
+============
+
+Version: 0.9.0  
+Author: Fang Liu, Rui He, and George Zaki
+
+Submodules
+==========
+.. toctree::
+   :maxdepth: 1
+
+   modules/index
+
+Functions
+=========
+.. toctree::
+   :maxdepth: 1
+
+   modules/spac.*.*
