+++ docs/source/modules/spac.transformations.rst
@@ -12,19 +12,19 @@ Functions
 .. toctree::
    :maxdepth: 1
 
-   transformations.phenograph_clustering
-   transformations.knn_clustering
-   transformations.get_cluster_info
-   transformations.tsne
-   transformations.run_umap
-   transformations._validate_transformation_inputs
-   transformations._select_input_features
-   transformations.batch_normalize
-   transformations.rename_annotations
-   transformations.normalize_features
-   transformations.normalize_features_core
-   transformations.arcsinh_transformation
-   transformations.arcsinh_transformation_core
-   transformations.z_score_normalization
-   transformations.apply_per_batch
-   transformations.run_utag_clustering
+   spac.transformations.phenograph_clustering
+   spac.transformations.knn_clustering
+   spac.transformations.get_cluster_info
+   spac.transformations.tsne
+   spac.transformations.run_umap
+   spac.transformations._validate_transformation_inputs
+   spac.transformations._select_input_features
+   spac.transformations.batch_normalize
+   spac.transformations.rename_annotations
+   spac.transformations.normalize_features
+   spac.transformations.normalize_features_core
+   spac.transformations.arcsinh_transformation
+   spac.transformations.arcsinh_transformation_core
+   spac.transformations.z_score_normalization
+   spac.transformations.apply_per_batch
+   spac.transformations.run_utag_clustering
