+++ docs/_sources/modules/spac.visualization.rst.txt
@@ -12,24 +12,24 @@ Functions
 .. toctree::
    :maxdepth: 1
 
-   visualization.visualize_2D_scatter
-   visualization.dimensionality_reduction_plot
-   visualization.tsne_plot
-   visualization.histogram
-   visualization.heatmap
-   visualization.hierarchical_heatmap
-   visualization.threshold_heatmap
-   visualization.spatial_plot
-   visualization.boxplot
-   visualization.boxplot_interactive
-   visualization.interactive_spatial_plot
-   visualization.sankey_plot
-   visualization.relational_heatmap
-   visualization.plot_ripley_l
-   visualization._prepare_spatial_distance_data
-   visualization._plot_spatial_distance_dispatch
-   visualization._css_rgb_or_hex_to_hex
-   visualization._ordered_unique_figs
-   visualization.visualize_nearest_neighbor
-   visualization.present_summary_as_html
-   visualization.present_summary_as_figure
+   spac.visualization.visualize_2D_scatter
+   spac.visualization.dimensionality_reduction_plot
+   spac.visualization.tsne_plot
+   spac.visualization.histogram
+   spac.visualization.heatmap
+   spac.visualization.hierarchical_heatmap
+   spac.visualization.threshold_heatmap
+   spac.visualization.spatial_plot
+   spac.visualization.boxplot
+   spac.visualization.boxplot_interactive
+   spac.visualization.interactive_spatial_plot
+   spac.visualization.sankey_plot
+   spac.visualization.relational_heatmap
+   spac.visualization.plot_ripley_l
+   spac.visualization._prepare_spatial_distance_data
+   spac.visualization._plot_spatial_distance_dispatch
+   spac.visualization._css_rgb_or_hex_to_hex
+   spac.visualization._ordered_unique_figs
+   spac.visualization.visualize_nearest_neighbor
+   spac.visualization.present_summary_as_html
+   spac.visualization.present_summary_as_figure
