+++ docs/_sources/modules/spac.utils.rst.txt
@@ -12,17 +12,17 @@ Functions
 .. toctree::
    :maxdepth: 1
 
-   utils.regex_search_list
-   utils.check_list_in_list
-   utils.check_table
-   utils.check_annotation
-   utils.check_feature
-   utils.check_column_name
-   utils.check_distances
-   utils.text_to_others
-   utils.annotation_category_relations
-   utils.color_mapping
-   utils.check_label
-   utils.spell_out_special_characters
-   utils.get_defined_color_map
-   utils.compute_boxplot_metrics
+   spac.utils.regex_search_list
+   spac.utils.check_list_in_list
+   spac.utils.check_table
+   spac.utils.check_annotation
+   spac.utils.check_feature
+   spac.utils.check_column_name
+   spac.utils.check_distances
+   spac.utils.text_to_others
+   spac.utils.annotation_category_relations
+   spac.utils.color_mapping
+   spac.utils.check_label
+   spac.utils.spell_out_special_characters
+   spac.utils.get_defined_color_map
+   spac.utils.compute_boxplot_metrics

[…]
