+++ docs/source/spac.rst
@@ -4,102 +4,33 @@ spac package
 Version: 0.9.0  
 Author: Fang Liu, Rui He, and George Zaki
 
-Submodules
-==========
-.. toctree::
-   :maxdepth: 1
+Submodules & Functions
+=======================
 
-   modules/index
+.. automodule:: spac.data_utils
+   :members:
+   :undoc-members:
 
-Functions
-=========
-.. toctree::
-   :maxdepth: 1
+.. automodule:: spac.phenotyping
+   :members:
+   :undoc-members:
 
-   modules/spac.data_utils.add_pin_color_rules
-   modules/spac.data_utils.add_rescaled_features
-   modules/spac.data_utils.append_annotation
-   modules/spac.data_utils.bin2cat
-   modules/spac.data_utils.calculate_centroid
-   modules/spac.data_utils.combine_annotations
-   modules/spac.data_utils.combine_dfs
-   modules/spac.data_utils.concatinate_regions
-   modules/spac.data_utils.downsample_cells
-   modules/spac.data_utils.ingest_cells
-   modules/spac.data_utils.load_csv_files
-   modules/spac.data_utils.rescale_features
-   modules/spac.data_utils._select_values_anndata
-   modules/spac.data_utils._select_values_dataframe
-   modules/spac.data_utils.select_values
-   modules/spac.data_utils.subtract_min_per_region
-   modules/spac.data_utils.subtract_min_quantile
-   modules/spac.data_utils.summarize_dataframe
-   modules/spac.phenotyping.apply_phenotypes
-   modules/spac.phenotyping.assign_manual_phenotypes
-   modules/spac.phenotyping.combine_phenotypes
-   modules/spac.phenotyping.decode_phenotype
-   modules/spac.phenotyping.generate_phenotypes_dict
-   modules/spac.phenotyping.is_binary_0_1
-   modules/spac.spatial_analysis.calculate_nearest_neighbor
-   modules/spac.spatial_analysis._neighborhood_profile_core
-   modules/spac.spatial_analysis.neighborhood_profile
-   modules/spac.spatial_analysis.ripley_l
-   modules/spac.spatial_analysis.spatial_interaction
-   modules/spac.transformations.apply_per_batch
-   modules/spac.transformations.arcsinh_transformation_core
-   modules/spac.transformations.arcsinh_transformation
-   modules/spac.transformations.batch_normalize
-   modules/spac.transformations.get_cluster_info
-   modules/spac.transformations.knn_clustering
-   modules/spac.transformations.normalize_features_core
-   modules/spac.transformations.normalize_features
-   modules/spac.transformations.phenograph_clustering
-   modules/spac.transformations.rename_annotations
-   modules/spac.transformations.run_umap
-   modules/spac.transformations.run_utag_clustering
-   modules/spac.transformations._select_input_features
-   modules/spac.transformations.tsne
-   modules/spac.transformations._validate_transformation_inputs
-   modules/spac.transformations.z_score_normalization
-   modules/spac.utag_functions.add_probabilities_to_centroid
-   modules/spac.utag_functions.custom_message_passing
-   modules/spac.utag_functions.low_variance_filter
-   modules/spac.utag_functions._parallel_message_pass
-   modules/spac.utag_functions.sparse_matrix_dstack
-   modules/spac.utag_functions.utag
-   modules/spac.utag_functions.z_score
-   modules/spac.utils.annotation_category_relations
-   modules/spac.utils.check_annotation
-   modules/spac.utils.check_column_name
-   modules/spac.utils.check_distances
-   modules/spac.utils.check_feature
-   modules/spac.utils.check_label
-   modules/spac.utils.check_list_in_list
[…]
