+++ tests/test_visualization/test_embedded_scatter_plot.py
@@ -1,0 +1,524 @@
+import unittest
+from unittest.mock import patch
+import anndata
+import numpy as np
+import matplotlib
+import matplotlib.pyplot as plt
+import itertools
+from spac.visualization import embedded_scatter_plot
+matplotlib.use('Agg')
+
+
+class TestStaticScatterPlot(unittest.TestCase):
+
+    def setUp(self):
+        self.adata = anndata.AnnData(X=np.random.rand(10, 10))
+        self.adata.obsm['X_tsne'] = np.random.rand(10, 2)
+        self.adata.obsm['X_umap'] = np.random.rand(10, 2)
+        self.adata.obsm['X_pca'] = np.random.rand(10, 2)
+        self.adata.obsm['sumap'] = np.random.rand(10, 2)
+        self.adata.obsm['3dsumap'] = np.random.rand(10, 3)
+        self.adata.obs['annotation_column'] = np.random.choice(
+            ['A', 'B', 'C'], size=10
+        )
+        self.adata.var_names = ['gene_' + str(i) for i in range(10)]
+
+    def test_missing_umap_coordinates(self):
+        del self.adata.obsm['X_umap']
+        with self.assertRaises(ValueError) as cm:
+            embedded_scatter_plot(self.adata, 'umap')
+        expected_msg = (
+            "X_umap coordinates not found in adata.obsm. "
+            "Please run UMAP before calling this function."
+        )
+        self.assertEqual(str(cm.exception), expected_msg)
+
+    def test_missing_tsne_coordinates(self):
+        del self.adata.obsm['X_tsne']
+        with self.assertRaises(ValueError) as cm:
+            embedded_scatter_plot(self.adata, 'tsne')
+        expected_msg = (
+            "X_tsne coordinates not found in adata.obsm. "
+            "Please run TSNE before calling this function."
+        )
+        self.assertEqual(str(cm.exception), expected_msg)
+
+    def test_annotation_and_feature(self):
+        with self.assertRaises(ValueError) as cm:
+            embedded_scatter_plot(
+                self.adata, 'tsne',
+                annotation='annotation_column',
+                feature='feature_column'
+            )
+        expected_msg = (
+            "Please specify either an annotation or a feature for coloring, "
+            "not both."
+        )
+        self.assertEqual(str(cm.exception), expected_msg)
+
+    def test_annotation_column(self):
+        fig, ax = embedded_scatter_plot(
+            self.adata, 'tsne', annotation='annotation_column'
+        )
+        self.assertIsNotNone(fig)
+        self.assertIsNotNone(ax)
+        self.assertEqual(ax.get_xlabel(), 't-SNE 1')
+        self.assertEqual(ax.get_ylabel(), 't-SNE 2')
+        self.assertEqual(ax.get_title(), 'TSNE-annotation_column')
+
+    def test_associated_table(self):
+        fig, ax = embedded_scatter_plot(
+            self.adata,
+            annotation='annotation_column',
+            associated_table='sumap'
+        )
+        self.assertIsNotNone(fig)
+        self.assertIsNotNone(ax)
+        self.assertEqual(ax.get_xlabel(), 'sumap 1')
+        self.assertEqual(ax.get_ylabel(), 'sumap 2')
+        self.assertEqual(ax.get_title(), 'sumap-annotation_column')
+
+    def test_feature_column(self):
+        fig, ax = embedded_scatter_plot(
+            self.adata, 'tsne', feature='gene_1'
+        )
+        self.assertIsNotNone(fig)
+        self.assertIsNotNone(ax)
+        self.assertEqual(ax.get_xlabel(), 't-SNE 1')
[…]
