### 2025-06-03 9fe387b
addition of pinned colors, must be rebased to pass
