### 2025-06-03 4e59324
addition of pinned colors, correcting rebase through cherry pick
