+++ src/spac/visualization.py
@@ -15,7 +15,10 @@ from spac.utils import check_table, check_annotation
 from spac.utils import check_feature, annotation_category_relations
 from spac.utils import check_label
 from spac.utils import get_defined_color_map
+<<<<<<< HEAD
 from spac.utils import compute_boxplot_metrics
+=======
+>>>>>>> 2a60092 (addition of pinned colors, must be rebased to pass)
 from functools import partial
 from spac.utils import color_mapping, spell_out_special_characters
 from spac.data_utils import select_values
