### 2025-06-03 3ef099e
Combined dimensionality and spatial plot
