### 2025-06-05 3246e42
init commit for stat_test
