### 2025-06-06 8b83d35
Added test to statistical test
