+++ tests/test_visualization/test_statistics.py
@@ -1,4 +1,4 @@
-import pytest
+import unittest
 import pandas as pd
 import numpy as np
 from scipy import stats
