### 2025-06-13 507ef7f
feat(downsample):AnnData input and output for downsample_cells
This commit modifies the 'downsample_cells' function and adds a  helper function '_get_downsampled_indices' to provide cell downsampling capabilities for both AnnData objects and Pandas DataFrames.
