### 2025-06-13 3e84383
Update test_downsample_cells.py
This test now checks anndata objects are accepted, downsampled correctly, and returned as annadata objects.
