### 2025-06-13 c110994
Fix downsample_cells in data_utils.py
Reordering the code to convert annotations to a list before extracting annotation information, ensuring it is in DataFrame format as required by subsequent downsample_cells code.
