### 2025-06-17 eda8b49
Format code and move function to utils.py
