### 2025-06-20 994bac4
feat(qc-metrics): add common single cell quality control metrics
