### 2025-06-20 59675ca
style(qc-metrics): fix spelling typo in nFeature metric
