### 2025-06-20 7bf4043
Update src/spac/transformations.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
