### 2025-06-23 0cf530b
fix(check_layer): use check_table spac function to evaluate if adata.layer is present
