### 2025-07-01 95ca56d
correction of errors and unittests
