### 2025-07-01 729cfee
test_visualize_2d test case update
