### 2025-07-03 a228e5e
feat(qc_summary_statistics): add summary statistics table for sc/spatial transcriptomics quality control metrics
