### 2025-07-11 5e03dc2
refactor(get_qc_summary_table): adjust code style to adhere to spac guidlines closer
