+++ tests/test_transformations/test_get_qc_summary_table.py
@@ -32,8 +32,8 @@ class TestGetQCSummaryTable(unittest.TestCase):
     def test_qc_summary_table_basic(self):
         adata = self.create_test_adata()
         get_qc_summary_table(adata)
-        self.assertIn("qc_summary_table", adata.uns)
         summary = adata.uns["qc_summary_table"]
+        self.assertIn("qc_summary_table", adata.uns)
         self.assertTrue(isinstance(summary, pd.DataFrame))
         # Check that all expected columns are present
         self.assertIn("mean", summary.columns)
@@ -44,7 +44,8 @@ class TestGetQCSummaryTable(unittest.TestCase):
         self.assertIn("lower_quantile", summary.columns)
         self.assertIn("Sample", summary.columns)
         # Check that the correct metrics are summarized
-        self.assertEqual(set(summary["metric_name"]), {"nFeature", "nCount", "percent.mt"})
+        self.assertEqual(set(summary["metric_name"]), 
+                         {"nFeature", "nCount", "percent.mt"})
         # Check that the sample label is correct when not grouping
         self.assertEqual(summary["Sample"].iloc[0], "All")
 
@@ -52,8 +53,11 @@ class TestGetQCSummaryTable(unittest.TestCase):
     def test_qc_summary_table_non_numeric(self):
         adata = self.create_test_adata()
         adata.obs["non_numeric"] = ["a", "b", "c"]
-        with self.assertRaises(TypeError):
-            get_qc_summary_table(adata, stat_columns_list=["nFeature", "non_numeric"])
+        with self.assertRaises(TypeError) as exc_info:
+            get_qc_summary_table(adata, 
+                                 stat_columns_list=["nFeature", "non_numeric"])
+        expected_msg = 'Column "non_numeric" must be numeric to compute statistics.'
+        self.assertEqual(str(exc_info.exception), expected_msg)
 
     # Test that summary statistics are computed correctly for nFeature and nCount
     def test_qc_summary_table_statistics(self):
@@ -68,16 +72,46 @@ class TestGetQCSummaryTable(unittest.TestCase):
         self.assertEqual(nfeature_row["lower_mad"], 2)
         self.assertEqual(nfeature_row["upper_quantile"], 2)
         self.assertEqual(nfeature_row["lower_quantile"], 2)
-        # Check mean, median, quantiles for nCount
+        # Check nCount statistics
+        # nCount per cell = [4, 6, 11] -> 
+        # mean 7.0, median 6.0, 95th pct 10.5, 5th pct 4.2
         ncount_row = summary[summary["metric_name"] == "nCount"].iloc[0]
-        expected_mean = np.mean([4, 6, 11])
-        expected_median = np.median([4, 6, 11])
-        expected_upper = np.percentile([4, 6, 11], 95)
-        expected_lower = np.percentile([4, 6, 11], 5)
-        self.assertAlmostEqual(ncount_row["mean"], expected_mean)
-        self.assertAlmostEqual(ncount_row["median"], expected_median)
-        self.assertAlmostEqual(ncount_row["upper_quantile"], expected_upper)
-        self.assertAlmostEqual(ncount_row["lower_quantile"], expected_lower)
+        self.assertAlmostEqual(ncount_row["mean"], 7.0) 
+        self.assertAlmostEqual(ncount_row["median"], 6.0)  
+        self.assertAlmostEqual(ncount_row["upper_quantile"], 10.5) 
+        self.assertAlmostEqual(ncount_row["lower_quantile"], 4.2)  
+
+    # Test that summary statistics is computed correctly with sample_column grouping
+    def test_qc_summary_table_grouping(self):
+        adata = self.create_test_adata()
[…]

[…]
