### 2025-07-11 239f981
Update src/spac/utils.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
