### 2025-07-11 df46aca
make sure that compute_summary_qc_stats has a default input for stat_columns_list
