+++ src/spac/utils.py
@@ -1199,7 +1199,7 @@ def compute_summary_qc_stats(
         n_mad: int = 5,
         upper_quantile: float = 0.95,
         lower_quantile: float = 0.05,
-        stat_columns_list: List[str]
+        stat_columns_list: List[str] = ['nFeature', 'nCount', 'percent.mt']
         ) -> pd.DataFrame:
     
     """
