### 2025-07-15 ce2bf74
feat(plot_qc_metrics): add plot to visualize sc/spatial transcriptomics quality control metrics
