### 2025-07-15 ea21a00
refactor(plot_qc_metrics): make adjustments based on copilot code revieww
