### 2025-07-15 d5061c4
refactor(get_qc_summary_table): refactor quality control summary statistics function and tests based on the PR review
