### 2025-07-15 #8211 @kevinyamauchi
Hey @adamltyson ! Nice to hear from you. I am just about to head on holiday through July 29. Thus, I likely wouldn't be able to complete the review until August 21. Would that be okay? (no worries if you want to pass the review to somebody with a quicker turnaround)
