### 2025-07-16 eaf0e16
feat(ripley_template): Add Ripley‑L template, helpers, fixtures and tests
