### 2025-07-16 3aa787d
fix: expose src/ on PYTHONPATH for template tests
