### 2025-07-16 b453cbf
Revert "fix: expose src/ on PYTHONPATH for template tests"
This reverts commit 3aa787df0a4c9bf8e01a95437a942c86257b748c.
