### 2025-07-16 fb7343e
feat(ripley_template): new test, revert PYTHONPATH hack, remove fixture, flake8 fixes
