### 2025-07-22 2fd6a71
feat(template_function): Add new template folders and analysis examples
