+++ src/spac/templates_2/template_utils.py
@@ -1,0 +1,419 @@
+"""
+Improved template utilities for SPAC templates.
+Designed to work seamlessly with NIDAP-style templates and maintain compatibility.
+"""
+from pathlib import Path
+import pickle
+from typing import Any, Dict, Union, Optional, Tuple
+import json
+import pandas as pd
+
+
+def load_input(file_path: Union[str, Path], format: Optional[str] = None):
+    """
+    Load input data from either h5ad or pickle file.
+    
+    Parameters
+    ----------
+    file_path : str or Path
+        Path to input file (h5ad or pickle)
+    format : str, optional
+        Force specific format ('h5ad' or 'pickle'). If None, auto-detect.
+    
+    Returns
+    -------
+    Loaded data object (typically AnnData)
+    """
+    path = Path(file_path)
+    
+    if not path.exists():
+        raise FileNotFoundError(f"Input file not found: {file_path}")
+
+    # Determine format
+    if format is None:
+        suffix = path.suffix.lower()
+        if suffix in ['.h5ad', '.h5']:
+            format = 'h5ad'
+        elif suffix in ['.pickle', '.pkl', '.p']:
+            format = 'pickle'
+        else:
+            # Try to auto-detect
+            format = 'auto'
+    
+    # Load based on format
+    if format == 'h5ad':
+        try:
+            import anndata as ad
+            return ad.read_h5ad(path)
+        except ImportError:
+            raise ImportError("anndata package required to read h5ad files")
+        except Exception as e:
+            if format == 'auto':
+                # Try pickle as fallback
+                format = 'pickle'
+            else:
+                raise ValueError(f"Error reading h5ad file: {e}")
+    
+    if format == 'pickle' or format == 'auto':
+        try:
+            with path.open('rb') as fh:
+                return pickle.load(fh)
+        except Exception as e:
+            raise ValueError(
+                f"Unable to load file '{file_path}'. "
+                f"Supported formats: h5ad, pickle. Error: {e}"
+            )
+
+
+def save_output(
+    data: Any,
+    output_path: Union[str, Path],
+    format: Optional[str] = None
+) -> str:
+    """
+    Save a single output to file. Returns the absolute path as a string.
+    This matches NIDAP behavior more closely.
+    
+    Parameters
+    ----------
+    data : Any
+        Data to save (AnnData, DataFrame, etc.)
+    output_path : str or Path
+        Output file path
+    format : str, optional
+        Force specific format. If None, auto-detect from extension.
+    
+    Returns
+    -------
+    str
+        Absolute path to saved file
+    """
+    output_path = Path(output_path)
+    output_path.parent.mkdir(parents=True, exist_ok=True)
+    
+    # Determine format
+    if format is None:
+        suffix = output_path.suffix.lower()
+        if suffix in ['.h5ad', '.h5']:
+            format = 'h5ad'
+        elif suffix in ['.pickle', '.pkl', '.p']:
+            format = 'pickle'
+        elif suffix == '.csv':
+            format = 'csv'
+        else:
+            # Default based on data type
+            if hasattr(data, 'write_h5ad'):
+                format = 'h5ad'
+            elif isinstance(data, pd.DataFrame):
+                format = 'csv'
+            else:
+                format = 'pickle'
+    
+    # Save based on format
+    if format == 'h5ad':
+        if hasattr(data, 'write_h5ad'):
+            data.write_h5ad(output_path)
+        else:
+            raise ValueError(f"Cannot save {type(data)} as h5ad")
+    
+    elif format == 'csv':
+        if isinstance(data, pd.DataFrame):
+            data.to_csv(output_path, index=False)
+        else:
+            raise ValueError(f"Cannot save {type(data)} as CSV")
[…]
