### 2025-07-22 17f9fa3
Revert "feat(template_function): Add new template folders and analysis examples"
This reverts commit 2fd6a718a6e38d7e2fa8f9c193a3003144432991.
