+++ src/spac/templates_2/__init__.py
@@ -1,13 +1,0 @@
-"""
-Canonical SPAC template sub‑package.
-
-Each template is a self‑contained module that
-  • reads parameters from JSON/dict
-  • runs a SPAC analysis function
-  • returns / saves results
-
-Available templates
--------------------
-- ripley_l_template.run_from_json
-"""
-
