### 2025-07-22 da4b3d7
fix(template_func): add template and tests
