+++ src/spac/templates/ripley_l_template.py
@@ -1,107 +1,242 @@
 """
-Platform‑agnostic Ripley‑L template.
+Platform-agnostic Ripley-L template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
 
 Usage
 -----
 >>> from spac.templates.ripley_l_template import run_from_json
 >>> run_from_json("examples/ripley_l_params.json")
 """
-from __future__ import annotations
 import json
+import sys
 from pathlib import Path
-from typing import Any, Dict, Union
+from typing import Any, Dict, Union, List
+import pandas as pd
 
-import anndata as ad
-from spac.utils import check_table
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent))
+
 from spac.spatial_analysis import ripley_l
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
 
-from .template_utils import load_pickle, save_pickle
 
+# def _prepare_ripley_uns_for_h5ad(adata) -> None:
+#     """
+#     Minimal fix for ripley_l results serialization.
+#     Converts object columns to appropriate types for H5AD storage.
+#     """
+#     if "ripley_l" not in adata.uns:
+#         return
 
-# public API
-def run_from_json(src: Union[str, Path, Dict[str, Any]]) -> ad.AnnData:
+#     rl = adata.uns.get("ripley_l")
+#     if isinstance(rl, pd.DataFrame):
+#         # Create a copy to avoid modifying the original
+#         clean_df = rl.copy()
+
+#         # Process each object column
+#         for col in clean_df.columns:
+#             if clean_df[col].dtype == "object":
+#                 # Try to convert to numeric
+#                 try:
+#                     clean_df[col] = pd.to_numeric(
+#                         clean_df[col], errors='raise'
+#                     )
+#                 except Exception:
+#                     # If that fails, convert to string
+#                     clean_df[col] = clean_df[col].astype(str)
+
+#         # Replace with cleaned version
+#         adata.uns["ripley_l"] = clean_df
+
+def _prepare_ripley_uns_for_h5ad(adata) -> None:
     """
-    Execute Ripley‑L with parameters supplied via *src*.
+    Enhanced fix for ripley_l results serialization.
+    Ensures proper data types and structure for H5AD storage.
+    """
+    if "ripley_l" not in adata.uns:
+        return
+
+    rl = adata.uns.get("ripley_l")
+    
+    # Handle case where ripley_l might be a string (corrupted data)
+    if isinstance(rl, str):
+        print(f"Warning: ripley_l data appears to be corrupted (string): {rl[:100]}...")
+        # Try to reconstruct or remove corrupted data
+        del adata.uns["ripley_l"]
+        print("Removed corrupted ripley_l data from adata.uns")
+        return
+    
+    # Handle DataFrame case
+    if isinstance(rl, pd.DataFrame):
+        # Create a copy to avoid modifying the original
+        clean_df = rl.copy()
+
+        # Process each object column
+        for col in clean_df.columns:
+            if clean_df[col].dtype == "object":
+                # Check if column contains dictionaries or other complex objects
+                sample_val = clean_df[col].iloc[0] if len(clean_df) > 0 else None
+                
+                if isinstance(sample_val, dict):
+                    # Convert dict columns to JSON strings for H5AD compatibility
+                    clean_df[col] = clean_df[col].apply(lambda x: json.dumps(x) if isinstance(x, dict) else str(x))
+                else:
+                    # Try to convert to numeric first
+                    try:
+                        clean_df[col] = pd.to_numeric(clean_df[col], errors='raise')
[…]
