+++ src/spac/templates/template_utils.py
@@ -1,16 +1,312 @@
 from pathlib import Path
 import pickle
+from typing import Any, Dict, Union, Optional
+import json
+import anndata as ad
 
 
-def load_pickle(path: str):
-    """Load any pickled Python object from *path*."""
-    with Path(path).open("rb") as fh:
-        return pickle.load(fh)
+def load_input(file_path: Union[str, Path]):
+    """
+    Load input data from either h5ad or pickle file.
+    
+    Parameters
+    ----------
+    file_path : str or Path
+        Path to input file (h5ad or pickle)
+    
+    Returns
+    -------
+    Loaded data object (typically AnnData)
+    """
+    path = Path(file_path)
+    
+    if not path.exists():
+        raise FileNotFoundError(f"Input file not found: {file_path}")
+
+    # Check file extension
+    suffix = path.suffix.lower()
+
+    if suffix in ['.h5ad', '.h5']:
+        # Load h5ad file
+        try:
+            import anndata as ad
+            return ad.read_h5ad(path)
+        except ImportError:
+            raise ImportError(
+                "anndata package required to read h5ad files"
+            )
+        except Exception as e:
+            raise ValueError(f"Error reading h5ad file: {e}")
+
+    elif suffix in ['.pickle', '.pkl', '.p']:
+        # Load pickle file
+        with path.open('rb') as fh:
+            return pickle.load(fh)
+
+    else:
+        # Try to detect file type by content
+        try:
+            # First try h5ad
+            import anndata as ad
+            return ad.read_h5ad(path)
+        except Exception:
+            # Fall back to pickle
+            try:
+                with path.open('rb') as fh:
+                    return pickle.load(fh)
+            except Exception as e:
+                raise ValueError(
+                    f"Unable to load file '{file_path}'. "
+                    f"Supported formats: h5ad, pickle. Error: {e}"
+                )
 
 
-def save_pickle(obj, path: str):
-    """Save *obj* to *path*; create parent directories if absent."""
-    p = Path(path)
-    p.parent.mkdir(parents=True, exist_ok=True)
-    with p.open("wb") as fh:
-        pickle.dump(obj, fh)
+def save_outputs(outputs: Dict[str, Any],
+                 output_dir: Union[str, Path] = ".") -> Dict[str, str]:
+    """
+    Save multiple outputs to files and return a dict {filename: absolute_path}.
+    (Always a dict, even if just one file.)
+    
+    Parameters
+    ----------
+    outputs : dict
+        Dictionary where:
+        - key: filename (with extension)
+        - value: object to save
+    output_dir : str or Path
+        Directory to save files
+    
+    Returns
+    -------
+    dict
+        Dictionary of saved file paths
+    
+    Example
+    -------
+    >>> outputs = {
+    ...     "adata.h5ad": adata,
+    ...     "results.csv": results_df,
+    ...     "adata.pickle": adata
+    ... }
+    >>> saved = save_outputs(outputs, "results/")
+    """
+    output_dir = Path(output_dir)
+    output_dir.mkdir(parents=True, exist_ok=True)
+    
+    saved_files = {}
+    
+    for filename, obj in outputs.items():
+        filepath = output_dir / filename
+        
+        # Save based on file extension
+        if filename.endswith('.csv'):
+            obj.to_csv(filepath, index=False)
+        elif filename.endswith('.h5ad'):
+            ad.write(filepath)
+        elif filename.endswith(('.pickle', '.pkl')):
+            with open(filepath, 'wb') as f:
+                pickle.dump(obj, f)
+        elif hasattr(obj, "savefig"):
+            obj.savefig(filepath.with_suffix('.png'))
+            filepath = filepath.with_suffix('.png')
+        else:
+            # Default to pickle
[…]

[…]
