### 2025-07-24 f2c4d56
WIP: full template refactor before PR split
