+++ tests/templates/test_visualize_ripley_template.py
@@ -1,0 +1,217 @@
+# tests/templates/test_visualize_ripley_template.py
+"""Unit tests for the Visualize Ripley L template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+from unittest.mock import patch, MagicMock
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.visualize_ripley_template import run_from_json
+
+
+def mock_adata_with_ripley(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData with Ripley L results for tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "phenotype": ["B cells", "CD8 T cells"] * (n_cells // 2)
+    })
+    x_mat = rng.normal(size=(n_cells, 2))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.obsm["spatial"] = rng.random((n_cells, 2)) * 50.0
+    
+    # Add mock Ripley L results in the expected format
+    # When using pickle, the structure is preserved as-is
+    adata.uns["ripley_l_B cells_CD8 T cells"] = {
+        "radius": [0, 50, 100],
+        "ripley_l": [0, 1.2, 2.5],
+        "simulations": np.array([
+            [0, 0.8, 1.9], [0, 1.1, 2.3], [0, 1.3, 2.7]
+        ])
+    }
+    return adata
+
+
+class TestVisualizeRipleyTemplate(unittest.TestCase):
+    """Unit tests for the Visualize Ripley L template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "ripley_output.pickle"
+        )
+        self.out_file = "plots.csv"
+
+        # Save minimal mock data with Ripley results as pickle
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata_with_ripley(), f)
+
+        # Minimal parameters
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Center_Phenotype": "B cells",
+            "Neighbor_Phenotype": "CD8 T cells",
+            "Plot_Specific_Regions": False,
+            "Regions_Labels": [],
+            "Plot_Simulations": True,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.visualize_ripley_template.plot_ripley_l')
+    def test_run_with_save(self, mock_plot_ripley) -> None:
+        """Test visualization with file saving."""
+        # Mock the plot_ripley_l function to return a figure and dataframe
+        mock_fig = MagicMock()
+        mock_df = pd.DataFrame({
+            'radius': [0, 50, 100],
+            'ripley_l': [0, 1.2, 2.5],
+            'lower_ci': [0, 0.8, 1.9],
+            'upper_ci': [0, 1.6, 3.1]
+        })
+        mock_plot_ripley.return_value = (mock_fig, mock_df)
+        
+        # Suppress warnings for cleaner test output
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+
+            # Test with save_results=True (default)
+            saved_files = run_from_json(self.params)
+            self.assertIn(self.out_file, saved_files)
+            
+            # Verify plot_ripley_l was called with correct parameters
[…]
