### 2025-07-24 b960684
feat(template_utils): add template_utils and unit tests
