+++ tests/templates/test_template_utils.py
@@ -1,0 +1,297 @@
+# tests/utils/test_template_utils.py
+"""Unit tests for template utilities."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    text_to_value,
+    convert_pickle_to_h5ad,
+    convert_to_floats
+)
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "cell_type": ["TypeA", "TypeB"] * (n_cells // 2)
+    })
+    x_mat = rng.normal(size=(n_cells, 2))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    return adata
+
+
+def mock_dataframe(n_rows: int = 5) -> pd.DataFrame:
+    """Return a minimal DataFrame for fast tests."""
+    return pd.DataFrame({
+        "col1": range(n_rows),
+        "col2": [f"value_{i}" for i in range(n_rows)]
+    })
+
+
+class TestTemplateUtils(unittest.TestCase):
+    """Unit tests for template utility functions."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.test_adata = mock_adata()
+        self.test_df = mock_dataframe()
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering all input/output scenarios."""
+        # Suppress warnings for cleaner test output
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+
+            # Test 1: Load h5ad file
+            h5ad_path = os.path.join(self.tmp_dir.name, "test.h5ad")
+            self.test_adata.write_h5ad(h5ad_path)
+            loaded_h5ad = load_input(h5ad_path)
+            self.assertEqual(loaded_h5ad.n_obs, 10)
+            self.assertIn("cell_type", loaded_h5ad.obs.columns)
+
+            # Test 2: Load pickle file
+            pickle_path = os.path.join(self.tmp_dir.name, "test.pickle")
+            with open(pickle_path, "wb") as f:
+                pickle.dump(self.test_adata, f)
+            loaded_pickle = load_input(pickle_path)
+            self.assertEqual(loaded_pickle.n_obs, 10)
+
+            # Test 3: Load .pkl extension
+            pkl_path = os.path.join(self.tmp_dir.name, "test.pkl")
+            with open(pkl_path, "wb") as f:
+                pickle.dump(self.test_adata, f)
+            loaded_pkl = load_input(pkl_path)
+            self.assertEqual(loaded_pkl.n_obs, 10)
+
+            # Test 4: Load .p extension
+            p_path = os.path.join(self.tmp_dir.name, "test.p")
+            with open(p_path, "wb") as f:
+                pickle.dump(self.test_adata, f)
+            loaded_p = load_input(p_path)
+            self.assertEqual(loaded_p.n_obs, 10)
+
+            # Test 5: Save outputs - multiple formats
+            outputs = {
+                "result.pickle": self.test_adata,  # Now preferred format
+                "data.csv": self.test_df,
+                "adata.pkl": self.test_adata,
+                "adata.h5ad": self.test_adata,  # Still supported
[…]
