### 2025-07-24 c889259
feat(ripley_l_template): add ripley_l_template and unit tests
