+++ tests/templates/test_ripley_l_template.py
@@ -1,0 +1,245 @@
+# tests/templates/test_ripley_l_template.py
+"""Unit‑tests for the Ripley‑L template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+from unittest.mock import patch, MagicMock
+import anndata as ad
+import numpy as np
+import pandas as pd
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.ripley_l_template import run_from_json
+
+
+def mock_adata(n_cells: int = 40) -> ad.AnnData:
+    """Return a tiny synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame(
+        {
+            "renamed_phenotypes": np.where(
+                rng.random(n_cells) > 0.5, "B cells", "CD8 T cells"
+            )
+        }
+    )
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.obsm["spatial"] = rng.random((n_cells, 2)) * 300.0
+    return adata
+
+
+class TestRipleyLTemplate(unittest.TestCase):
+    """Light‑weight sanity checks for the Ripley‑L template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(self.tmp_dir.name, "input.pickle")
+        self.out_file = os.path.join(self.tmp_dir.name, "output.pickle")
+
+        # Save as pickle
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Radii": [0, 50, 100],
+            "Annotation": "renamed_phenotypes",
+            "Center_Phenotype": "B cells",
+            "Neighbor_Phenotype": "CD8 T cells",
+            "Output_path": self.out_file,
+            "Stratify_By": "None",
+            "Number_of_Simulations": 100,
+            "Area": "None",
+            "Seed": 42,
+            "Spatial_Key": "spatial",
+            "Edge_Correction": True
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.ripley_l_template.ripley_l')
+    def test_run_from_dict_with_save(self, mock_ripley_l) -> None:
+        """Test run_from_json with dict parameters and file saving."""
+        # Mock the ripley_l function 
+        def mock_ripley_side_effect(adata, **kwargs):
+            # Simulate what ripley_l does - adds results to adata.uns
+            phenotypes = kwargs.get('phenotypes', ['B cells', 'CD8 T cells'])
+            key = f"ripley_l_{phenotypes[0]}_{phenotypes[1]}"
+            adata.uns[key] = {
+                "radius": kwargs.get('distances', [0, 50, 100]),
+                "ripley_l": [0.0, 1.2, 2.5],
+                "simulations": [[0.0, 0.8, 1.9], [0.0, 1.1, 2.3]],
+            }
+            return None
+        
+        mock_ripley_l.side_effect = mock_ripley_side_effect
+
+        # Run with save_results=True (default)
+        saved_files = run_from_json(self.params)
+        
+        # Check that ripley_l was called with correct parameters
+        mock_ripley_l.assert_called_once()
+        call_args = mock_ripley_l.call_args
+        
+        # Verify the call arguments
+        self.assertEqual(call_args[1]['annotation'], "renamed_phenotypes")
[…]
