### 2025-07-24 48608e2
feat(visualize_ripley_template): add visualize_ripley_template and unit tests
