### 2025-07-24 0e2747e
fix(visualize_ripley): make plt.show() conditional based on show_plot parameter
