### 2025-07-24 e9f0883
fix(template_utils): address review comments
