### 2025-07-24 c94b219
fix(template_utils): address review comments again
