### 2025-07-24 9914716
fix(ripley_template): address review comments - replace debug prints with logging
