+++ tests/templates/test_ripley_l_template.py
@@ -1,5 +1,5 @@
 # tests/templates/test_ripley_l_template.py
-"""Unit‑tests for the Ripley‑L template."""
+"""Unit-tests for the Ripley-L template."""
 
 import json
 import os
@@ -36,7 +36,7 @@ def mock_adata(n_cells: int = 40) -> ad.AnnData:
 
 
 class TestRipleyLTemplate(unittest.TestCase):
-    """Light‑weight sanity checks for the Ripley‑L template."""
+    """Light-weight sanity checks for the Ripley-L template."""
 
     def setUp(self) -> None:
         self.tmp_dir = tempfile.TemporaryDirectory()
@@ -68,7 +68,7 @@ class TestRipleyLTemplate(unittest.TestCase):
     @patch('spac.templates.ripley_l_template.ripley_l')
     def test_run_from_dict_with_save(self, mock_ripley_l) -> None:
         """Test run_from_json with dict parameters and file saving."""
-        # Mock the ripley_l function 
+        # Mock the ripley_l function
         def mock_ripley_side_effect(adata, **kwargs):
             # Simulate what ripley_l does - adds results to adata.uns
             phenotypes = kwargs.get('phenotypes', ['B cells', 'CD8 T cells'])
@@ -79,16 +79,16 @@ class TestRipleyLTemplate(unittest.TestCase):
                 "simulations": [[0.0, 0.8, 1.9], [0.0, 1.1, 2.3]],
             }
             return None
-        
+
         mock_ripley_l.side_effect = mock_ripley_side_effect
 
         # Run with save_results=True (default)
         saved_files = run_from_json(self.params)
-        
+
         # Check that ripley_l was called with correct parameters
         mock_ripley_l.assert_called_once()
         call_args = mock_ripley_l.call_args
-        
+
         # Verify the call arguments
         self.assertEqual(call_args[1]['annotation'], "renamed_phenotypes")
         self.assertEqual(call_args[1]['phenotypes'], ["B cells", "CD8 T cells"])
@@ -97,15 +97,15 @@ class TestRipleyLTemplate(unittest.TestCase):
         self.assertEqual(call_args[1]['seed'], 42)
         self.assertEqual(call_args[1]['spatial_key'], "spatial")
         self.assertEqual(call_args[1]['edge_correction'], True)
-        
+
         # Check that output file was created - check if any file was saved
         self.assertTrue(
             len(saved_files) > 0,
             f"Expected files to be saved, but got {saved_files}"
         )
-        
+
         # Check that at least one pickle file was created
-        pickle_files = [f for f in saved_files.values() 
+        pickle_files = [f for f in saved_files.values()
                        if f.endswith('.pickle')]
         self.assertTrue(
             len(pickle_files) > 0,
@@ -124,18 +124,18 @@ class TestRipleyLTemplate(unittest.TestCase):
                 "ripley_l": [0.0, 1.2, 2.5],
             }
             return None
-        
+
         mock_ripley_l.side_effect = mock_ripley_side_effect
-        
+
         # Run with save_results=False
         adata_result = run_from_json(self.params, save_results=False)
-        
+
         # Check that ripley_l was called
         mock_ripley_l.assert_called_once()
-        
+
         # Check that we got an AnnData object back
         self.assertIsInstance(adata_result, ad.AnnData)
-        
+
         # Check that results are in the object
         ripley_key = "ripley_l_B cells_CD8 T cells"
         self.assertIn(ripley_key, adata_result.uns)
@@ -153,18 +153,18 @@ class TestRipleyLTemplate(unittest.TestCase):
                 "simulations": [[0.0, 0.8, 1.9], [0.0, 1.1, 2.3]],
             }
             return None
-        
+
         mock_ripley_l.side_effect = mock_ripley_side_effect
-        
+
[…]
