### 2025-07-25 415df89
fix(ripley_template): address review comments - merge dev into the branch and fix unit test
