### 2025-07-27 5456658
feat(load_csv): Add load_csv template function with configuration support
- Add load_csv_files() to template_utils.py for loading and combining CSV files
- Add spell_out_special_characters() to handle biological marker names
- Add load_csv_files_with_config.py template wrapper for NIDAP compatibility
- Add comprehensive unit tests for both functions
- Support column name cleaning, metadata mapping, and string column enforcement
