+++ src/spac/templates/template_utils.py
@@ -2,9 +2,12 @@ from pathlib import Path
 import pickle
 from typing import Any, Dict, Union, Optional, List
 import json
+import pandas as pd
 import anndata as ad
+import re
 import logging
 logger = logging.getLogger(__name__)
+
 
 def load_input(file_path: Union[str, Path]):
     """
@@ -108,7 +111,8 @@ def save_outputs(
             # Still support h5ad, but not the default
             if type(obj) is not ad.AnnData:
                 raise TypeError(
-                    f"Object for '{str(filename)}' must be AnnData, got {type(obj)}"
+                    f"Object for '{str(filename)}' must be AnnData, "
+                    f"got {type(obj)}"
                 )
             logger.info(f"Saving AnnData to {str(filepath)}")
             logger.debug(f"AnnData object: {obj}")
@@ -357,3 +361,253 @@ def convert_pickle_to_h5ad(
     adata.write_h5ad(h5ad_path)
 
     return str(h5ad_path)
+
+
+def spell_out_special_characters(text: str) -> str:
+    """
+    Clean column names by replacing special characters with text equivalents.
+
+    Handles biological marker names like:
+    - "CD4+" → "CD4_pos"
+    - "CD8-" → "CD8_neg"
+    - "CD4+CD20-" → "CD4_pos_CD20_neg"
+    - "CD4+/CD20-" → "CD4_pos_slashCD20_neg"
+    - "CD4+ CD20-" → "CD4_pos_CD20_neg"
+    - "Area µm²" → "Area_um2"
+
+    Parameters
+    ----------
+    text : str
+        The text to clean
+
+    Returns
+    -------
+    str
+        Cleaned text with special characters replaced
+    """
+    # Replace spaces with underscores
+    text = text.replace(' ', '_')
+
+    # Replace specific substrings for units
+    text = text.replace('µm²', 'um2')
+    text = text.replace('µm', 'um')
+
+    # Handle hyphens between alphanumeric characters FIRST
+    # (before + and - replacements)
+    # This pattern matches a hyphen that has alphanumeric on both sides
+    text = re.sub(r'(?<=[A-Za-z0-9])-(?=[A-Za-z0-9])', '_', text)
+
+    # Now replace remaining '+' with '_pos_' and '-' with '_neg_'
+    text = text.replace('+', '_pos_')
+    text = text.replace('-', '_neg_')
+
+    # Mapping for specific characters
+    special_char_map = {
+        'µ': 'u',       # Micro symbol replaced with 'u'
+        '²': '2',       # Superscript two replaced with '2'
+        '@': 'at',
+        '#': 'hash',
+        '$': 'dollar',
+        '%': 'percent',
+        '&': 'and',
+        '*': 'asterisk',
+        '/': 'slash',
+        '\\': 'backslash',
+        '=': 'equals',
+        '^': 'caret',
+        '!': 'exclamation',
+        '?': 'question',
+        '~': 'tilde',
+        '|': 'pipe',
+        ',': '',        # Remove commas
+        '(': '',        # Remove parentheses
+        ')': '',        # Remove parentheses
+        '[': '',        # Remove brackets
+        ']': '',        # Remove brackets
+        '{': '',        # Remove braces
+        '}': '',        # Remove braces
+    }
+
+    # Replace special characters using special_char_map
+    for char, replacement in special_char_map.items():
+        text = text.replace(char, replacement)
+
+    # Remove any remaining disallowed characters 
+    # (keep only alphanumeric and underscore)
+    text = re.sub(r'[^a-zA-Z0-9_]', '', text)
+
+    # Remove multiple consecutive underscores and 
+    # replace with single underscore
[…]
