+++ tests/templates/test_load_csv_files_with_config.py
@@ -1,0 +1,275 @@
+# tests/templates/test_load_csv_template.py
+"""Unit tests for the Load CSV Files template."""
+
+import json
+import os
+import sys
+import tempfile
+import unittest
+from pathlib import Path
+from unittest.mock import patch, MagicMock
+
+import pandas as pd
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.load_csv_files_with_config import run_from_json
+
+
+def create_mock_csv_files(tmp_dir: Path) -> tuple:
+    """Create minimal CSV files for testing."""
+    # Create first CSV file
+    csv1_data = pd.DataFrame({
+        'CellID': [1, 2, 3],
+        'X_centroid': [10.0, 20.0, 30.0],
+        'Y_centroid': [15.0, 25.0, 35.0],
+        'cell_type': ['TypeA', 'TypeB', 'TypeA']
+    })
+    csv1_path = tmp_dir / 'sample1.csv'
+    csv1_data.to_csv(csv1_path, index=False)
+
+    # Create second CSV file
+    csv2_data = pd.DataFrame({
+        'CellID': [4, 5, 6],
+        'X_centroid': [40.0, 50.0, 60.0],
+        'Y_centroid': [45.0, 55.0, 65.0],
+        'cell_type': ['TypeB', 'TypeB', 'TypeA']
+    })
+    csv2_path = tmp_dir / 'sample2.csv'
+    csv2_data.to_csv(csv2_path, index=False)
+
+    # Create configuration file
+    config_data = pd.DataFrame({
+        'file_name': ['sample1.csv', 'sample2.csv'],
+        'slide_number': ['S1', 'S2']
+    })
+    config_path = tmp_dir / 'config.csv'
+    config_data.to_csv(config_path, index=False)
+
+    return csv1_path, csv2_path, config_path
+
+
+class TestLoadCSVFilesWithConfig(unittest.TestCase):
+    """Unit tests for the Load CSV Files template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.tmp_path = Path(self.tmp_dir.name)
+
+        # Create mock CSV files
+        self.csv1, self.csv2, self.config = create_mock_csv_files(
+            self.tmp_path
+        )
+
+        # Minimal parameters
+        self.params = {
+            "CSV_Files": str(self.tmp_path),
+            "CSV_Files_Configuration": str(self.config),
+            "String_Columns": [""],
+            "Output_File": "combined.csv"
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering all input/output scenarios."""
+        # Test 1: Run with save_results=True
+        saved_files = run_from_json(self.params)
+        self.assertIn("combined.csv", saved_files)
+        output_path = Path(saved_files["combined.csv"])
+        self.assertTrue(output_path.exists())
+
+        # Verify content
+        result_df = pd.read_csv(output_path)
+        self.assertEqual(len(result_df), 6)  # 3 + 3 rows
+        self.assertIn('file_name', result_df.columns)
+        self.assertIn('slide_number', result_df.columns)
+        self.assertIn('CellID', result_df.columns)
+
+        # Test 2: Run with save_results=False
+        df_result = run_from_json(self.params, save_results=False)
+        self.assertIsInstance(df_result, pd.DataFrame)
+        self.assertEqual(len(df_result), 6)
+
+        # Test 3: JSON file input
+        json_path = self.tmp_path / "params.json"
+        with open(json_path, "w") as f:
[…]
