### 2025-07-28 bbfa2f6
fix(template_utils): Use applymap instead of map for pandas compatibility
