### 2025-07-28 1cfb39e
feat(setup_analysis_template): add setup_analysis_template function and unit tests
