+++ src/spac/templates/setup_analysis_template.py
@@ -1,0 +1,158 @@
+"""
+Platform-agnostic Setup Analysis template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.setup_analysis_template import run_from_json
+>>> run_from_json("examples/setup_analysis_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, List, Optional
+import pandas as pd
+import logging
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.data_utils import ingest_cells
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True
+) -> Union[Dict[str, str], Any]:
+    """
+    Execute Setup Analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. If False, returns the adata object
+        directly for in-memory workflows. Default is True.
+
+    Returns
+    -------
+    dict or AnnData
+        If save_results=True: Dictionary of saved file paths
+        If save_results=False: The processed AnnData object
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Extract parameters
+    upstream_dataset = params["Upstream_Dataset"]
+    feature_names = params["Features_to_Analyze"]
+    regex_str = params.get("Feature_Regex", [])
+    x_col = params["X_Coordinate_Column"]
+    y_col = params["Y_Coordinate_Column"]
+    annotation = params["Annotation_s_"]
+
+    # Load upstream data - could be DataFrame or CSV
+    if isinstance(upstream_dataset, (str, Path)):
+        # Check if it's a pickle/h5ad file or CSV
+        path = Path(upstream_dataset)
+        if path.suffix.lower() in ['.pickle', '.pkl', '.p', '.h5ad']:
+            input_dataset = load_input(upstream_dataset)
+        else:
+            # Assume it's a CSV file
+            input_dataset = pd.read_csv(upstream_dataset)
+    else:
+        # Already a DataFrame
+        input_dataset = upstream_dataset
+
+    # Process annotation parameter
+    if isinstance(annotation, str):
+        annotation = [annotation]
+
+    if len(annotation) == 1 and annotation[0] == "None":
+        annotation = None
+
+    if annotation and len(annotation) != 1 and "None" in annotation:
+        error_msg = 'String "None" found in the annotation list'
+        raise ValueError(error_msg)
+
+    # Process coordinate columns
+    x_col = text_to_value(x_col, default_none_text="None")
+    y_col = text_to_value(y_col, default_none_text="None")
+
+    # Process feature names and regex
+    if isinstance(feature_names, str):
+        feature_names = [feature_names]
+    if isinstance(regex_str, str):
+        import ast
+        try:
+            regex_str = ast.literal_eval(regex_str)
+        except (ValueError, SyntaxError):
+            regex_str = [regex_str] if regex_str else []
+
+    # Processing two search methods
+    for feature in feature_names:
+        regex_str.append(f"^{feature}$")
+
+    # Sanitizing search list
+    regex_str_set = set(regex_str)
+    regex_str_list = list(regex_str_set)
+
+    # Run the ingestion
+    ingested_anndata = ingest_cells(
+        dataframe=input_dataset,
+        regex_str=regex_str_list,
[…]
