### 2025-07-28 dddc33a
fix(setup_analysis_template): fix setup_analysis_template function
