### 2025-07-28 eb810ab
feat(boxplot_template): add boxplot_template function and unit tests
