+++ tests/templates/test_boxplot_template.py
@@ -1,0 +1,422 @@
+# tests/templates/test_boxplot_template.py
+"""Unit tests for the Boxplot template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.boxplot_template import run_from_json
+
+
+def mock_adata_for_boxplot(n_cells: int = 20) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for boxplot tests."""
+    rng = np.random.default_rng(0)
+    
+    # Create observations with annotations
+    obs = pd.DataFrame({
+        "cell_type": ["B cells", "T cells"] * (n_cells // 2),
+        "condition": ["Control", "Treatment"] * (n_cells // 2)
+    })
+    
+    # Create expression matrix with 3 features
+    n_features = 3
+    x_mat = rng.poisson(lam=5, size=(n_cells, n_features)) + 1.0
+    
+    # Create var dataframe with feature names
+    var = pd.DataFrame(
+        index=[f"Gene_{i}" for i in range(n_features)]
+    )
+    
+    adata = ad.AnnData(X=x_mat, obs=obs, var=var)
+    
+    # Add a normalized layer
+    adata.layers["normalized"] = np.log1p(adata.X)
+    
+    return adata
+
+
+class TestBoxplotTemplate(unittest.TestCase):
+    """Unit tests for the Boxplot template."""
+
+    def _create_mock_boxplot_return(self, df_data=None):
+        """Helper to create standard mock returns."""
+        mock_fig = MagicMock()
+        mock_ax = MagicMock()
+        if df_data is None:
+            df_data = {'Gene_0': [1]}
+        mock_df = pd.DataFrame(df_data)
+        return mock_fig, mock_ax, mock_df
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input_data.pickle"
+        )
+        self.out_file = "boxplot_summary.csv"
+
+        # Save minimal mock data as pickle
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata_for_boxplot(), f)
+
+        # Minimal parameters
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Primary_Annotation": "cell_type",
+            "Secondary_Annotation": "None",
+            "Table_to_Visualize": "Original",
+            "Feature_s_to_Plot": ["All"],
+            "Value_Axis_Log_Scale": False,
+            "Figure_Title": "BoxPlot",
+            "Horizontal_Plot": False,
+            "Figure_Width": 12,
+            "Figure_Height": 8,
+            "Figure_DPI": 300,
+            "Font_Size": 10,
+            "Keep_Outliers": True,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.boxplot_template.boxplot')
+    @patch('matplotlib.pyplot.show')
+    def test_run_with_save(self, mock_show, mock_boxplot) -> None:
+        """Test boxplot with file saving."""
+        # Mock the boxplot function to return figure, ax, and dataframe
+        mock_fig, mock_ax, mock_df = self._create_mock_boxplot_return({
[…]
