### 2025-07-29 cd5abd0
fix(boxplot_template): address minor comments from copilot
