### 2025-07-31 3380427
feat(histogram_template): add histogram_template and unit tests
