+++ tests/templates/test_histogram_template.py
@@ -1,0 +1,222 @@
+# tests/templates/test_histogram_template.py
+"""Unit tests for the Histogram template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.histogram_template import run_from_json
+
+
+def mock_adata_with_features(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    
+    # Simple expression data
+    x_mat = rng.normal(size=(n_cells, 3))
+    
+    # Simple observations
+    obs = pd.DataFrame({
+        "cell_type": ["TypeA", "TypeB"] * (n_cells // 2)
+    })
+    
+    # Simple var names
+    var = pd.DataFrame(index=["Gene1", "Gene2", "Gene3"])
+    
+    return ad.AnnData(X=x_mat, obs=obs, var=var)
+
+
+class TestHistogramTemplate(unittest.TestCase):
+    """Unit tests for the Histogram template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "plots.csv"
+
+        # Save minimal mock data as pickle
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata_with_features(), f)
+
+        # Minimal parameters
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Plot_By": "Annotation",
+            "Annotation": "cell_type",
+            "Feature": "None",
+            "Table_": "Original",
+            "Group_by": "None",
+            "Together": True,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.histogram_template.histogram')
+    @patch('seaborn.move_legend')  # Mock seaborn to avoid legend issues
+    def test_run_with_save(self, mock_move_legend, mock_histogram) -> None:
+        """Test basic run with file saving."""
+        # Mock the histogram function
+        mock_fig = MagicMock()
+        mock_fig.number = 1
+        mock_fig.set_size_inches = MagicMock()
+        mock_fig.set_dpi = MagicMock()
+        
+        mock_ax = MagicMock()
+        mock_ax.get_legend.return_value = None
+        mock_ax.tick_params = MagicMock()
+        mock_ax.set_title = MagicMock()
+        
+        mock_df = pd.DataFrame({
+            'category': ['TypeA', 'TypeB'],
+            'count': [5, 5]
+        })
+        
+        mock_histogram.return_value = {
+            "fig": mock_fig,
+            "axs": mock_ax,
+            "df": mock_df
+        }
+        
+        # Run with save_results=True (default)
+        saved_files = run_from_json(self.params)
+        
+        # Check that file was saved
+        self.assertIn(self.out_file, saved_files)
+        self.assertTrue(os.path.exists(saved_files[self.out_file]))
+        
+        # Verify histogram was called
+        mock_histogram.assert_called_once()
+
+    @patch('spac.templates.histogram_template.histogram')
+    @patch('seaborn.move_legend')
[…]
