+++ src/spac/templates/histogram_template.py
@@ -1,0 +1,282 @@
+"""
+Platform-agnostic Histogram template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.histogram_template import run_from_json
+>>> run_from_json("examples/histogram_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, Optional, Tuple
+import pandas as pd
+import matplotlib.pyplot as plt
+import seaborn as sns
+import warnings
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.visualization import histogram
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True,
+    show_plot: bool = False
+) -> Union[Dict[str, str], Tuple[Any, pd.DataFrame]]:
+    """
+    Execute Histogram analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. If False, returns the figure and
+        dataframe directly for in-memory workflows. Default is True.
+    show_plot : bool, optional
+        Whether to display the plot. Default is False.
+
+    Returns
+    -------
+    dict or tuple
+        If save_results=True: Dictionary of saved file paths
+        If save_results=False: Tuple of (figure, dataframe)
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Load the upstream analysis data
+    adata = load_input(params["Upstream_Analysis"])
+
+    # Extract parameters
+    feature = text_to_value(params.get("Feature", "None"))
+    annotation = text_to_value(params.get("Annotation", "None"))
+    layer = params.get("Table_", "Original")
+    group_by = params.get("Group_by", "None")
+    together = params.get("Together", True)
+    fig_width = params.get("Figure_Width", 8)
+    fig_height = params.get("Figure_Height", 6)
+    font_size = params.get("Font_Size", 12)
+    fig_dpi = params.get("Figure_DPI", 300)
+    legend_location = params.get("Legend_Location", "best")
+    legend_in_figure = params.get("Legend_in_Figure", False)
+    take_X_log = params.get("Take_X_Log", False)
+    take_Y_log = params.get("Take_Y_log", False)
+    multiple = params.get("Multiple", "dodge")
+    shrink = params.get("Shrink_Number", 1)
+    bins = params.get("Bins", "auto")
+    alpha = params.get("Bin_Transparency", 0.75)
+    stat = params.get("Stat", "count")
+    x_rotate = params.get("X_Axis_Label_Rotation", 0)
+    histplot_by = params.get("Plot_By", "Annotation")
+
+    # Close all existing figures to prevent extra plots
+    plt.close('all')
+    existing_fig_nums = plt.get_fignums()
+
+    plt.rcParams.update({'font.size': font_size})
+
+    # Adjust feature and annotation based on histplot_by
+    if histplot_by == "Annotation":
+        feature = None
+    else:
+        annotation = None
+
+    # If both feature and annotation are None, set default
+    if feature is None and annotation is None:
+        if histplot_by == "Annotation":
+            if adata.obs.columns.size > 0:
+                annotation = adata.obs.columns[0]
+                print(
+                    f'No annotation specified. Using the first annotation '
+                    f'"{annotation}" as default.'
+                )
+            else:
[…]
