### 2025-07-31 51ba1c4
fix(histogram_template): fix odd number of cells in test
