### 2025-08-02 0f26c08
feat(spatial_plot_temp): add spatial_plot_template.py and unit tests
