+++ tests/templates/test_spatial_plot_template.py
@@ -1,0 +1,270 @@
+# tests/templates/test_spatial_plot_template.py
+"""Unit tests for the Spatial Plot template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock, call
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.spatial_plot_template import run_from_json
+
+
+def mock_adata_spatial(n_cells: int = 100) -> ad.AnnData:
+    """Return a minimal synthetic AnnData with spatial coordinates."""
+    rng = np.random.default_rng(0)
+    
+    # Create expression matrix
+    n_features = 5
+    X = rng.normal(size=(n_cells, n_features))
+    
+    # Create annotations
+    obs = pd.DataFrame({
+        "cell_type": rng.choice(["TypeA", "TypeB", "TypeC"], n_cells),
+        "region": rng.choice(["Region1", "Region2"], n_cells),
+        "slide": rng.choice(["Slide1", "Slide2"], n_cells)
+    })
+    
+    # Create spatial coordinates
+    spatial_coords = rng.random((n_cells, 2)) * 1000
+    
+    # Create AnnData object
+    adata = ad.AnnData(X=X, obs=obs)
+    adata.obsm["spatial"] = spatial_coords
+    
+    # Add feature names
+    adata.var_names = [f"Feature_{i}" for i in range(n_features)]
+    
+    return adata
+
+
+class TestSpatialPlotTemplate(unittest.TestCase):
+    """Unit tests for the Spatial Plot template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "spatial_data.pickle"
+        )
+        self.out_file = "spatial_plot"
+
+        # Save minimal mock data with spatial info
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata_spatial(), f)
+
+        # Minimal parameters
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Annotation_to_Highlight": "cell_type",
+            "Feature_to_Highlight": "",
+            "Table": "Original",
+            "Dot_Transparency": 0.5,
+            "Dot_Size": 25,
+            "Figure_Height": 6,
+            "Figure_Width": 12,
+            "Figure_DPI": 200,
+            "Font_Size": 12,
+            "Lower_Colorbar_Bound": 999,
+            "Upper_Colorbar_Bound": -999,
+            "Color_By": "Annotation",
+            "Stratify": False,
+            "Stratify_By": [],
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.spatial_plot_template.spatial_plot')
+    @patch('matplotlib.pyplot.show')
+    def test_run_single_plot_annotation(self, mock_show, mock_spatial) -> None:
+        """Test single plot with annotation coloring."""
+        # Mock the spatial_plot function to return axes
+        mock_ax = MagicMock()
+        mock_spatial.return_value = [mock_ax]
+        
+        # Test with annotation coloring
+        saved_files = run_from_json(self.params)
+        
+        # Verify spatial_plot was called correctly
+        mock_spatial.assert_called_once()
+        call_kwargs = mock_spatial.call_args[1]
[…]
