### 2025-08-02 028a049
fix(spatial_plot_temp): addrss copilot comments spatial_plot_template.py
