### 2025-08-02 ff6cce4
feat(arcsinh_normalization_template): add arcsinh_normalization_template function and unit tests
