### 2025-08-02 afca7ff
fix(test_arcsinh_normalization_template): Handle odd numbers with better list slicing
