+++ tests/templates/test_zscore_normalization_template.py
@@ -1,0 +1,211 @@
+# tests/templates/test_zscore_normalization_template.py
+"""Unit tests for the Z-Score Normalization template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.zscore_normalization_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "cell_type": (["TypeA", "TypeB"] * ((n_cells + 1) // 2))[:n_cells],
+        "sample": (["S1", "S2"] * ((n_cells + 1) // 2))[:n_cells]
+    })
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = ["Marker1", "Marker2", "Marker3"]
+    
+    # Add a test layer with different values
+    adata.layers["test_layer"] = rng.normal(loc=5, scale=2, size=(n_cells, 3))
+    
+    return adata
+
+
+class TestZScoreNormalizationTemplate(unittest.TestCase):
+    """Unit tests for the Z-Score Normalization template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "normalized_output"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters from JSON template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Table_to_Process": "Original",
+            "Output_Table_Name": "z_scores",
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.zscore_normalization_template.'
+           'z_score_normalization')
+    def test_complete_io_workflow(self, mock_z_score_norm) -> None:
+        """Single I/O test covering input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Mock the z_score_normalization function
+            def mock_z_score_side_effect(adata, **kwargs):
+                # Simulate what z_score_normalization does
+                output_layer = kwargs.get('output_layer', 'z_scores')
+                input_layer = kwargs.get('input_layer', None)
+                
+                # Create normalized data
+                if input_layer is None:
+                    data = adata.X
+                else:
+                    data = adata.layers[input_layer]
+                
+                # Simple z-score normalization simulation
+                normalized = (data - np.mean(data, axis=0)) / np.std(data, axis=0)
+                adata.layers[output_layer] = normalized
+                return None
+            
+            mock_z_score_norm.side_effect = mock_z_score_side_effect
+            
+            # Test 1: Run with default parameters (Original layer)
+            result = run_from_json(self.params)
+            self.assertIsInstance(result, dict)
+            
+            # Verify z_score_normalization was called correctly
+            mock_z_score_norm.assert_called_once()
+            call_args = mock_z_score_norm.call_args
+            self.assertEqual(call_args[1]['output_layer'], "z_scores")
+            # Original -> None
[…]
