### 2025-08-02 06d8feb
feat(summarize_dataframe_template): add summarize_dataframe_template function and unit tests
