+++ tests/templates/test_summarize_dataframe_template.py
@@ -1,0 +1,253 @@
+# tests/templates/test_summarize_dataframe_template.py
+"""Unit tests for the Summarize DataFrame template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.summarize_dataframe_template import run_from_json
+
+
+def mock_dataframe(n_rows: int = 10) -> pd.DataFrame:
+    """Return a minimal synthetic DataFrame for fast tests."""
+    rng = np.random.default_rng(0)
+    df = pd.DataFrame({
+        "file_name": [f"file_{i}.csv" for i in range(n_rows)],
+        "CD3D": rng.random(n_rows) * 100,
+        "FOXP3": rng.random(n_rows) * 50,
+        "PDL1": rng.random(n_rows) * 75,
+    })
+    # Add some NaN values for testing
+    df.loc[2, "CD3D"] = np.nan
+    df.loc[5, "FOXP3"] = np.nan
+    return df
+
+
+class TestSummarizeDataFrameTemplate(unittest.TestCase):
+    """Unit tests for the Summarize DataFrame template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.csv"
+        )
+        self.out_file = "summary"
+
+        # Save minimal mock data
+        mock_dataframe().to_csv(self.in_file, index=False)
+
+        # Minimal parameters - from the example
+        self.params = {
+            "Calculate_Centroids": self.in_file,
+            "Columns": ["file_name", "CD3D", "FOXP3", "PDL1"],
+            "Print_Missing_Location": True,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Test 1: Run with default parameters
+            with patch('spac.templates.summarize_dataframe_template.'
+                      'summarize_dataframe') as mock_summarize:
+                with patch('spac.templates.summarize_dataframe_template.'
+                          'present_summary_as_figure') as mock_fig:
+                    # Mock the summary dataframe
+                    mock_summarize.return_value = pd.DataFrame(
+                        {"test": [1, 2, 3]}
+                    )
+                    
+                    # Mock the Plotly figure
+                    mock_plotly_fig = MagicMock()
+                    mock_plotly_fig.write_html = MagicMock()
+                    mock_plotly_fig.show = MagicMock()
+                    mock_fig.return_value = mock_plotly_fig
+                    
+                    result = run_from_json(self.params)
+                    self.assertIsInstance(result, dict)
+                    self.assertIn("summary.html", result)
+                    
+                    # Verify figure methods were called
+                    mock_plotly_fig.show.assert_called_once()
+                    mock_plotly_fig.write_html.assert_called_once_with(
+                        "summary.html"
+                    )
+            
+            # Test 2: Run without saving
+            with patch('spac.templates.summarize_dataframe_template.'
+                      'summarize_dataframe') as mock_summarize:
+                with patch('spac.templates.summarize_dataframe_template.'
+                          'present_summary_as_figure') as mock_fig:
+                    mock_summarize.return_value = pd.DataFrame(
+                        {"test": [1, 2, 3]}
+                    )
+                    mock_plotly_fig = MagicMock()
[…]
