### 2025-08-02 30118f9
fix(summarize_dataframe_template): address comments from copilot
