### 2025-08-02 829a4bd
feat(combine_annotations_template): add combine_annotations_template function and unit tests
