+++ tests/templates/test_combine_annotations_template.py
@@ -1,0 +1,210 @@
+# tests/templates/test_combine_annotations_template.py
+"""Unit tests for the Combine Annotations template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.combine_annotations_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "detailed_cell_type": (
+            ["B_cell", "T_cell", "NK_cell"] * ((n_cells + 2) // 3)
+        )[:n_cells],
+        "broad_cell_type": (
+            ["Immune", "Immune", "Immune"] * ((n_cells + 2) // 3)
+        )[:n_cells]
+    })
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = ["Gene1", "Gene2", "Gene3"]
+    return adata
+
+
+class TestCombineAnnotationsTemplate(unittest.TestCase):
+    """Unit tests for the Combine Annotations template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "transform_output"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters matching NIDAP template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Annotations_Names": ["detailed_cell_type", "broad_cell_type"],
+            "New_Annotation_Name": "combined_annotation",
+            "Separator": "_",
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering all input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Test 1: Run with default parameters
+            result = run_from_json(self.params)
+            self.assertIsInstance(result, dict)
+            # Should have 2 files: pickle and CSV
+            self.assertEqual(len(result), 2)
+            # Check pickle file exists
+            self.assertIn(
+                f"{self.out_file}.pickle", result
+            )
+            # Check CSV file exists
+            self.assertIn("combined_annotation_counts.csv", result)
+            
+            # Test 2: Run without saving
+            adata_result = run_from_json(self.params, save_results=False)
+            # Check we got AnnData back
+            self.assertIsInstance(adata_result, ad.AnnData)
+            # Check new annotation was created
+            self.assertIn("combined_annotation", adata_result.obs.columns)
+            # Check annotation values are correct format
+            sample_val = adata_result.obs["combined_annotation"].iloc[0]
+            self.assertIn("_", sample_val)  # Should have separator
+            
+            # Test 3: JSON file input
+            json_path = os.path.join(self.tmp_dir.name, "params.json")
+            with open(json_path, "w") as f:
+                json.dump(self.params, f)
+            
+            result_json = run_from_json(json_path)
+            self.assertIsInstance(result_json, dict)
[…]
