### 2025-08-02 9d8582a
fix(combine_annotations_template): address comments from copilot CR for combine_annotations_template function
