### 2025-08-02 941d641
feat(manual_phenotyping_template): add manual_phenotyping_template function and unit tests
