+++ tests/templates/test_manual_phenotyping_template.py
@@ -1,0 +1,290 @@
+#!/usr/bin/env python3
+"""Unit tests for the Manual Phenotyping template."""
+
+import json
+import os
+import sys
+import tempfile
+import unittest
+from pathlib import Path
+from unittest.mock import patch, MagicMock
+
+import pandas as pd
+import numpy as np
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.manual_phenotyping_template import run_from_json
+
+
+def create_mock_data_and_phenotypes(tmp_dir: Path) -> tuple:
+    """Create minimal data and phenotypes files for testing."""
+    # Create mock expression data
+    n_cells = 20  # Simple scenario as per rule 1
+    data = pd.DataFrame({
+        'cell_id': [f'cell_{i}' for i in range(n_cells)],
+        'CD3D_expression': np.random.choice([0, 1], n_cells),
+        'CD4_expression': np.random.choice([0, 1], n_cells),
+        'CD8A_expression': np.random.choice([0, 1], n_cells),
+        'FOXP3_expression': np.random.choice([0, 1], n_cells),
+        'CD68_expression': np.random.choice([0, 1], n_cells),
+        'CD20_expression': np.random.choice([0, 1], n_cells),
+        'CD21_expression': np.random.choice([0, 1], n_cells),
+        'CD56_expression': np.random.choice([0, 1], n_cells),
+    })
+    data_path = tmp_dir / 'input_data.csv'
+    data.to_csv(data_path, index=False)
+    
+    # Create phenotypes definition
+    phenotypes = pd.DataFrame({
+        'phenotype_code': [
+            'CD3D+CD4+FOXP3+',
+            'CD3D+CD4+',
+            'CD3D+CD8A+',
+            'CD68+',
+            'CD20+'
+        ],
+        'phenotype_name': [
+            'Regulatory T Cell',
+            'Helper T Cell',
+            'Cytotoxic T Cell',
+            'Macrophage',
+            'B Cell'
+        ]
+    })
+    phenotypes_path = tmp_dir / 'phenotypes.csv'
+    phenotypes.to_csv(phenotypes_path, index=False)
+    
+    return data_path, phenotypes_path
+
+
+class TestManualPhenotypingTemplate(unittest.TestCase):
+    """Unit tests for the Manual Phenotyping template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.tmp_path = Path(self.tmp_dir.name)
+        
+        # Create mock files
+        self.data_path, self.phenotypes_path = \
+            create_mock_data_and_phenotypes(self.tmp_path)
+        
+        # Minimal parameters
+        self.params = {
+            "Upstream_Dataset": str(self.data_path),
+            "Phenotypes_Code": str(self.phenotypes_path),
+            "Classification_Column_Prefix": "",
+            "Classification_Column_Suffix": "_expression",
+            "Allow_Multiple_Phenotypes": True,
+            "Manual_Annotation_Name": "manual_phenotype",
+            "Output_File": "phenotyped_data.csv"
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.manual_phenotyping_template.'
+           'assign_manual_phenotypes')
+    def test_complete_io_workflow(self, mock_assign) -> None:
+        """Single I/O test covering all input/output scenarios (Rule 2)."""
+        # Mock the assign_manual_phenotypes function
+        def mock_assign_func(df, pheno, **kwargs):
+            # Add phenotype column with the correct annotation name
+            annotation_name = kwargs.get('annotation', 'manual_phenotype')
[…]
