### 2025-08-02 e4d61cf
fix(test_manual_phenotyping_temp): address comments of copilot review for unit tests
