### 2025-08-02 67e5a80
feat(hierarchical_heatmap_template): add hierarchical_heatmap_template and unit tests
