+++ tests/templates/test_hierarchical_heatmap_template.py
@@ -1,0 +1,209 @@
+# tests/templates/test_hierarchical_heatmap_template.py
+"""Unit tests for the Hierarchical Heatmap template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.hierarchical_heatmap_template import run_from_json
+
+
+def mock_adata(n_cells: int = 100) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "phenograph_k60_r1": (
+            ["Cluster1", "Cluster2", "Cluster3"] * ((n_cells + 2) // 3)
+        )[:n_cells],
+        "cell_type": (["TypeA", "TypeB"] * ((n_cells + 1) // 2))[:n_cells]
+    })
+    # Create expression data with 9 markers as in the example
+    x_mat = rng.normal(size=(n_cells, 9))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = [
+        "Hif1a", "NOS2", "COX2", "β-catenin", "vimentin", 
+        "E-cadherin", "Ki67", "PIMO", "aSMA"
+    ]
+    # Add a z-score normalized layer for testing
+    adata.layers["arcsinh_z_scores"] = (
+        (x_mat - x_mat.mean(axis=0)) / x_mat.std(axis=0)
+    )
+    return adata
+
+
+class TestHierarchicalHeatmapTemplate(unittest.TestCase):
+    """Unit tests for the Hierarchical Heatmap template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "mean_intensity.csv"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters - from the JSON template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Annotation": "phenograph_k60_r1",
+            "Table_to_Visualize": "arcsinh_z_scores",
+            "Feature_s_": ["All"],
+            "Standard_Scale_": "None",
+            "Z_Score": "None",
+            "Feature_Dendrogram": True,
+            "Annotation_Dendrogram": True,
+            "Figure_Title": "Hierarchical Heatmap",
+            "Figure_Width": 15,
+            "Figure_Height": 12,
+            "Figure_DPI": 300,
+            "Font_Size": 14,
+            "Matrix_Plot_Ratio": 0.8,
+            "Swap_Axes": False,
+            "Rotate_Label_": False,
+            "Horizontal_Dendrogram_Display_Ratio": 0.2,
+            "Vertical_Dendrogram_Display_Ratio": 0.2,
+            "Value_Min": "-3",
+            "Value_Max": "3",
+            "Color_Map": "seismic",
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
[…]
