### 2025-08-02 6466e2f
feat(hierarchical_heatmap_template): add hierarchical_heatmap_template function and unit tests
