### 2025-08-02 743fb10
feat(utag_clustering_template): add utag_clustering_template and unit tests
