### 2025-08-02 6da3985
feat(utag_clustering_template): add utag_clustering_template and unit tests
