### 2025-08-02 e79fd78
feat(umap_transformation_template): add umap_transformation_template function and unit tests
