### 2025-08-02 9d24638
fix(umap_transformation_template): return adata in place and fix comments of copilot
