### 2025-08-03 d67f6c7
feat(umap_tsne_pca_template): add umap_tsne_pca_template function and unit tests
