+++ tests/templates/test_umap_tsne_pca_template.py
@@ -1,0 +1,181 @@
+# tests/templates/test_umap_tsne_pca_template.py
+"""Unit tests for the UMAP\tSNE\PCA Visualization template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.umap_tsne_pca_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "cell_type": (["TypeA", "TypeB"] * ((n_cells + 1) // 2))[:n_cells],
+        "sample": (["S1", "S2"] * ((n_cells + 1) // 2))[:n_cells]
+    })
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = ["Gene1", "Gene2", "Gene3"]
+    # Add UMAP coordinates for visualization
+    adata.obsm["X_umap"] = rng.random((n_cells, 2)) * 10
+    adata.obsm["X_tsne"] = rng.random((n_cells, 2)) * 50
+    adata.obsm["X_pca"] = rng.random((n_cells, 2)) * 5
+    return adata
+
+
+class TestUmapTsnePcaTemplate(unittest.TestCase):
+    """Unit tests for the UMAP\tSNE\PCA Visualization template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "visualization"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters - adjust based on template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Annotation_to_Highlight": "cell_type",
+            "Feature_to_Highlight": "None",
+            "Table": "Original",
+            "Dimension_Reduction_Method": "umap",
+            "Figure_Width": 12,
+            "Figure_Height": 12,
+            "Font_Size": 12,
+            "Figure_DPI": 300,
+            "Legend_Location": "best",
+            "Legend_Font_Size": 16,
+            "Legend_Marker_Size": 5.0,
+            "Color_By": "Annotation",
+            "Dot_Size": 1,
+            "Value_Min": "None",
+            "Value_Max": "None",
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Mock the dimensionality_reduction_plot function
+            with patch('spac.templates.umap_tsne_pca_template.'
+                      'dimensionality_reduction_plot') as mock_plot:
+                # Create mock figure and axis
+                mock_fig = MagicMock()
+                mock_ax = MagicMock()
+                mock_ax.get_figure.return_value = mock_fig
+                mock_ax.get_legend.return_value = MagicMock()
+                mock_fig.savefig = MagicMock()
+                
[…]
