### 2025-08-03 7fb9b2c
fix(umap_tsne_pca_template): address the comments from copilot
