### 2025-08-03 c57075d
feat(relational_heatmap_template): add relational_heatmap_template function and unit tests
