+++ tests/templates/test_relational_heatmap_template.py
@@ -1,0 +1,214 @@
+# tests/templates/test_relational_heatmap_template.py
+"""Unit tests for the Relational Heatmap template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock, Mock
+
+import matplotlib
+import matplotlib.pyplot as plt
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.relational_heatmap_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "phenograph_k60_r1": (["cluster1", "cluster2", "cluster3"] * 
+                             ((n_cells + 2) // 3))[:n_cells],
+        "renamed_phenotypes": (["phenotype_A", "phenotype_B"] * 
+                              ((n_cells + 1) // 2))[:n_cells]
+    })
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = ["Gene1", "Gene2", "Gene3"]
+    return adata
+
+
+class TestRelationalHeatmapTemplate(unittest.TestCase):
+    """Unit tests for the Relational Heatmap template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "relational_heatmap"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters from NIDAP template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Source_Annotation_Name": "phenograph_k60_r1",
+            "Target_Annotation_Name": "renamed_phenotypes",
+            "Colormap": "darkmint",
+            "Figure_Width_inch": 8,
+            "Figure_Height_inch": 10,
+            "Figure_DPI": 300,
+            "Font_Size": 8,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.relational_heatmap_template.relational_heatmap')
+    @patch('plotly.io.write_image')
+    @patch('matplotlib.pyplot.show')  # Mock plt.show()
+    def test_complete_io_workflow(
+        self, mock_plt_show, mock_write_image, mock_relational
+    ) -> None:
+        """Single I/O test covering input/output scenarios."""
+        # Mock the relational_heatmap function
+        mock_fig = Mock()
+        mock_fig.show = Mock()  # Mock the fig.show() method
+        
+        mock_df = pd.DataFrame({
+            'source': ['cluster1', 'cluster2'],
+            'target': ['phenotype_A', 'phenotype_B'],
+            'value': [5, 3]
+        })
+        
+        mock_relational.return_value = {
+            'file_name': 'relational_heatmap.csv',
+            'data': mock_df,
+            'figure': mock_fig
+        }
+        
+        # Mock the plotly write_image to create a dummy image
+        def create_dummy_image(fig, path, **kwargs):
+            # Create a minimal PNG file
+            fig, ax = plt.subplots(figsize=(1, 1))
[…]
