### 2025-08-03 5662bdb
fix(relational_heatmap_template): address the issue of insecure temporary file and comments from copilot
