### 2025-08-03 34b4eee
feat(sankey_plot_template): add sankey_plot_template function and unit tests
