### 2025-08-03 07baeb9
fix(sankey_plot_template): address the comments from copilot
