### 2025-08-03 824d131
feat(neighborhood_profile_template): add neighborhood_profile_template function and unit tests
