+++ src/spac/templates/neighborhood_profile_template.py
@@ -1,0 +1,237 @@
+"""
+Platform-agnostic Neighborhood Profile template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.neighborhood_profile_template import run_from_json
+>>> run_from_json("examples/neighborhood_profile_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, List, Optional, Tuple
+import pandas as pd
+import numpy as np
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.spatial_analysis import neighborhood_profile
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True
+) -> Union[Dict[str, str], Dict[Tuple[str, str], pd.DataFrame]]:
+    """
+    Execute Neighborhood Profile analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. If False, returns the dataframes
+        directly for in-memory workflows. Default is True.
+
+    Returns
+    -------
+    dict
+        If save_results=True: Dictionary of saved file paths
+        If save_results=False: Dictionary of (anchor, neighbor) tuples 
+        to DataFrames
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Load the upstream analysis data
+    adata = load_input(params["Upstream_Analysis"])
+
+    # Extract parameters
+    cell_types_annotation = params["Annotation_of_interest"]
+    bins = params["Bins"]
+    slide_names = params.get("Stratify_By", "None")
+    normalization = None
+    output_table = "neighborhood_profile"
+
+    anchor_neighbor_list = params["Anchor_Neighbor_List"]
+    anchor_neighbor_list = [
+        tuple(map(str.strip, item.split(";"))) 
+        for item in anchor_neighbor_list
+    ]
+
+    # Call the spatial umap calculation
+
+    bins = [float(radius) for radius in bins]
+
+    slide_names = text_to_value(slide_names)
+
+    neighborhood_profile(
+        adata,
+        phenotypes=cell_types_annotation,
+        distances=bins,
+        regions=slide_names,
+        spatial_key="spatial",
+        normalize=normalization,
+        associated_table_name=output_table
+    )
+
+    print(adata)
+    print(adata.obsm[output_table].shape)
+    print(adata.uns[output_table])
+
+    dataframes, filenames = neighborhood_profiles_for_pairs(
+        adata,
+        cell_types_annotation,
+        slide_names,
+        bins,
+        anchor_neighbor_list,
+        output_table
+    )
+
+    # Handle results based on save_results flag
+    if save_results:
+        # Save outputs
+        saved_files = {}
+        
+        for (anchor_label, neighbor_label), filename in zip(
+            dataframes.keys(), filenames
+        ):
+            df = dataframes[(anchor_label, neighbor_label)]
+            saved_files[filename] = df
+
+        # Save all CSV files
+        saved_files = save_outputs(saved_files)
+        
+        for filename in saved_files:
+            print(f"Saved: {filename}")
+        
+        print(
+            f"Neighborhood Profile completed → {len(saved_files)} files"
+        )
+        return saved_files
+    else:
+        # Return the dataframes directly for in-memory workflows
[…]
