### 2025-08-03 448a980
feat(analysis_to_csv_template): add analysis_to_csv_template function and unit tests
