### 2025-08-03 3f4336b
feat(interactive_spatial_plot_template): add interactive_spatial_plot_template function and unit tests
