+++ tests/templates/test_interactive_spatial_plot_template.py
@@ -1,0 +1,200 @@
+# tests/templates/test_interactive_spatial_plot_template.py
+"""Unit tests for the Interactive Spatial Plot template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.interactive_spatial_plot_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "renamed_phenotypes": (
+            ["TypeA", "TypeB"] * ((n_cells + 1) // 2)
+        )[:n_cells],
+        "sample": (["S1", "S2"] * ((n_cells + 1) // 2))[:n_cells]
+    })
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = ["Gene1", "Gene2", "Gene3"]
+    # Add spatial coordinates required for spatial plots
+    adata.obsm["spatial"] = rng.random((n_cells, 2)) * 100
+    # Add color mapping
+    adata.uns["_spac_colors"] = {
+        "TypeA": "#FF0000",
+        "TypeB": "#0000FF"
+    }
+    return adata
+
+
+class TestInteractiveSpatialPlotTemplate(unittest.TestCase):
+    """Unit tests for the Interactive Spatial Plot template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "interactive_plot"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters - adjust based on template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Color_By": "Annotation",
+            "Annotation_s_to_Highlight": ["renamed_phenotypes"],
+            "Feature_to_Highlight": "None",
+            "Table": "Original",
+            "Dot_Size": 3,
+            "Dot_Transparency": 0.75,
+            "Feature_Color_Scale": "balance",
+            "Figure_Width": 12,
+            "Figure_Height": 12,
+            "Figure_DPI": 200,
+            "Font_Size": 12,
+            "Stratify_By": "None",
+            "Define_Label_Color_Mapping": "_spac_colors",
+            "Lower_Colorbar_Bound": 999,
+            "Upper_Colorbar_Bound": -999,
+            "Flip_Vertical_Axis": False,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.interactive_spatial_plot_template.'
+           'interactive_spatial_plot')
+    @patch('plotly.io.to_html')
+    def test_complete_io_workflow(
+        self,
+        mock_to_html,
+        mock_spatial_plot,
+    ) -> None:
+        """Single I/O test covering input/output scenarios."""
+        # Mock the interactive_spatial_plot function
+        mock_fig1 = MagicMock()
+        mock_fig1.show = MagicMock()
+        
+        mock_spatial_plot.return_value = [
+            {
+                'image_name': 'plot_1',
[…]
