### 2025-08-03 64ff302
fix(interactive_spatial_plot_template): remove nidap comments
