### 2025-08-04 a7b1349
feat(spatial_interaction_template): add spatial_interaction_template and unit tests
