+++ src/spac/templates/spatial_interaction_template.py
@@ -1,0 +1,273 @@
+"""
+Platform-agnostic Spatial Interaction template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.spatial_interaction_template import run_from_json
+>>> run_from_json("examples/spatial_interaction_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, List, Optional
+import pandas as pd
+import numpy as np
+from PIL import Image
+from pprint import pprint
+import matplotlib.pyplot as plt
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.spatial_analysis import spatial_interaction
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True,
+    show_plot: bool = True
+) -> Union[Dict[str, str], None]:
+    """
+    Execute Spatial Interaction analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. Default is True.
+    show_plot : bool, optional
+        Whether to display the plot. Default is True.
+
+    Returns
+    -------
+    dict or None
+        If save_results=True: Dictionary of saved file paths containing:
+            - PNG files: Heatmap visualizations of spatial interactions
+            - CSV files: Matrices with interaction scores/counts between cell types
+        If save_results=False: None (plots are displayed only)
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Load the upstream analysis data
+    adata = load_input(params["Upstream_Analysis"])
+
+    # Extract parameters
+    annotation = params["Annotation"]
+    analysis_method = params["Spatial_Analysis_Method"]
+    # Two analysis methods available:
+    # 1. "Neighborhood Enrichment": Calculates how often pairs of cell types
+    #    are neighbors compared to random chance. Positive scores indicate
+    #    attraction/co-location, negative scores indicate avoidance.
+    #    Output: z-scores (can be positive or negative)
+    #    Files: neighborhood_enrichment_{identifier}.csv
+    # 2. "Cluster Interaction Matrix": Counts the number of edges/connections
+    #    between different cell types in the spatial neighborhood graph.
+    #    Shows absolute interaction frequencies rather than enrichment.
+    #    Output: raw counts (always positive integers)
+    #    Files: cluster_interaction_matrix_{identifier}.csv
+    # Both methods produce the same data structure, just different values
+    stratify_by = params.get("Stratify_By", ["None"])
+    seed = params.get("Seed", "None")
+    coord_type = params.get("Coordinate_Type", "None")
+    n_rings = 1
+    n_neighs = params.get("K_Nearest_Neighbors", 6)
+    radius = params.get("Radius", "None")
+    image_width = params.get("Figure_Width", 15)
+    image_height = params.get("Figure_Height", 12)
+    dpi = params.get("Figure_DPI", 200)
+    font_size = params.get("Font_Size", 12)
+    color_bar_range = params.get("Color_Bar_Range", "Automatic")
+
+    def save_matrix(matrix):
+        for file_name in matrix:
+            data_df = matrix[file_name]
+            print("\n")
+            print(file_name)
+            print(data_df)
+            # In SPAC, collect matrices for later saving instead of 
+            # direct file write
+            matrices[file_name] = data_df
+
+    def update_nidap_display(
+        axs,
+        image_width,
[…]
