### 2025-08-04 1a3d03d
fix(spatial_interaction_template): fix typo
