### 2025-08-04 07ecdfa
feat(visualize_nearest_neighbor_template): add visualize_nearest_neighbor_template function and unit tests
