+++ tests/templates/test_visualize_nearest_neighbor_template.py
@@ -1,0 +1,237 @@
+# tests/templates/test_visualize_nearest_neighbor_template.py
+"""Unit tests for the Visualize Nearest Neighbor template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import matplotlib
+matplotlib.use("Agg")  # Headless backend for CI
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.visualize_nearest_neighbor_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    obs = pd.DataFrame({
+        "renamed_phenotypes": (["Tfh", "B_cells"] * ((n_cells + 1) // 2))[:n_cells],
+        "image_id": (["Image1", "Image2"] * ((n_cells + 1) // 2))[:n_cells]
+    })
+    x_mat = rng.normal(size=(n_cells, 3))
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = ["Marker1", "Marker2", "Marker3"]
+    
+    # Add spatial coordinates
+    adata.obsm["spatial"] = rng.random((n_cells, 2)) * 100
+    
+    # Add mock distance matrix with proper structure
+    # Create a mock distance matrix in obsm
+    # For nearest neighbor, we need distances from each phenotype
+    unique_phenotypes = obs["renamed_phenotypes"].unique()
+    n_phenotypes = len(unique_phenotypes)
+    
+    # Create random distance matrix
+    distance_matrix = rng.exponential(scale=10, size=(n_cells, n_phenotypes))
+    distance_df = pd.DataFrame(
+        distance_matrix,
+        columns=[f"distance_to_{pheno}" for pheno in unique_phenotypes],
+        index=adata.obs.index  # Use the same index as adata.obs
+    )
+    adata.obsm["spatial_distance"] = distance_df
+    
+    return adata
+
+
+class TestVisualizeNearestNeighborTemplate(unittest.TestCase):
+    """Unit tests for the Visualize Nearest Neighbor template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "nearest_neighbor_plots.csv"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters - adjust based on template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Annotation": "renamed_phenotypes",
+            "ImageID": "None",
+            "Plot_Method": "numeric",
+            "Plot_Type": "boxen",
+            "Source_Anchor_Cell_Label": "Tfh",
+            "Target_Cell_Label": "All",
+            "Nearest_Neighbor_Associated_Table": "spatial_distance",
+            "Log_Scale": False,
+            "Facet_Plot": False,
+            "X_Axis_Label_Rotation": 0,
+            "Shared_X_Axis_Title_": True,
+            "X_Axis_Title_Font_Size": "None",
+            "Defined_Color_Mapping": "None",
+            "Figure_Width": 12,
+            "Figure_Height": 6,
+            "FIgure_DPI": 300,
+            "Font_Size": 12,
[…]
