+++ src/spac/templates/visualize_nearest_neighbor_template.py
@@ -1,0 +1,429 @@
+"""
+Platform-agnostic Visualize Nearest Neighbor template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.visualize_nearest_neighbor_template import (
+...     run_from_json
+... )
+>>> run_from_json("examples/visualize_nearest_neighbor_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, Tuple, List
+import pandas as pd
+import numpy as np
+from matplotlib.axes import Axes
+import matplotlib.pyplot as plt
+import matplotlib.patches as mpatches
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.visualization import visualize_nearest_neighbor
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True,
+    show_plot: bool = True
+) -> Union[Dict[str, str], Tuple[Any, pd.DataFrame]]:
+    """
+    Execute Visualize Nearest Neighbor analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. If False, returns the figure and
+        dataframe directly for in-memory workflows. Default is True.
+    show_plot : bool, optional
+        Whether to display the plot. Default is True.
+
+    Returns
+    -------
+    dict or tuple
+        If save_results=True: Dictionary of saved file paths
+        If save_results=False: Tuple of (figure(s), dataframe)
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Load the upstream analysis data
+    adata = load_input(params["Upstream_Analysis"])
+
+    # Extract parameters
+    # Use direct dictionary access for required parameters
+    # Will raise KeyError if missing
+    annotation = params["Annotation"]
+    source_label = params["Source_Anchor_Cell_Label"]
+
+    # Use .get() with defaults for optional parameters from JSON template
+    image_id = params.get("ImageID", "None")
+    method = params.get("Plot_Method", "numeric")
+    plot_type = params.get("Plot_Type", "boxen")
+    target_label = params.get("Target_Cell_Label", "All")
+    distance_key = params.get(
+        "Nearest_Neighbor_Associated_Table", "spatial_distance"
+    )
+    log_scale = params.get("Log_Scale", False)
+    facet_plot = params.get("Facet_Plot", False)
+    x_axis_title_rotation = params.get("X_Axis_Label_Rotation", 0)
+    shared_x_axis_title = params.get("Shared_X_Axis_Title_", True)
+    x_axis_title_fontsize = params.get("X_Axis_Title_Font_Size", "None")
+
+    defined_color_map = text_to_value(
+        params.get("Defined_Color_Mapping", "None"),
+        param_name="Define Label Color Mapping"
+    )
+    annotation_colorscale = "rainbow"
+
+    fig_width = params.get("Figure_Width", 12)
+    fig_height = params.get("Figure_Height", 6)
+    fig_dpi = params.get("FIgure_DPI", 300)
+    global_font_size = params.get("Font_Size", 12)
+    fig_title = (
+        f'Nearest Neighbor Distance Distribution Measured from '
+        f'"{source_label}"'
+    )
+
+    image_id = text_to_value(
+        image_id,
[…]
