### 2025-08-04 52a4ee6
fix(visualize_nearest_neighbor_template): fix typo
