### 2025-08-04 2477266
feat(add_pin_color_rule_template): add add_pin_color_rule_template fnction and unit tests
