### 2025-08-04 5e68e02
feat(append_annotation_template): add append_annotation_template function and unit tests
