+++ tests/templates/test_append_annotation_template.py
@@ -1,0 +1,223 @@
+# tests/templates/test_append_annotation_template.py
+"""Unit tests for the Append Annotation template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.append_annotation_template import run_from_json
+
+
+def mock_dataframe(n_rows: int = 10) -> pd.DataFrame:
+    """Return a minimal synthetic DataFrame for fast tests."""
+    data = {
+        "cell_id": list(range(n_rows)),
+        "intensity": [i * 0.5 for i in range(n_rows)],
+        "existing_annotation": (["TypeA", "TypeB"] * 
+                               ((n_rows + 1) // 2))[:n_rows]
+    }
+    return pd.DataFrame(data)
+
+
+class TestAppendAnnotationTemplate(unittest.TestCase):
+    """Unit tests for the Append Annotation template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.csv"
+        )
+        self.out_file = "append_observations.csv"
+
+        # Save minimal mock data as CSV
+        mock_df = mock_dataframe()
+        mock_df.to_csv(self.in_file, index=False)
+
+        # Minimal parameters from NIDAP template
+        self.params = {
+            "Upstream_Dataset": self.in_file,
+            "Annotation_Pair_List": ["region:region-A", "day:day1"],
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Test 1: Run with CSV input and save results
+            result = run_from_json(self.params)
+            self.assertIsInstance(result, dict)
+            self.assertIn(self.out_file, result)
+            
+            # Verify the output file exists
+            output_path = result[self.out_file]
+            self.assertTrue(os.path.exists(output_path))
+            
+            # Load and verify content
+            output_df = pd.read_csv(output_path)
+            self.assertIn("region", output_df.columns)
+            self.assertIn("day", output_df.columns)
+            self.assertTrue(all(output_df["region"] == "region-A"))
+            self.assertTrue(all(output_df["day"] == "day1"))
+            
+            # Test 2: Run without saving
+            result_no_save = run_from_json(self.params, save_results=False)
+            # Check appropriate return type
+            self.assertIsInstance(result_no_save, pd.DataFrame)
+            self.assertIn("region", result_no_save.columns)
+            self.assertIn("day", result_no_save.columns)
+            
+            # Test 3: JSON file input
+            json_path = os.path.join(self.tmp_dir.name, "params.json")
+            with open(json_path, "w") as f:
+                json.dump(self.params, f)
+            
+            result_json = run_from_json(json_path)
+            self.assertIsInstance(result_json, dict)
+
+    def test_error_validation(self) -> None:
+        """Test exact error message for invalid parameters."""
+        # Test invalid annotation pair format (missing colon)
+        params_bad = self.params.copy()
+        params_bad["Annotation_Pair_List"] = ["invalidformat"]
+        
+        with self.assertRaises(ValueError) as context:
+            run_from_json(params_bad, save_results=False)
+        
+        # The error will come from the split operation
+        self.assertIn("not enough values to unpack", str(context.exception))
+
[…]
