### 2025-08-04 8e500ec
feat(binary_to_categorical_annotation_template): add binary_to_categorical_annotation_template function and unit tests
