### 2025-08-04 4fea9c3
feat(calculate_centroid_template): add calculate_centroid_template function and unit tests
