### 2025-08-04 47adf3e
feat(downsample_cells_template): add downsample_cells_template function and unit tests
