+++ tests/templates/test_downsample_cells_template.py
@@ -1,0 +1,161 @@
+# tests/templates/test_downsample_cells_template.py
+"""Unit tests for the Downsample Cells template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import pandas as pd
+import numpy as np
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.downsample_cells_template import run_from_json
+
+
+def mock_dataframe(n_rows: int = 1000) -> pd.DataFrame:
+    """Return a minimal synthetic DataFrame for fast tests."""
+    rng = np.random.default_rng(0)
+    
+    # Create a dataframe with multiple annotations for downsampling
+    data = {
+        "cell_id": range(n_rows),
+        "region": rng.choice(["region1", "region2", "region3"], n_rows),
+        "day": rng.choice(["day1", "day2", "day3", "day4"], n_rows),
+        "cell_type": rng.choice(["TypeA", "TypeB", "TypeC"], n_rows),
+        "marker1": rng.normal(100, 15, n_rows),
+        "marker2": rng.normal(50, 10, n_rows),
+    }
+    
+    return pd.DataFrame(data)
+
+
+class TestDownsampleCellsTemplate(unittest.TestCase):
+    """Unit tests for the Downsample Cells template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.csv"
+        )
+        self.out_file = "downsampled_data"
+
+        # Save minimal mock data as CSV
+        mock_df = mock_dataframe()
+        mock_df.to_csv(self.in_file, index=False)
+
+        # Minimal parameters from NIDAP template
+        self.params = {
+            "Upstream_Dataset": self.in_file,
+            "Annotations_List": ["region", "day"],
+            "Number_of_Samples": 100,
+            "Stratify_Option": False,
+            "Random_Selection": True,
+            "New_Combined_Annotation_Name": "_combined_",
+            "Minimum_Threshold": 5,
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    def test_complete_io_workflow(self) -> None:
+        """Single I/O test covering input/output scenarios."""
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Test 1: Run with default parameters
+            result = run_from_json(self.params)
+            self.assertIsInstance(result, dict)
+            
+            # Verify file was saved with correct extension
+            self.assertTrue(len(result) > 0)
+            saved_file = list(result.values())[0]
+            self.assertTrue(saved_file.endswith('.csv'))
+            
+            # Test 2: Run without saving
+            result_no_save = run_from_json(self.params, save_results=False)
+            # Check appropriate return type
+            self.assertIsInstance(result_no_save, pd.DataFrame)
+            # Verify downsampling occurred
+            self.assertLessEqual(len(result_no_save), 1000)
+            
+            # Test 3: JSON file input
+            json_path = os.path.join(self.tmp_dir.name, "params.json")
+            with open(json_path, "w") as f:
+                json.dump(self.params, f)
+            
+            result_json = run_from_json(json_path)
+            self.assertIsInstance(result_json, dict)
+
+    def test_different_input_formats(self) -> None:
+        """Test loading from different file formats."""
+        # Test pickle input
+        pickle_file = os.path.join(self.tmp_dir.name, "input.pickle")
+        mock_df = mock_dataframe(500)
[…]
