### 2025-08-04 1db00a8
feat(subset_analysis_template): add subset_analysis_template function and unit tests
