### 2025-08-04 aefcb29
fix(subset_analysis_template): fix typo and enhance function
