### 2025-08-04 542f985
feat(quantile_scaling_template): refactor nidap code, add quantile_scaling_template function and unit tests
