+++ tests/templates/test_quantile_scaling_template.py
@@ -1,0 +1,210 @@
+# tests/templates/test_quantile_scaling_template.py
+"""Unit tests for the Quantile Scaling template."""
+
+import json
+import os
+import pickle
+import sys
+import tempfile
+import unittest
+import warnings
+from unittest.mock import patch, MagicMock
+
+import anndata as ad
+import numpy as np
+import pandas as pd
+from pathlib import Path
+
+sys.path.append(
+    os.path.dirname(os.path.realpath(__file__)) + "/../../src"
+)
+
+from spac.templates.quantile_scaling_template import run_from_json
+
+
+def mock_adata(n_cells: int = 10) -> ad.AnnData:
+    """Return a minimal synthetic AnnData for fast tests."""
+    rng = np.random.default_rng(0)
+    # Create expression data with variance for normalization
+    x_mat = rng.exponential(scale=5, size=(n_cells, 5))
+    obs = pd.DataFrame({
+        "cell_type": (["TypeA", "TypeB"] * ((n_cells + 1) // 2))[:n_cells],
+        "batch": (["Batch1", "Batch2"] * ((n_cells + 1) // 2))[:n_cells]
+    })
+    adata = ad.AnnData(X=x_mat, obs=obs)
+    adata.var_names = [f"Marker{i}" for i in range(5)]
+    # Add an existing layer for testing
+    adata.layers["arcsinh"] = np.arcsinh(x_mat / 5)
+    return adata
+
+
+class TestQuantileScalingTemplate(unittest.TestCase):
+    """Unit tests for the Quantile Scaling template."""
+
+    def setUp(self) -> None:
+        self.tmp_dir = tempfile.TemporaryDirectory()
+        self.in_file = os.path.join(
+            self.tmp_dir.name, "input.pickle"
+        )
+        self.out_file = "output"
+
+        # Save minimal mock data
+        with open(self.in_file, 'wb') as f:
+            pickle.dump(mock_adata(), f)
+
+        # Minimal parameters from NIDAP template
+        self.params = {
+            "Upstream_Analysis": self.in_file,
+            "Table_to_Process": "arcsinh",
+            "Low_Quantile": "0.02",
+            "High_Quantile": "0.98",
+            "Interpolation": "nearest",
+            "Output_Table_Name": "scaled_arcsinh",
+            "Per_Batch": "False",
+            "Annotation": "",
+            "Output_File": self.out_file,
+        }
+
+    def tearDown(self) -> None:
+        self.tmp_dir.cleanup()
+
+    @patch('spac.templates.quantile_scaling_template.normalize_features')
+    def test_complete_io_workflow(self, mock_normalize) -> None:
+        """Single I/O test covering input/output scenarios."""
+        # Mock the normalize_features function
+        def mock_transform(adata, **kwargs):
+            # Simulate normalization by adding a layer
+            output_layer = kwargs['output_layer']
+            input_layer = kwargs.get('input_layer')
+            data = adata.layers[input_layer] if input_layer else adata.X
+            # Simple quantile scaling simulation
+            adata.layers[output_layer] = (
+                (data - np.min(data)) / (np.max(data) - np.min(data))
+            )
+            return adata
+
+        mock_normalize.side_effect = mock_transform
+
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            
+            # Test 1: Run with default parameters
+            result = run_from_json(self.params)
+            self.assertIsInstance(result, dict)
+            # Verify file was saved
+            self.assertTrue(len(result) > 0)
+            pickle_files = [f for f in result.values() if '.pickle' in str(f)]
+            self.assertTrue(len(pickle_files) > 0)
+            
+            # Test 2: Run without saving
+            result_no_save = run_from_json(
[…]
