+++ src/spac/templates/quantile_scaling_template.py
@@ -1,0 +1,262 @@
+"""
+Platform-agnostic Quantile Scaling template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.quantile_scaling_template import run_from_json
+>>> run_from_json("examples/quantile_scaling_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, Tuple
+import pandas as pd
+import plotly.graph_objects as go
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.transformations import normalize_features
+from spac.templates.template_utils import (
+    load_input,
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True,
+    show_plot: bool = True
+) -> Union[Dict[str, str], Tuple[Any, pd.DataFrame]]:
+    """
+    Execute Quantile Scaling analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. If False, returns the adata object
+        and figure directly for in-memory workflows. Default is True.
+    show_plot : bool, optional
+        Whether to display the plot. Default is True.
+
+    Returns
+    -------
+    dict or tuple
+        If save_results=True: Dictionary of saved file paths
+        If save_results=False: Tuple of (adata, figure)
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Load the upstream analysis data
+    adata = load_input(params["Upstream_Analysis"])
+
+    # Extract parameters using .get() with defaults from JSON template
+    low_quantile = params.get("Low_Quantile", "0.02")
+    high_quantile = params.get("High_Quantile", "0.98")
+    interpolation = params.get("Interpolation", "nearest")
+    input_layer = params.get("Table_to_Process", "Original")
+    output_layer = params.get("Output_Table_Name", "normalized_feature")
+    per_batch = params.get("Per_Batch", "False")
+    # Annotation may be None, '', 'None', or a real name
+    annotation = params.get("Annotation")
+
+    # Convert parameters using text_to_value
+    if input_layer == "Original":
+        input_layer = None
+    
+    low_quantile = text_to_value(
+        low_quantile,
+        to_float=True,
+        param_name='Low_Quantile'
+    )
+    
+    high_quantile = text_to_value(
+        high_quantile,
+        to_float=True,
+        param_name='High_Quantile'
+    )
+    
+    # Convert "True"/"False" string to boolean (case-insensitive)
+    per_batch = str(per_batch).strip().lower() == "true"
+    
+    # Annotation is optional - empty string or "None" becomes None
+    annotation = text_to_value(annotation)
+    
+    # Validate annotation is provided when per_batch is True
+    if per_batch and annotation is None:
+        raise ValueError(
+            'Parameter "Annotation" is required when "Per Batch" is set '
+            'to True.'
+        )
+    
+    # Check if output_layer already exists in adata
+    print(f"Checking if output layer '{output_layer}' exists in adata "
+          f"layers...")
+    if output_layer in adata.layers.keys():
+        raise ValueError(
+            f"Output Table Name '{output_layer}' already exists, "
+            f"please rename it."
+        )
+    else:
+        print(f"Output layer '{output_layer}' does not exist. "
+              f"Proceeding with normalization.")
+    
+    def df_as_html(
+        df,
[…]
