### 2025-08-04 de6ee91
fix(quantile_scaling_template): fix typo in both function and unit tests
