### 2025-08-04 a71e865
feat(normalize_batch_template): add normalize_batch_template functionand unit tests
