### 2025-08-04 9edf8ce
fix(normalize_batch_template): fix typo and unused import
