+++ src/spac/templates/normalize_batch_template.py
@@ -11,7 +11,6 @@ import json
 import sys
 from pathlib import Path
 from typing import Any, Dict, Union
-import pandas as pd
 
 # Add parent directory to path for imports
 sys.path.append(str(Path(__file__).parent.parent.parent))
@@ -21,7 +20,6 @@ from spac.templates.template_utils import (
     load_input,
     save_outputs,
     parse_params,
-    text_to_value,
 )
 
 
