### 2025-08-04 abda610
feat(tsne_analysis_template): add tsne_analysis_template function and unit tests
