### 2025-08-04 bbb53f7
feat(posit_it_python_template): add posit_it_python_template functionand unit tests
