+++ src/spac/templates/posit_it_python_template.py
@@ -1,0 +1,220 @@
+"""
+Platform-agnostic Post-It-Python template converted from NIDAP.
+Maintains the exact logic from the NIDAP template.
+
+Usage
+-----
+>>> from spac.templates.posit_it_python_template import run_from_json
+>>> run_from_json("examples/posit_it_python_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, Optional, Tuple
+import matplotlib.pyplot as plt
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.templates.template_utils import (
+    save_outputs,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results: bool = True,
+    show_plot: bool = False
+) -> Union[Dict[str, str], Any]:
+    """
+    Execute Post-It-Python analysis with parameters from JSON.
+    Replicates the NIDAP template functionality exactly.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results : bool, optional
+        Whether to save results to file. If False, returns the figure
+        directly for in-memory workflows. Default is True.
+    show_plot : bool, optional
+        Whether to display the plot. Default is False.
+
+    Returns
+    -------
+    dict or figure
+        If save_results=True: Dictionary of saved file paths
+        If save_results=False: The matplotlib figure object
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Extract parameters using .get() with defaults from JSON template
+    text = params.get("Label", "Post-It")
+    text_color = params.get("Label_font_color", "Black")
+    text_size = params.get("Label_font_size", "80")
+    text_fontface = params.get("Label_font_type", "normal")
+    text_fontfamily = params.get("Label_font_family", "Arial")
+    bold = params.get("Label_Bold", "False")
+
+    # background params
+    fill_color = params.get("Background_fill_color", "Yellow1")
+    fill_alpha = params.get("Background_fill_opacity", "10")
+    border_alpha = 1
+    border_width = 0
+
+    # image params
+    image_width = params.get("Page_width", "18")
+    image_height = params.get("Page_height", "6")
+    image_resolution = params.get("Page_DPI", "300")
+
+    # output value
+    tag = "CCBR"
+
+    # Convert string parameters to appropriate types
+    text_size = text_to_value(
+        text_size,
+        to_int=True,
+        param_name="Label_font_size"
+    )
+    
+    bold = text_to_value(bold) == "True"
+    
+    fill_alpha = text_to_value(
+        fill_alpha,
+        to_float=True,
+        param_name="Background_fill_opacity"
+    )
+    
+    image_width = text_to_value(
+        image_width,
+        to_float=True,
+        param_name="Page_width"
+    )
+    
+    image_height = text_to_value(
+        image_height,
+        to_float=True,
+        param_name="Page_height"
+    )
+    
+    image_resolution = text_to_value(
+        image_resolution,
+        to_int=True,
+        param_name="Page_DPI"
+    )
+
+    # RUN ====   
+
+    # paints    
+    paints = {
+        'White': '#FFFFFF',
+        'LightGrey': '#D3D3D3',
+        'Grey': '#999999',
+        'Black': '#000000',
+        'Red1': '#F44E3B',
+        'Red2': '#D33115',
+        'Red3': '#9F0500',
+        'Orange1': '#FE9200',
[…]
