### 2025-08-04 e70f547
fix(posit_it_python_template): fixed typo
