### 2025-08-13 #8211 @adamltyson
It looks like everyone is busy, so I think we can go ahead and get started. Thanks @jluethi, @kevinyamauchi!
