### 2025-08-26 #8787 @kevinyamauchi
Hello everyone! Thank you for the interesting project, @fangliu117. I have taken a first pass through the review checklist. Overall, it seems like a well-written library. Based on the checklist, I have a few questions below. Please let me know if you need any clarification!

## Functionality

>  Functionality: Have the functional claims of the software been confirmed?

I am sorry if I missed it, but is there any sample data available? As I do not have access to any suitable data, I cannot evaluate the functionality of the software.

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

I agree with @jluethi  that the statement in the paper about enhanced performance is unclear ("SPAC enhances computational efficiency by over 5x by integrating optimized numerical routines from NumPy's compiled C-based backend"). For what tasks is this speedup observed? How was it measured? Compare to which other software?

Also it is stated in the paper that "Unlike generalized Ripley’s implementations (e.g., Squidpy), SPAC explicitly distinguishes phenotype pairings and employs edge correction by excluding cells located near the region's borders within the analytical radius, mitigating edge-effect biases and enhancing statistical reliability. " I think it would be helpful if there was some benchmarking to support this statement.

## Documentation
>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I think the statement of could more explicitly state what types of analysis SPAC is designed to perform. I think it is clear what type of data SPAC is designed to analyze, but it is not clear what questions SPAC can answer.

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I was unable to find any examples of how to use the library. I was also unable to find any sample data. I suggest that the authors add some example scripts and/or notebooks with sample data and the expected outputs documented. Ideally these examples would cover both analysis and visualization.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

It is a bit unclear to me how the CI works. From my quick look, it appears to me that unit tests get run on PRs, but they do not on push to `main`. Is that correct? If so, what is the reason they aren't run on push to main? If they aren't run on push to main, I find it difficult to tell if the tests are currently passing.

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

There is a contributing guide, but I did not see any information on how to report issues or seek support. Would it be possible to add such a statement to the README?

## Software paper
> A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I think both the "statement of need" and "state of the field" should be clarified. In the paper it is stated that existing tools "lack sufficient flexibility and customization options necessary to meet the diverse scalable analysis and visualization needs of non-technical users." However, it is unclear to me in which ways to current tools are inflexible and what the specific missing options are that SPAC address.
