### 2025-08-27 #8787 @ruiheesi
> I had a first look at it today, super interesting! I'll have a few first questions / comments today and will follow up more in-depth later.
>
> [@fangliu117](https://github.com/fangliu117) Are there any public example data available? And do you have example processing workflows? It would help a lot to review the functionality of SPAC to run it locally and have some examples to run (both by getting example data and some pointers on how to run a standard workflow). Such example workflows would also help future users better understand how to use SPAC, what input data structures are supported etc.
>
> Do you have any additional documentation besides https://fnlcr-dmap.github.io/SCSAWorkflow/ ? I was particularly wondering about requirements for the data inputs (e.g. supported format) and whether you expect image processing for segmentation and measurements to happen before SPAC or whether SPAC performs them.
>
> Besides that, here's a few initial questions or comments that came up:
>
> > SPAC enhances computational efficiency by over 5x by integrating optimized numerical routines from NumPy's compiled C-based backend.
>
> 5x compared to what?
>
> Line 66:
>
> > toggle of feature
>
> => toggle off
>
> > enterprise-level software engineering standards, featuring extensive unit testing, rigorous edge-case evaluation, comprehensive logging, and clear, context-rich error handling
>
> Do you measure your test coverage somewhere (a Github CI, github badges or such)? Also, are you running your unit testing in the GitHub CI? I saw that you're using CodeQL (which I'm not familiar with myself yet), but couldn't understand if that would also run your tests.

Hi @kevinyamauchi, thanks for the review and I can answer the CI/CD part. The package was established to support NIH Integrated Data Analysis Platform (NIDAP) usage that require the package to be formatted into conda package with specific deployment/testing steps. In other words it requires deployment to a specific test environment before releasing to the production.

We adapt the [A successful Git branching model - Vincent Driessen](https://nvie.com/posts/a-successful-git-branching-model/) to manage development and implement CI/CD to work with this strategy. We built [gitflow-py](https://github.com/FNLCR-DMAP/gitflow-py) and [gitflow-R](https://github.com/FNLCR-DMAP/gitflow-R) reusable github action workflow to manage the tests and deployment. We can talk further about these action packages if you are interested, but in short:

CI happens at:
1. Local feature branch pushing to GitHub repo branch, to accelerated the process we run test only on changed functions and their corresponding unittests. Example is this action run [action 1081](https://github.com/FNLCR-DMAP/SCSAWorkflow/actions/runs/16282422260/job/45974516903)
2. When PR opened to merge the feature branch into dev branch again on the PR branch, this time runs all unittests on the feature branch to see if the newly merging code would live with others prior to merging into the dev. Example [PR 394](https://github.com/FNLCR-DMAP/SCSAWorkflow/pull/394)
3. When PR open to merge into release_dev, since release dev is the deployment to test environment, it again run all unittests prior to package creation. example is [PR 340](https://github.com/FNLCR-DMAP/SCSAWorkflow/actions/runs/14996219724).

CI 1, 2, 3 make sure the package is working prior to deployment, the test are done in prebuilt ubuntu docker container inside a conda environment to be compliant to the target deployment environment.

CD happens at:
1. Success merged into release_dev, this will create a dev version of conda package deployment. This is mainly to test compliant with the product environment.
2. Success merged into main, this will  create the prod version for user consumption.
The deployment history of each conda package can be found in the [Conda_Package branch](https://github.com/FNLCR-DMAP/SCSAWorkflow/tree/Conda_Package).

Please let me know if this answers your question, thanks.
Now when
