### 2025-09-01 #8787 @ruiheesi
> Thanks a lot [@ruiheesi](https://github.com/ruiheesi) for the explanation on the CI/CD workflows. Interesting process for sure.
>
> What I don't follow is whether you/we can see somewhere what kind of test coverage you achieve. Is this visible somewhere?
>
> And how would one know whether the changes that have been merged do pass the tests? Is that knowledge purely process based (e.g. "we wouldn't merge to main if tests were failing") or do any tests run on the main branch? The CodeQL actions to my naive perspective don't look like they are running tests (though I may be missing something).
>
> > I am sorry if I missed it, but is there any sample data available? As I do not have access to any suitable data, I cannot evaluate the functionality of the software.
>
> I am also stuck at this and don't see a clear way to review the functionality of the package without sample data and preferably also some sample workflows being made available. Is this something you can do?

The current test is based on the unittests we wrote for each of the function that we implement. There is not an automated coverage metric per se for now as we strictly implement each function pair with tests before able to deploy. Now thinking about it, it would be very helpful to have this metric moving forward.

As for merge strategy, all unittest are ran twice for each main release and we do lock main down. We do not directly make change to the main/production to maintain the production stability. Thus changes made to any specific code should pass three rounds of tests before reaching production:
1. When first push to the feature branch
2. When a pull request is opened to propose merging into the dev branch.
3. When a pull request is opened to propose merging into the release (release_dev) branch.

As for sample data, we are using Scimap's tutorial data to establish demonstration workflow on NIDAP, maybe @fangliu117 can share the data with you.
