### 2025-09-02 5243268
feat: Address reviewer comments: Add comprehensive example notebook and clarify documentation
