### 2025-09-16 #8787 @fangliu117
Hi @kevinyamauchi @adamltyson , thanks for checking in, and for all the feedback. I had a brief pause but I’m back on it now and will post the full revision and point‑by‑point response by Friday, September 19. I don’t need any clarification at the moment; if anything comes up as I finalize, I’ll ask here. Thank you all very much again!
