### 2025-09-22 706cf79
fix: revised manuscript for joss
