+++ paper/paper.md
@@ -44,28 +44,37 @@ Multiplexed immunofluorescence microscopy captures detailed measurements of spat
 
 # Statement of Need
 
-Advanced multiplex imaging technologies, such as CODEX [@Goltsev:2018], MxIF [@Gerdes:2013], CyCIF [@Lin:2018], generate high dimensional dataset capable of profiling up to dozens of biomarkers simultaneously. Analyzing and interpreting these complex spatial protein data pose significant computational challenges, especially given that high-resolution whole-slide imaging data can reach hundreds of gigabytes in size and contain millions of cells across extensive tissue areas. Currently, many spatial biology tools (e.g., Seurat [@Hao:2021], GraphST [@Long:2023], and bento [@Mah:2024]), primarily address spatial transcriptomics and cannot directly handle multiplexed protein imaging data. Other specialized software such as SPIA [@Feng:2023], Giotto [@Dries:2021], Squidpy [@Palla:2022], and SCIMAP [@Nirmal:2024] provides valuable capabilities tailored for spatial protein analyses. However, these tools lack sufficient flexibility and customization options necessary to meet the diverse scalable analysis and visualization needs of non-technical users.
+Multiplexed protein imaging (e.g., CODEX [@Goltsev:2018], CyCIF [@Lin:2018], MxIF [@Gerdes:2013]) yields gigabyte-scale, single-cell datasets with millions of cells per study. Extracting biological signal from these data requires reproducible workflows that span preprocessing, phenotyping, spatial statistics, and interactive visualization. While recent tools have advanced the analysis of spatial omics, biologists without coding expertise still face steep barriers to conduct end-to-end analyses independently and iteratively. This slows hypothesis testing and risks brittle, one-off scripts that are hard to reuse. SPAC is designed to close this gap.
 
-To address this gap, we developed the SPAC Python package and the web-based SPAC Shiny application, which together enhance analytical capabilities through intuitive terminology, optimized computational performance, specialized spatial statistics, and extensive visualization configurations. Results computed using the SPAC Python package are stored as AnnData objects, which can be interactively explored in real time via the SPAC Shiny web application, enabling researchers to dynamically visualize data, toggle annotations, inspect cell populations, and compare experimental conditions without requiring extensive computational expertise.
+SPAC is a Python package for downstream analysis of single-cell multiplexed imaging after segmentation. It ingests per-cell tables (CSV/H5AD/tabular) with spatial coordinates and marker intensities and exposes a reproducible, AnnData‑compliant Python API. All analysis artifacts and figures are preserved in AnnData, supporting sharing, re‑runs, and downstream reuse. As the core analytical layer of the broader SPAC ecosystem [@Liu:2025], the package connects to interactive pipelines and real‑time dashboards on enterprise platforms via a modular, layered architecture, enabling large‑scale analyses without requiring users to write code. Its purpose is to lower the barrier for non‑programmers while preserving rigorous, reusable workflows; for example, SPAC uses biologist-friendly terminology (e.g., “cells,” “features,” “tables,” “associated tables,” “annotations”) and clear messages so users can interpret results and diagnostics without a bioinformatic analyst’s assistance. 
 
-Specifically, SPAC uses biologist-friendly terminology to simplify technical AnnData concepts. In SPAC, \"cells\" are rows in the data matrix, \"features\" denote protein expression levels, \"tables\" contain transformed data layers, \"associated tables\" store spatial coordinates or dimensional reductions (e.g., UMAP embeddings), and \"annotations\" indicate cell phenotypes, experimental labels, slide identifiers, and other categorical data.
[…]
