+++ paper/paper.bib
@@ -1,5 +1,56 @@
+@misc{CodeOcean,
+    author = {{Code Ocean}},
+    title = {{Code Ocean}},
+    url = {https://codeocean.com/},
+    urldate = {2025-04-15},
+    year = {2019}
+}
+
+@article{Dries:2021,
+    author = {Dries, Ruben and Zhu, Qian and Dong, Rui and Eng, Chee Huat Linus and Li, Huipeng and Liu, Kan and Fu, Yuntian and Zhao, Tianxiao and Sarkar, Arpan and Bao, Feng and George, Rani E and Pierson, Nico and Cai, Long and Yuan, Guo Cheng},
+    doi = {10.1186/s13059-021-02286-2},
+    issn = {1474760X},
+    journal = {Genome Biology},
+    month = {dec},
+    number = {1},
+    pages = {78},
+    pmid = {33685491},
+    publisher = {BioMed Central Ltd},
+    title = {{Giotto: a toolbox for integrative analysis and visualization of spatial expression data}},
+    volume = {22},
+    year = {2021}
+}
+
+@article{Feng:2023,
+    author = {Feng, Yuzhou and Yang, Tianpei and Zhu, John and Li, Mabel and Doyle, Maria and Ozcoban, Volkan and Bass, Greg T and Pizzolla, Angela and Cain, Lachlan and Weng, Sirui and Pasam, Anupama and Kocovski, Nikolce and Huang, Yu Kuan and Keam, Simon P and Speed, Terence P and Neeson, Paul J and Pearson, Richard B and Sandhu, Shahneen and Goode, David L and Trigos, Anna S},
+    doi = {10.1038/s41467-023-37822-0},
+    issn = {20411723},
+    journal = {Nature Communications},
+    month = {dec},
+    number = {1},
+    pages = {2787},
+    pmid = {37188662},
+    publisher = {Nature Research},
+    title = {{Spatial analysis with SPIAT and spaSim to characterize and simulate tissue microenvironments}},
+    volume = {14},
+    year = {2023}
+}
+
+@article{Jalili:2020,
+    author = {Jalili, Vahid and Afgan, Enis and Gu, Qiang and Clements, Dave and Blankenberg, Daniel and Goecks, Jeremy and Taylor, James and Nekrutenko, Anton},
+    doi = {10.1093/nar/gkaa554},
+    journal = {Nucleic Acids Research},
+    month = {jul},
+    number = {W1},
+    pages = {W395--W402},
+    pmid = {32637963},
+    publisher = {Oxford University Press},
+    title = {{The Galaxy platform for accessible, reproducible and collaborative biomedical analyses: 2020 update}},
+    volume = {48},
+    year = {2020}
+}
+
 @article{Gerdes:2013,
[…]
