### 2025-09-22 #8787 @fangliu117
> Also it is stated in the paper that "Unlike generalized Ripley’s implementations (e.g., Squidpy), SPAC explicitly distinguishes phenotype pairings and employs edge correction by excluding cells located near the region's borders within the analytical radius, mitigating edge-effect biases and enhancing statistical reliability. " I think it would be helpful if there was some benchmarking to support this statement.
>
Hi @kevinyamauchi @jluethi @adamltyson Thank you all again for your interests and constructive feedback. Thank you for the suggestion of benchmark for Ripley's implementation. We added a concise benchmark to the example notebook. that contrasts edge correction OFF vs. ON for a bivariate Ripley’s L analysis:

- Data/Pairing: centers = follicular helper T cells (Tfh); neighbors = proliferative B cells.
- Null model: 100 Poisson simulations per radius to form a simulation envelope.
- [Notebook:](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/address-reviewer-comments/paper/examples/lymphnode_analysis.ipynb) paper/examples/lymphnode_analysis.ipynb on the address-reviewer-comments branch.
What changes with correction? When edge correction is OFF, the simulated null (orange) falls below the theoretical L=r line and the bias grows at larger radi, this is the known boundary undercount that underestimates L(r). When we turn edge correction ON (SPAC excludes center cells within r of the boundary), the simulated null aligns with L=r across radii, as expected for a homogeneous Poisson process.
Biological inference: in both cases the empirical curve (blue) lies above the null, indicating clustering of Tfh around proliferative B cells. With correction ON, that deviation persists over a broad range of radii, demonstrating that the effect is not an edge artifact. Please find more detailed analysis in the example notebook.
I also invite the corresponding author @georgenih for answering further concerns.
