### 2025-09-22 #8787 @fangliu117
> ## Software paper
> > A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
>
> > State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> I think both the "statement of need" and "state of the field" should be clarified. In the paper it is stated that existing tools "lack sufficient flexibility and customization options necessary to meet the diverse scalable analysis and visualization needs of non-technical users." However, it is unclear to me in which ways to current tools are inflexible and what the specific missing options are that SPAC address.

Hi @kevinyamauchi @jluethi @adamltyson thanks again for the detailed feedback. Per your comments, we have revised the paper to clarify both the Statement of need and State of the field. Here is a **summary of statements**:

**Statement of need (Who the paper is for and what problem it solves).**
Single cell spatial analysis from processed multiplexed imaging is challenging due to larger scale and complexity of dataset. While recent tools have advanced the analysis of spatial omics, biologists without coding expertise still face steep barriers.

Targeting those non-technical users, we have developed SPAC ecosystem that provides enterprise-level scalability, interactive pipelines, and real-time visualization via a modular, layered architecture. As a core layer of SPAC ecosystem, the SPAC python package is the foundation on which we build tools for non- technical users.  It has distinctive features to low barriers for biologists, for example, it adopts biologist friendly terminology and messages so users can interpret errors and warnings without consulting a bioinformatic analyst; pins colors to annotations to maintain visual consistency across all plots and sessions; and stratified and gridded plotting enable comparison (e.g., across conditions or phenotypes) clear in a single view and publication-ready. It accelerates the computation by custom NumPy based implementations (e.g., histograms/boxplot) for near real time performance even with tens of millions of cells.

SPAC is engineered to support upper‑layer interfaces (e.g., Code Workbook/NIDAP, Galaxy, Code Ocean and Shiny dashboards) without coupling the package to any single platform. This separation of concerns lets software engineers extend methods in Python while biologists interact through guided, no code front ends.

**State of the field (how SPAC relates to existing packages):**
We explicitly acknowledge and build on community standards (AnnData/Scanpy) and spatial packages (Squidpy, SCIMAP). SPAC leverages these packages for the functionality they provide, but adds a layer that enables us build tools for non-technical users, and implement new features or enhancements as needed. In particular, we highlight: (i) integrated, flexible visualization (stratification, subsetting, faceting, pinned colors) and optimized summary statistics method; (ii) new features added: e.g., neighborhood profile to quantify the numbers of cells without equidistant bins from center phenotype;  (iii) augmented existing functions: e.g., a phenotype‑pair-specific Ripley’s L (explicit center vs neighbor phenotypes) with radius‑dependent guard‑zone edge correction (excluding center cells whose r‑neighborhood intersects the ROI boundary), which reduces edge‑induced inflation and yields more interpretable L(r) curves for cell-cell interactions.

These points are now stated directly in the Statement of Need and State of the Field section in the updated [paper.md.](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/address-reviewer-comments/paper/paper.md) on the address-reviewer-comments branch. For readers who want more ecosystem context (how this Python package underpins interactive/no‑code layers), we reference the separate SPAC ecosystem manuscript; the JOSS paper here stays focused on the Python package itself.

We added a public, self‑contained example (notebook + small CSV) showing an end‑to‑end workflow, with updated Ripley’s L with and without edge exclusion [,](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/address-reviewer-comments/paper/examples/lymphnode_analysis.ipynb) so readers can reproduce the figures and see the edge‑correction effect. The plotting template notes how many center cells are used when edge correction is on and explains how to interpret L(r) vs. CSR envelopes.

Please let us know if this addresses your request, or if you’d like any further tightening of wording in either paragraph. Thank you again for all your interests and constructive comments. We are looking forward to your further feedback.
