### 2025-09-22 8480f90
fix: revised and reload manuscript for joss
