+++ paper/paper.bib
@@ -36,7 +36,7 @@
     year = {2023}
 }
 
-@article{Jalili:2020,
+@article{Galaxy:2020,
     author = {Jalili, Vahid and Afgan, Enis and Gu, Qiang and Clements, Dave and Blankenberg, Daniel and Goecks, Jeremy and Taylor, James and Nekrutenko, Anton},
     doi = {10.1093/nar/gkaa554},
     journal = {Nucleic Acids Research},
