+++ paper/paper.md
@@ -54,14 +54,11 @@ To address real-time scalability challenges in analyzing large multiplex imaging
 
 SPAC provides customizable visualization methods. Pin colors to annotations maintain consistent color mapping across figures and sessions. Interactive spatial plots (Plotly) allow users to toggle off features (e.g., biomarkers) and switch among multiple annotations simultaneously, enabling intuitive exploration of spatial relationships and patterns. Stratified and gridded plotting with optimized statistical summary make subgroup comparisons (e.g., across conditions or phenotypes) clear within a single view. 
 
-SPAC provides customizable visualization methods, leveraging Plotly\'s interactive capabilities for dynamic exploration of spatial data. Interactive spatial plots allow users to toggle off features (e.g., biomarkers) and multiple annotations simultaneously, while a pin-color option ensures consistent color mapping across analyses. These designs help researchers intuitively explore spatial relationships by switching between different cell populations and identify patterns before performing detailed quantitative analyses. In addition, SPAC supports comparative visualization, such as overlaying manual classifications with unsupervised clustering or comparing spatial distributions across experimental conditions or treatments. It also enhances core analytical functions (e.g., nearest neighbor computations using SCIMAP\'s spatial distance calculations) by integrating extensive visualization configurations, including subgroup analyses, subset plots, and faceted layouts, allowing tailored visual outputs for various experimental contexts and research questions.
-
 # Statement of Need
 
 The spatial omics ecosystem includes general frameworks for single cell data and spatial analysis (AnnData/Scanpy, Squidpy) and toolkits focused on multiplex imaging and spatial proteomics (SCIMAP), alongside methods aimed primarily at spatial transcriptomics (Seurat, GraphST, Bento). These packages collectively cover data structures, clustering, dimensionality reduction, neighborhood graphs, and spatial interaction metrics, and they are widely adopted for programmatic analyses. 
 
 Within this landscape, SPAC builds directly on community standards such as AnnData and established methods, but contributes a coherent foundation that enhances key functions for spatial analyses and visualization. Highlighting a few examples, SPAC enhances core analyses (e.g., nearest‑neighbor metrics leveraging SCIMAP\’s spatial‑distance utilities) with flexible, built‑in plotting for publication ready figures, including native support for stratification, subsetting, and faceting. Users can fine‑tune plot aesthetics and export layout (margins, size, DPI) to produce clear, consistent figures. SPAC strengthens the functionality of existing packages: SPAC implements a phenotype‑pair specific Ripley’s L that treats one phenotype as “centers” and another as “neighbors,” and applies radius‑dependent guard-zone edge correction by excluding center cells whose r‑neighborhood intersects the ROI boundary. Unlike generalized implementations (e.g., Squidpy), this design reduces edge‑induced inflation at larger radii and yields more reliable, interpretable L(r) curves for concrete cell‑cell interactions. SPAC adds new features. Among its features, SPAC implements efficient KDTree‐based neighborhood profiling that quantifies the distribution of neighboring cell phenotypes in user‑defined radial distance bins around each center cell of a specified phenotype. The resulting three‑dimensional array, which summarizes the local cellular microenvironment, is stored in the AnnData object and supports dimensionality‑reduction methods such as spatial UMAP [@Giraldo:2021]. 
-
 
 # Structure and Implementation
 
@@ -69,7 +66,7 @@ The SPAC package is available at [GitHub](https://github.com/FNLCR-DMAP/SCSAWork
 
[…]
