### 2025-09-22 f09a0de
fix: revised manuscript for joss at 2nd time
