+++ paper/paper.md
@@ -34,44 +34,42 @@ affiliations:
     name: Axle Informatics, United States
   - index: 4
     name: Purdue University, United States
-date: 12 April 2025
+date: 15 April 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-Multiplexed immunofluorescence microscopy captures detailed measurements of spatially resolved, multiple biomarkers simultaneously, revealing tissue composition and cellular interactions in situ among single cells. The growing scale and dimensional complexity of these datasets demand reproducible, comprehensive and user-friendly computational tools. To address this need, we developed SPAC **(SPA**tial single-**C**ell analysis), a Python-based package and a corresponding shiny application within an integrated, modular SPAC ecosystem designed specifically for biologists without extensive coding expertise. Following image segmentation and extraction of spatially resolved single-cell data, SPAC streamlines downstream phenotyping and spatial analysis, facilitating characterization of cellular heterogeneity and spatial organization within tissues. Through scalable performance, specialized spatial statistics, highly customizable visualizations, and seamless workflows from dataset to insights, SPAC significantly lowers barriers to sophisticated spatial analyses.
+Multiplexed immunofluorescence microscopy captures detailed measurements of spatially resolved, multiple biomarkers simultaneously, revealing tissue composition and cellular interactions in situ among single cells. The growing scale and dimensional complexity of these datasets demand reproducible, comprehensive and user-friendly computational tools. To address this need, we developed SPAC **(SPA**tial single-**C**ell analysis), a Python-based package and a corresponding Shiny application within an integrated, modular SPAC ecosystem designed specifically for biologists without extensive coding expertise. Following image segmentation and extraction of spatially resolved single-cell data, SPAC streamlines downstream phenotyping and spatial analysis, facilitating characterization of cellular heterogeneity and spatial organization within tissues. Through scalable performance, specialized spatial statistics, highly customizable visualizations, and seamless workflows from dataset to insights, SPAC significantly lowers barriers to sophisticated spatial analyses.
 
 # Statement of Need
 
 Multiplexed protein imaging (e.g., CODEX [@Goltsev:2018], CyCIF [@Lin:2018], MxIF [@Gerdes:2013]) yields gigabyte-scale, single-cell datasets with millions of cells per study. Extracting biological signal from these data requires reproducible workflows that span preprocessing, phenotyping, spatial statistics, and interactive visualization. While recent tools have advanced the analysis of spatial omics, biologists without coding expertise still face steep barriers to conduct end-to-end analyses independently and iteratively. This slows hypothesis testing and risks brittle, one-off scripts that are hard to reuse. SPAC is designed to close this gap.
 
-SPAC is a Python package for downstream analysis of single-cell multiplexed imaging after segmentation. It ingests per-cell tables (CSV/H5AD/tabular) with spatial coordinates and marker intensities and exposes a reproducible, AnnData‑compliant Python API. All analysis artifacts and figures are preserved in AnnData, supporting sharing, re‑runs, and downstream reuse. As the core analytical layer of the broader SPAC ecosystem [@Liu:2025], the package connects to interactive pipelines and real‑time dashboards on enterprise platforms via a modular, layered architecture, enabling large‑scale analyses without requiring users to write code. Its purpose is to lower the barrier for non‑programmers while preserving rigorous, reusable workflows; for example, SPAC uses biologist-friendly terminology (e.g., “cells,” “features,” “tables,” “associated tables,” “annotations”) and clear messages so users can interpret results and diagnostics without a bioinformatic analyst’s assistance. 
[…]
