### 2025-09-30 978c18e
fix: Update paper for reviewer comments
