### 2025-10-01 9e3bea0
feat: Add SPAC boxplot Galaxy tool for Docker deployment
