### 2025-10-01 #8787 @fangliu117
Thanks again for taking another look, much appreciated! Please take the time you need @kevinyamauchi @jluethi, if you run into any issues setting up the environment, we'd share a ready‑to‑run Docker container (with all dependencies). We are happy to help troubleshoot, e.g., running the example workbook, in any other way that’s useful.
