+++ galaxy_tools/spac_load_csv_files/run_spac_template.sh
@@ -1,0 +1,403 @@
+#!/usr/bin/env bash
+# run_spac_template.sh - SPAC wrapper with Load CSV fixes
+# Version: 5.3.0 - Complete version with Load CSV handling
+set -euo pipefail
+
+PARAMS_JSON="${1:?Missing params.json path}"
+TEMPLATE_BASE="${2:?Missing template base name}"
+
+# Handle both base names and full .py filenames for backward compatibility
+if [[ "$TEMPLATE_BASE" == *.py ]]; then
+    TEMPLATE_PY="$TEMPLATE_BASE"
+elif [[ "$TEMPLATE_BASE" == "load_csv_files_with_config" ]]; then
+    TEMPLATE_PY="load_csv_files_with_config.py"
+else
+    TEMPLATE_PY="${TEMPLATE_BASE}_template.py"
+fi
+
+# Use SPAC Python environment
+SPAC_PYTHON="${SPAC_PYTHON:-python3}"
+
+echo "=== SPAC Template Wrapper v5.3 ==="
+echo "Parameters: $PARAMS_JSON"
+echo "Template base: $TEMPLATE_BASE"
+echo "Template file: $TEMPLATE_PY"
+echo "Python: $SPAC_PYTHON"
+
+# Run template through Python
+"$SPAC_PYTHON" - <<'PYTHON_RUNNER' "$PARAMS_JSON" "$TEMPLATE_PY" 2>&1 | tee tool_stdout.txt
+import json
+import os
+import sys
+import copy
+import traceback
+import inspect
+import shutil
+
+# Get arguments
+params_path = sys.argv[1]
+template_filename = sys.argv[2]
+
+print(f"[Runner] Loading parameters from: {params_path}")
+print(f"[Runner] Template: {template_filename}")
+
+# Load parameters
+with open(params_path, 'r') as f:
+    params = json.load(f)
+
+# ---------------------------------------------------------------------------
+# De-sanitization and parsing helpers
+# ---------------------------------------------------------------------------
+def _unsanitize(s: str) -> str:
+    """Remove Galaxy's parameter sanitization tokens"""
+    if not isinstance(s, str):
+        return s
+    replacements = {
+        '__ob__': '[', '__cb__': ']',
+        '__oc__': '{', '__cc__': '}',
+        '__dq__': '"', '__sq__': "'",
+        '__gt__': '>', '__lt__': '<',
+        '__cn__': '\n', '__cr__': '\r',
+        '__tc__': '\t', '__pd__': '#',
+        '__at__': '@', '__cm__': ','
+    }
+    for token, char in replacements.items():
+        s = s.replace(token, char)
+    return s
+
+def _maybe_parse(v):
+    """Recursively de-sanitize and JSON-parse strings where possible."""
+    if isinstance(v, str):
+        u = _unsanitize(v).strip()
+        if (u.startswith('[') and u.endswith(']')) or (u.startswith('{') and u.endswith('}')):
+            try:
+                return json.loads(u)
+            except Exception:
+                return u
+        return u
+    elif isinstance(v, dict):
+        return {k: _maybe_parse(val) for k, val in v.items()}
+    elif isinstance(v, list):
+        return [_maybe_parse(item) for item in v]
+    return v
+
+# First normalize the whole params tree
+params = _maybe_parse(params)
+
+# Extract template name
+template_name = os.path.basename(template_filename).replace('_template.py', '').replace('.py', '')
+
+# Helper function to coerce singleton lists to strings for load_csv
+def _coerce_singleton_paths_for_load_csv(params, template_name):
+    """For load_csv templates, flatten 1-item lists to strings for path-like params."""
+    if 'load_csv' not in template_name:
+        return params
[…]
