### 2025-10-08 009d010
refactor: streamline galaxy tools implementation
