+++ galaxy_tools/spac_setup_analysis/run_spac_template.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 # run_spac_template.sh - SPAC wrapper with column index conversion
-# Version: 5.4.1 - Integrated column conversion
+# Version: 5.5.0 - Enhanced text input handling for setup_analysis
 set -euo pipefail
 
 PARAMS_JSON="${1:?Missing params.json path}"
@@ -18,7 +18,7 @@ fi
 # Use SPAC Python environment
 SPAC_PYTHON="${SPAC_PYTHON:-python3}"
 
-echo "=== SPAC Template Wrapper v5.3 ==="
+echo "=== SPAC Template Wrapper v5.5 ==="
 echo "Parameters: $PARAMS_JSON"
 echo "Template base: $TEMPLATE_BASE"
 echo "Template file: $TEMPLATE_PY"
@@ -90,11 +90,88 @@ def _maybe_parse(v):
 params = _maybe_parse(params)
 
 # ===========================================================================
-# COLUMN INDEX CONVERSION - CRITICAL FOR SETUP ANALYSIS
+# SETUP ANALYSIS SPECIAL HANDLING - Process text inputs before column conversion
+# ===========================================================================
+def process_setup_analysis_text_inputs(params, template_name):
+    """Process text-based column inputs for setup_analysis template"""
+    if 'setup_analysis' not in template_name:
+        return params
+    
+    print("[Runner] Processing setup_analysis text inputs")
+    
+    # Handle X_centroid and Y_centroid (single text values)
+    for coord_key in ['X_centroid', 'Y_centroid']:
+        if coord_key in params:
+            value = params[coord_key]
+            if isinstance(value, list) and len(value) == 1:
+                params[coord_key] = value[0]
+            # Ensure it's a string
+            if value:
+                params[coord_key] = str(value).strip()
+                print(f"[Runner] {coord_key} = '{params[coord_key]}'")
+    
+    # Handle Annotation_s_ (text area, can be comma-separated or newline-separated)
+    if 'Annotation_s_' in params:
+        value = params['Annotation_s_']
+        if value:
+            # Convert to list if it's a string
+            if isinstance(value, str):
+                # Check for comma separation first, then newline
+                if ',' in value:
+                    items = [item.strip() for item in value.split(',') if item.strip()]
+                elif '\n' in value:
+                    items = [item.strip() for item in value.split('\n') if item.strip()]
+                else:
+                    # Single value
+                    items = [value.strip()] if value.strip() else []
+                params['Annotation_s_'] = items
+                print(f"[Runner] Parsed Annotation_s_: {len(items)} items -> {items}")
+            elif not isinstance(value, list):
+                params['Annotation_s_'] = []
+        else:
+            params['Annotation_s_'] = []
+    
+    # Handle Feature_s_ (text area, can be comma-separated or newline-separated)
+    if 'Feature_s_' in params:
+        value = params['Feature_s_']
+        if value:
+            # Convert to list if it's a string
+            if isinstance(value, str):
+                # Check for comma separation first, then newline
+                if ',' in value:
+                    items = [item.strip() for item in value.split(',') if item.strip()]
+                elif '\n' in value:
+                    items = [item.strip() for item in value.split('\n') if item.strip()]
+                else:
+                    # Single value
+                    items = [value.strip()] if value.strip() else []
+                params['Feature_s_'] = items
+                print(f"[Runner] Parsed Feature_s_: {len(items)} items")
+                if len(items) <= 10:
+                    print(f"[Runner]   Features: {items}")
+            elif not isinstance(value, list):
+                params['Feature_s_'] = []
+        else:
+            params['Feature_s_'] = []
+    
+    # Handle Feature_Regex (optional text field)
+    if 'Feature_Regex' in params:
[…]
