+++ galaxy_tools/spac_boxplot/run_spac_template.sh
@@ -1,186 +1,62 @@
 #!/usr/bin/env bash
-# run_spac_template.sh - Docker version for Galaxy
-# Version: 4.0.0 - Imports templates from installed SPAC package
+# run_spac_template.sh - SPAC wrapper with column index conversion
+# Version: 5.4.2 - Integrated column conversion
 set -euo pipefail
 
-# Log everything to tool_stdout.txt
-exec > >(tee -a tool_stdout.txt) 2>&1
-
 PARAMS_JSON="${1:?Missing params.json path}"
-TEMPLATE_NAME="${2:?Missing template name}"  # Just the name, not filename
+TEMPLATE_BASE="${2:?Missing template base name}"
 
-# Use system Python inside Docker container
+# Handle both base names and full .py filenames
+if [[ "$TEMPLATE_BASE" == *.py ]]; then
+    TEMPLATE_PY="$TEMPLATE_BASE"
+elif [[ "$TEMPLATE_BASE" == "load_csv_files_with_config" ]]; then
+    TEMPLATE_PY="load_csv_files_with_config.py"
+else
+    TEMPLATE_PY="${TEMPLATE_BASE}_template.py"
+fi
+
+# Use SPAC Python environment
 SPAC_PYTHON="${SPAC_PYTHON:-python3}"
 
-echo "=== SPAC Template Wrapper v4.0 (Docker) ==="
+echo "=== SPAC Template Wrapper v5.4 ==="
 echo "Parameters: $PARAMS_JSON"
-echo "Template: $TEMPLATE_NAME"
+echo "Template base: $TEMPLATE_BASE"
+echo "Template file: $TEMPLATE_PY"
 echo "Python: $SPAC_PYTHON"
-echo "Working directory: $(pwd)"
 
-# Run template through Python interpreter
-"$SPAC_PYTHON" - <<'PYTHON_RUNNER' "$PARAMS_JSON" "$TEMPLATE_NAME"
+# Run template through Python
+"$SPAC_PYTHON" - <<'PYTHON_RUNNER' "$PARAMS_JSON" "$TEMPLATE_PY" 2>&1 | tee tool_stdout.txt
 import json
 import os
 import sys
 import copy
-import importlib
 import traceback
 import inspect
+import shutil
+import re
+import csv
 
-# Get command line arguments
+# Get arguments
 params_path = sys.argv[1]
-template_name = sys.argv[2]  # Just the name like "boxplot", not "boxplot_template.py"
+template_filename = sys.argv[2]
 
-print(f"[Runner] Starting execution for template: {template_name}")
-print(f"[Runner] Python version: {sys.version}")
+print(f"[Runner] Loading parameters from: {params_path}")
+print(f"[Runner] Template: {template_filename}")
 
-# ============================================================================
-# HELPER FUNCTIONS
-# ============================================================================
+# Load parameters
+with open(params_path, 'r') as f:
+    params = json.load(f)
 
-def determine_outputs_from_params(params):
-    """Read outputs from params if available, otherwise use defaults"""
-    if 'outputs' in params and isinstance(params['outputs'], dict):
-        outputs = params['outputs']
-        print(f"[Runner] Using outputs from params: {list(outputs.keys())}")
-        return outputs
-    
-    # Fallback to defaults based on template name
-    print(f"[Runner] No outputs in params, using defaults for {template_name}")
-    
-    if template_name in ['boxplot', 'histogram', 'violin_plot', 'scatter_plot',
-                         'hierarchical_heatmap', 'relational_heatmap']:
-        return {'figures': 'figure_folder', 'DataFrames': 'dataframe_folder'}
-    elif template_name in ['interactive_spatial_plot', 'interactive_scatter_plot']:
-        return {'html': 'html_folder'}
[…]
