### 2025-10-16 196f511
init benchmark script
