### 2025-10-19 4906439
fix(boxplot): replace deprecated append call with concat
