### 2025-10-19 862e523
feat(test_performance): add performance tests for boxplot/histogram
