+++ tests/test_performance/test_histogram_performance.py
@@ -1,0 +1,372 @@
+import unittest
+import time
+import warnings
+import numpy as np
+import pandas as pd
+import anndata as ad
+import matplotlib
+import matplotlib.pyplot as plt
+import seaborn as sns
+from sklearn.datasets import make_blobs
+from sklearn.preprocessing import StandardScaler
+from spac.visualization import histogram
+from spac.utils import check_annotation, check_feature, check_table
+
+matplotlib.use('Agg')  # Set the backend to 'Agg' to suppress plot window
+
+
+class TestHistogramPerformance(unittest.TestCase):
+    """Performance comparison tests for histogram vs histogram_old."""
+
+    @classmethod
+    def setUpClass(cls):
+        """Generate large datasets once for all tests."""
+        print("\n" + "=" * 70)
+        print("Setting up large datasets for histogram performance tests...")
+        print("=" * 70)
+        
+        # Generate 1M cell dataset
+        print("\nGenerating 1M cell dataset...")
+        start = time.time()
+        cls.adata_1m = cls._generate_dataset(n_obs=1_000_000, random_state=42)
+        print(f"  Completed in {time.time() - start:.2f} seconds")
+        
+        # Generate 5M cell dataset
+        print("\nGenerating 5M cell dataset...")
+        start = time.time()
+        cls.adata_5m = cls._generate_dataset(n_obs=5_000_000, random_state=42)
+        print(f"  Completed in {time.time() - start:.2f} seconds")
+        
+        # Generate 10M cell dataset
+        print("\nGenerating 10M cell dataset...")
+        start = time.time()
+        cls.adata_10m = cls._generate_dataset(n_obs=10_000_000, random_state=42)
+        print(f"  Completed in {time.time() - start:.2f} seconds")
+        print("=" * 70 + "\n")
+
+    @staticmethod
+    def _generate_dataset(n_obs: int, random_state: int = 42) -> ad.AnnData:
+        """
+        Generate a synthetic AnnData object with realistic clustering.
+        
+        Creates dataset with:
+        - 5 features (marker_1 to marker_5)
+        - 5 annotations (cell_type, phenotype, region, batch, treatment)
+        - 3 layers (normalized, log_transformed, scaled)
+        """
+        np.random.seed(random_state)
+        
+        # Generate base data with natural clustering
+        n_features = 5
+        n_centers = 5
+        
+        X, cluster_labels = make_blobs(
+            n_samples=n_obs,
+            n_features=n_features,
+            centers=n_centers,
+            cluster_std=1.5,
+            random_state=random_state
+        )
+        
+        # Make values positive and add variation
+        X = np.abs(X) + np.random.exponential(scale=2.0, size=X.shape)
+        
+        # Create feature names
+        feature_names = [f"marker_{i+1}" for i in range(n_features)]
+        
+        # Create annotations based on clusters
+        cell_types = [f"Type_{chr(65+i)}" for i in range(5)]
+        cell_type = np.array([cell_types[i % 5] for i in cluster_labels])
+        
+        phenotypes = [f"Pheno_{i+1}" for i in range(4)]
+        phenotype = np.array([phenotypes[i % 4] for i in cluster_labels])
+        random_mask = np.random.random(n_obs) < 0.2
+        phenotype[random_mask] = np.random.choice(phenotypes, size=random_mask.sum())
+        
+        regions = ["Region_X", "Region_Y", "Region_Z"]
+        region = np.random.choice(regions, size=n_obs)
+        
+        batches = ["Batch_1", "Batch_2", "Batch_3"]
+        batch = np.random.choice(batches, size=n_obs)
+        
+        treatments = ["Control", "Treated"]
[…]
