+++ tests/test_performance/__init__.py
@@ -1,0 +1,1 @@
+"""Performance tests for SPAC visualization functions."""
