### 2025-10-19 c0762c3
fix(scripts): remove old performance testing script
