+++ scripts/visualization_benchmark.py
@@ -1,798 +1,0 @@
-#!/usr/bin/env python3
-"""
-Visualization Benchmark Script
-===============================
-
-A command-line tool for benchmarking visualization functions from the SPAC
-package. This script measures the performance of various plotting functions
-with different dataset sizes and configurations.
-
-Usage:
-    # Benchmark with a custom dataset
-    python visualization_benchmark.py --dataset path/to/data.pkl --visualizations boxplot histogram
-
-    # Benchmark with generated random data
-    python visualization_benchmark.py --min-datapoints 1000 --max-datapoints 10000 \
-        --increment 1000 --visualizations both --output-plot
-
-Author: SPAC Development Team
-Date: October 2025
-"""
-
-import argparse
-import sys
-import time
-import pickle
-from pathlib import Path
-from typing import Dict, List, Tuple
-
-import numpy as np
-import pandas as pd
-import anndata as ad
-import matplotlib.pyplot as plt
-from sklearn.datasets import make_blobs
-from sklearn.preprocessing import StandardScaler
-
-# Import SPAC visualization functions
-from spac.visualization import boxplot, boxplot_interactive
-
-
-def parse_arguments() -> argparse.Namespace:
-    """
-    Parse command-line arguments for the benchmark script.
-
-    Returns
-    -------
-    argparse.Namespace
-        Parsed command-line arguments containing benchmark configuration.
-    """
-    parser = argparse.ArgumentParser(
-        description="Benchmark visualization functions from the SPAC package",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  # Benchmark with custom dataset
-  %(prog)s --dataset data.pkl --visualizations boxplot histogram --output-plot
-
-  # Benchmark with a single dataset size
-  %(prog)s --size 5000 --visualizations both --output-plot
-
-  # Benchmark with generated data range
-  %(prog)s --min-datapoints 1000 --max-datapoints 5000 --increment 1000 \\
-           --visualizations both --output-plot
-
-  # Run only histogram benchmarks with single size
-  %(prog)s --size 2000 --visualizations histogram
-        """
-    )
-
-    # Dataset input options
-    data_group = parser.add_argument_group("Dataset Options")
-    data_group.add_argument(
-        "--dataset",
-        type=str,
-        default=None,
-        metavar="PATH",
-        help="Path to input dataset file (pickle or anndata-readable format). "
-             "If not provided, random datasets will be generated."
-    )
-
-    # Random data generation options
-    random_group = parser.add_argument_group(
-        "Random Data Generation Options",
-        "Used when --dataset is not provided"
-    )
-    random_group.add_argument(
-        "--size",
-        type=int,
-        default=None,
-        metavar="N",
-        help="Single dataset size to benchmark. If specified, overrides "
-             "min/max/increment options."
-    )
-    random_group.add_argument(
-        "--min-datapoints",
-        type=int,
-        default=1000,
-        metavar="N",
-        help="Minimum number of datapoints for generated datasets (default: 1000). "
-             "Ignored if --size is specified."
-    )
-    random_group.add_argument(
-        "--max-datapoints",
-        type=int,
-        default=10000,
-        metavar="N",
-        help="Maximum number of datapoints for generated datasets (default: 10000). "
-             "Ignored if --size is specified."
-    )
-    random_group.add_argument(
-        "--increment",
-        type=int,
-        default=1000,
-        metavar="N",
-        help="Increment step for datapoint sizes (default: 1000). "
-             "Ignored if --size is specified."
-    )
-
-    # Visualization options
[…]
