### 2025-10-20 c31c3ff
fix(performance_test): fix the speedup calculation logic
