### 2025-10-21 179482e
fix(histogram_performance): add clarifying comment for old hist implementation
