### 2025-10-23 633e31d
feat: Add comprehensive Docker setup and testing workflow for reviewers
- Fix README.md installation instructions with missing 'pip install -e .' step
- Add complete Dockerfile with volume mapping support for reproducible environment
- Create standalone test_notebook_execution.sh script with:
  * SPAC installation verification
  * Timestamped output files to prevent conflicts
  * Interactive Jupyter launch option
  * Smart Docker vs local environment detection
- Update environment.yml to use pip for scimap (ARM64 compatibility)
- Add Docker port conflict resolution and cleanup commands
- Provi
[…]
