### 2025-10-23 3099ddd
Add performance validation documentation and test script
- Added comprehensive performance validation section to README.md
- Added test_performance_validation.sh script for Docker-based performance testing
- Documents how to run boxplot and histogram performance benchmarks
- Includes system requirements and output file descriptions
- Validates performance improvements with up to 10M cell datasets
