### 2025-10-25 0c2df5c
Update README with conda setup instructions
Added instructions for checking out the address_reviewer_comments branch and activating conda.
