### 2025-10-25 ba0c6d5
Modify dependencies in environment.yml
Updated numpy version to be greater than or equal to 1.26.4 and less than 2.0.0. Added phenograph dependency under pip.
