+++ environment.yml
@@ -6,13 +6,12 @@ channels:
   - https://fnlcr-dmap.github.io/scimap/
 dependencies:
   - python=3.9.13               # Core Python Package
-  - numpy=1.26.4                # Core Python Package
+  - numpy>=1.26.4,<2.0.0        # Core Python Package
   - pandas=1.5.3                # Data Manipulation and Analysis
   - anndata=0.10.9              # Single-Cell Data Manipulation and Analysis (scverse family)
   - scanpy=1.10.3               # Single-Cell Data Manipulation and Analysis (scverse family)
   - zarr=2.10.3                 # Data Manipulation and Analysis
   - numba=0.58.1                # Data Manipulation and Analysis
-  - phenograph=1.5.7            # Data Manipulation and Analysis
   - scikit-learn=1.5.2          # Machine Learning and Statistics
   - matplotlib=3.9.2            # Visualization and Plotting
   - seaborn=0.13.2              # Visualization and Plotting
@@ -29,3 +28,5 @@ dependencies:
     - gendocs
     - parmap==1.7.0
     - scimap
+    - phenograph=1.5.7
+
