### 2025-10-25 05dd565
Fix git checkout command in README
