### 2025-10-25 97d41bb
Fix formatting and dependencies in environment.yml
Updated the environment.yml file to correct formatting and ensure proper dependency listing.
