### 2025-10-25 ae5b6c8
Update environment.yml
