+++ environment.yml
@@ -6,12 +6,14 @@ channels:
   - https://fnlcr-dmap.github.io/scimap/
 dependencies:
   - python=3.9.13               # Core Python Package
-  - numpy>=1.26.4,<2.0.0        # Core Python Package
+  - numpy=1.26.4                # Core Python Package
   - pandas=1.5.3                # Data Manipulation and Analysis
   - anndata=0.10.9              # Single-Cell Data Manipulation and Analysis (scverse family)
+  - scimap=2.1.3_dmap_pandas153 # Single-Cell Data Data Manipulation and Analysis
   - scanpy=1.10.3               # Single-Cell Data Manipulation and Analysis (scverse family)
   - zarr=2.10.3                 # Data Manipulation and Analysis
   - numba=0.58.1                # Data Manipulation and Analysis
+  - phenograph=1.5.7            # Data Manipulation and Analysis
   - scikit-learn=1.5.2          # Machine Learning and Statistics
   - matplotlib=3.9.2            # Visualization and Plotting
   - seaborn=0.13.2              # Visualization and Plotting
@@ -27,7 +29,3 @@ dependencies:
     - flake8
     - gendocs
     - parmap==1.7.0
-    - scimap
-    - phenograph==1.5.7
-
-
