### 2025-10-25 1c042df
Correct README instructions for Conda setup
Fix typo in README and update checkout command.
