### 2025-10-27 b28174a
feat: update manuscript and jupyter notebook
