+++ paper/paper.bib
@@ -210,3 +210,37 @@
     urldate = {2025-04-15},
     year = {2003}
 }
+
+
+@article{Chen:2023,
+  title = {Giotto Suite: a multi-scale and technology-agnostic spatial multi-omics analysis ecosystem},
+  author = {Chen, Jiaji George and Ch{\'a}vez-Fuentes, Joselyn Cristina and O'Brien, Matthew and Xu, Junxiang and Ruiz, Edward and Wang, Wen and Amin, Iqra and Sheridan, Jeffrey and Shin, Sujung Crystal and Hasyagar, Sanjana Varada and Sarfraz, Irzam and Guckhool, Pratishtha and Sistig, Adriana and Jarzabek, Veronica and Yuan, Guo-Cheng and Dries, Ruben},
+  journal = {bioRxiv},
+  year = {2023},
+  doi = {10.1101/2023.11.26.568752},
+  url = {https://www.biorxiv.org/content/10.1101/2023.11.26.568752v1},
+  note = {Preprint}
+}
+
+
+@article{Virshup:2024,
+  author  = {Virshup, Isaac and Rybakov, Sergey and Theis, Fabian J. and Angerer, Philipp and Wolf, Falko A.},
+  title   = {anndata: Access and store annotated data matrices},
+  journal = {Journal of Open Source Software},
+  year    = {2024},
+  volume  = {9},
+  pages   = {4371},
+  doi     = {10.21105/joss.04371},
+  url     = {https://doi.org/10.21105/joss.04371}
+}
+
+
+@article{Wolf:2018,
+  author  = {Wolf, Falko A. and Angerer, Philipp and Theis, Fabian J.},
+  title   = {SCANPY: Large-scale single-cell gene expression data analysis},
+  journal = {Genome Biology},
+  year    = {2018},
+  volume  = {19},
+  doi     = {10.1186/s13059-017-1382-0},
+  url     = {https://doi.org/10.1186/s13059-017-1382-0}
+}
