### 2025-10-27 #8787 @fangliu117
> Thanks for the updates, [@fangliu117](https://github.com/fangliu117) ! I apologize for the delay. I have written my review below, focusing on each open point on the reviewer checklist.
>
> ## General checks
> ### Human and animal research
> The example notebook says the data is "Normal human lymph node tissue microarray (TMA) from CyCIF (Cyclic Immunofluorescence) imaging, preprocessed by MCMICRO, downloaded as "example_lymphnode_data.csv". The source of this data is not clear and thus, I am not sure if it needs an any reporting on the human subjects that data came from.
>
Thank you @kevinyamauchi for your suggestion, we should have the source of the lymph node data more clear. We have added a data availability statement to the [example notebook](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/address-reviewer-comments/paper/examples/lymphnode_analysis.ipynb) clarifying that this is publicly available, de-identified data that does not require additional human subjects reporting. The data statement now reads: "Source: This example uses scimap demo dataset (https://scimap.xyz/tutorials/md/demo_data_scimap/), with additional annotations added for phenotyping analysis named as "example_lymphnode_data.csv". The original data is publicly available normal human lymph node tissue microarray data from CyCIF imaging, preprocessed by MCMICRO. No additional ethical approvals are required for this de-identified published dataset."
