### 2025-10-30 9c4d606
fix(test_performance): set the path to include spac
