+++ tests/test_performance/__init__.py
@@ -1,1 +1,3 @@
-"""Performance tests for SPAC visualization functions."""
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.realpath(__file__)) + "/../../src")
