### 2025-11-03 315b3bf
feat: mvp_boxplot galaxy tool and refact templates
