+++ src/spac/templates/template_utils_refactored.py
@@ -1,0 +1,805 @@
+from pathlib import Path
+import pickle
+from typing import Any, Dict, Union, Optional, List
+import json
+import pandas as pd
+import anndata as ad
+import re
+import logging
+import matplotlib.pyplot as plt
+
+logger = logging.getLogger(__name__)
+
+
+def load_input(file_path: Union[str, Path]):
+    """
+    Load input data from either h5ad or pickle file.
+
+    Parameters
+    ----------
+    file_path : str or Path
+        Path to input file (h5ad or pickle)
+
+    Returns
+    -------
+    Loaded data object (typically AnnData)
+    """
+    path = Path(file_path)
+
+    if not path.exists():
+        raise FileNotFoundError(f"Input file not found: {file_path}")
+
+    # Check file extension
+    suffix = path.suffix.lower()
+
+    if suffix in ['.h5ad', '.h5']:
+        # Load h5ad file
+        try:
+            import anndata as ad
+            return ad.read_h5ad(path)
+        except ImportError:
+            raise ImportError(
+                "anndata package required to read h5ad files"
+            )
+        except Exception as e:
+            raise ValueError(f"Error reading h5ad file: {e}")
+
+    elif suffix in ['.pickle', '.pkl', '.p']:
+        # Load pickle file
+        with path.open('rb') as fh:
+            return pickle.load(fh)
+
+    else:
+        # Try to detect file type by content
+        try:
+            # First try h5ad
+            import anndata as ad
+            return ad.read_h5ad(path)
+        except Exception:
+            # Fall back to pickle
+            try:
+                with path.open('rb') as fh:
+                    return pickle.load(fh)
+            except Exception as e:
+                raise ValueError(
+                    f"Unable to load file '{file_path}'. "
+                    f"Supported formats: h5ad, pickle. Error: {e}"
+                )
+
+
+def save_results(
+    results: Dict[str, Any],
+    outputs_config: Dict[str, str],
+    output_base_dir: Union[str, Path] = ".",
+    params: Optional[Dict[str, Any]] = None
+) -> Dict[str, List[str]]:
+    """
+    Save results to appropriate directories based on blueprint outputs configuration.
+    
+    Parameters
+    ----------
+    results : dict
+        Dictionary of results to save where:
+        - key: result type ("analysis", "dataframes", "figures", "html") or specific filename
+        - value: object(s) to save (can be single object or list/dict of objects)
+    outputs_config : dict
+        Output configuration from blueprint JSON, e.g.:
+        {"analysis": "transform_output.pickle", "DataFrames": "dataframe_folder", 
+         "figures": "figure_folder", "html": "html_folder"}
+    output_base_dir : str or Path
+        Base directory for all outputs
+    params : dict, optional
+        Optional parameters dict for accessing specific output paths
+        
+    Returns
+    -------
+    dict
+        Dictionary mapping output types to lists of saved file paths
+        
+    Examples
+    --------
+    >>> # For boxplot with multiple figures and dataframes
+    >>> results = {
+    ...     "figures": [fig1, fig2, fig3],
+    ...     "dataframes": {"summary": summary_df, "stats": stats_df}
+    ... }
+    >>> outputs_config = {"DataFrames": "dataframe_folder", "figures": "figure_folder"}
+    >>> saved = save_results(results, outputs_config)
+    """
+    output_base_dir = Path(output_base_dir)
+    saved_files = {}
+    
+    # Normalize outputs_config keys to lowercase for matching
+    config_map = {k.lower(): (k, v) for k, v in outputs_config.items()}
+    
+    for result_type, data in results.items():
+        result_key = result_type.lower()
+        
+        # Find matching config
+        if result_key in config_map:
+            original_key, output_dest = config_map[result_key]
+        else:
[…]
