+++ src/spac/templates/boxplot_template_refactored.py
@@ -1,0 +1,234 @@
+"""
+Platform-agnostic Boxplot template converted from NIDAP.
+Refactored to use generalized save_results with blueprint outputs configuration.
+
+Usage
+-----
+>>> from spac.templates.boxplot_template import run_from_json
+>>> run_from_json("examples/boxplot_params.json")
+"""
+import json
+import sys
+from pathlib import Path
+from typing import Any, Dict, Union, List, Optional, Tuple
+import logging
+import pandas as pd
+import matplotlib.pyplot as plt
+import seaborn as sns
+
+# Add parent directory to path for imports
+sys.path.append(str(Path(__file__).parent.parent.parent))
+
+from spac.visualization import boxplot
+from spac.templates.template_utils import (
+    load_input,
+    save_results,
+    parse_params,
+    text_to_value,
+)
+
+
+def run_from_json(
+    json_path: Union[str, Path, Dict[str, Any]],
+    save_results_flag: bool = True,
+    show_plot: bool = True,
+    outputs_config: Optional[Dict[str, str]] = None
+) -> Union[Dict[str, List[str]], Tuple[Any, pd.DataFrame]]:
+    """
+    Execute Boxplot analysis with parameters from JSON.
+    Replicates the NIDAP template functionality with generalized output handling.
+
+    Parameters
+    ----------
+    json_path : str, Path, or dict
+        Path to JSON file, JSON string, or parameter dictionary
+    save_results_flag : bool, optional
+        Whether to save results to file. If False, returns the figure and
+        summary dataframe directly for in-memory workflows. Default is True.
+    show_plot : bool, optional
+        Whether to display the plot. Default is True.
+    outputs_config : dict, optional
+        Output configuration from blueprint JSON. If None, uses default:
+        {"DataFrames": "dataframe_folder", "figures": "figure_folder"}
+
+    Returns
+    -------
+    dict or tuple
+        If save_results_flag=True: Dictionary of saved file paths
+        If save_results_flag=False: Tuple of (figure, summary_dataframe)
+    """
+    # Parse parameters from JSON
+    params = parse_params(json_path)
+
+    # Get outputs configuration - from params or use default
+    if outputs_config is None:
+        # Try to get from params (if blueprint JSON was passed)
+        outputs_config = params.get("outputs", {
+            "DataFrames": "dataframe_folder",
+            "figures": "figure_folder"
+        })
+
+    # Load the upstream analysis data
+    adata = load_input(params["Upstream_Analysis"])
+
+    # Extract parameters
+    annotation = params.get("Primary_Annotation", "None")
+    second_annotation = params.get("Secondary_Annotation", "None")
+    layer_to_plot = params.get("Table_to_Visualize", "Original")
+    feature_to_plot = params.get("Feature_s_to_Plot", ["All"])
+    log_scale = params.get("Value_Axis_Log_Scale", False)
+    
+    # Figure parameters
+    figure_title = params.get("Figure_Title", "BoxPlot")
+    figure_horizontal = params.get("Horizontal_Plot", False)
+    fig_width = params.get("Figure_Width", 12)
+    fig_height = params.get("Figure_Height", 8)
+    fig_dpi = params.get("Figure_DPI", 300)
+    font_size = params.get("Font_Size", 10)
+    showfliers = params.get("Keep_Outliers", True)
+
+    # Process parameters exactly as in NIDAP template
+    if layer_to_plot == "Original":
+        layer_to_plot = None
+    
+    if second_annotation == "None":
+        second_annotation = None
+    
+    if annotation == "None":
+        annotation = None
+
+    if figure_horizontal:
+        figure_orientation = "h"
+    else:
+        figure_orientation = "v"
+
[…]

[…]
