### 2025-11-05 79335a0
refactor: template outputs: integrate save_results into template_utils
