+++ tests/templates/test_template_utils.py
@@ -13,6 +13,7 @@ import anndata as ad
 import numpy as np
 import pandas as pd
 from pathlib import Path
+import matplotlib.pyplot as plt
 
 sys.path.append(
     os.path.dirname(os.path.realpath(__file__)) + "/../../src"
@@ -20,7 +21,8 @@ sys.path.append(
 
 from spac.templates.template_utils import (
     load_input,
-    save_outputs,
+    save_results,
+    _save_single_object,
     text_to_value,
     convert_pickle_to_h5ad,
     convert_to_floats,
@@ -93,29 +95,7 @@ class TestTemplateUtils(unittest.TestCase):
             loaded_p = load_input(p_path)
             self.assertEqual(loaded_p.n_obs, 10)
 
-            # Test 5: Save outputs - multiple formats
-            outputs = {
-                "result.pickle": self.test_adata,  # Now preferred format
-                "data.csv": self.test_df,
-                "adata.pkl": self.test_adata,
-                "adata.h5ad": self.test_adata,  # Still supported
-                "other_data": {"key": "value"}  # Defaults to pickle
-            }
-            saved_files = save_outputs(outputs, self.tmp_dir.name)
-
-            # Verify all files were saved
-            self.assertEqual(len(saved_files), 5)
-            for filename, filepath in saved_files.items():
-                self.assertTrue(os.path.exists(filepath))
-                self.assertIn(filename, saved_files)
-
-            # Verify CSV content
-            csv_path = saved_files["data.csv"]
-            loaded_df = pd.read_csv(csv_path)
-            self.assertEqual(len(loaded_df), 5)
-            self.assertIn("col1", loaded_df.columns)
-
-            # Test 6: Convert pickle to h5ad
+            # Test 5: Convert pickle to h5ad
             pickle_src = os.path.join(
                 self.tmp_dir.name, "convert_src.pickle"
             )
@@ -281,18 +261,6 @@ class TestTemplateUtils(unittest.TestCase):
         expected_msg = "Loaded object is not AnnData, got <class 'dict'>"
         actual_msg = str(context.exception)
         self.assertEqual(expected_msg, actual_msg)
-
-    def test_default_pickle_extension(self) -> None:
-        """Test that files without extension default to pickle."""
-        outputs = {
-            "no_extension": self.test_adata
-        }
-        saved_files = save_outputs(outputs, self.tmp_dir.name)
-
-        # Should have .pickle extension
-        filepath = saved_files["no_extension"]
-        self.assertTrue(filepath.endswith('.pickle'))
-        self.assertTrue(os.path.exists(filepath))
 
     def test_spell_out_special_characters(self) -> None:
         """Test spell_out_special_characters function."""
@@ -518,6 +486,318 @@ class TestTemplateUtils(unittest.TestCase):
         final_info = [msg for msg in print_calls 
                      if 'Final Dataframe Info' in msg]
         self.assertTrue(len(final_info) > 0)
+
+    def test_save_results_single_csv_file(self) -> None:
+        """Test saving DataFrame as single CSV file using save_results."""
+        # Setup
+        df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
+        
+        config = {
+            "DataFrames": {"type": "file", "name": "data.csv"}
+        }
+        
+        results = {
+            "dataframes": df
+        }
+        
+        # Execute
+        saved = save_results(results, config, self.tmp_dir.name)
+        
+        # Verify
+        csv_path = Path(self.tmp_dir.name) / "data.csv"
+        self.assertTrue(csv_path.exists())
+        self.assertTrue(csv_path.is_file())
+        
+        # Check content
[…]
