+++ src/spac/templates/template_utils.py
@@ -6,6 +6,8 @@ import pandas as pd
 import anndata as ad
 import re
 import logging
+import matplotlib.pyplot as plt
+
 logger = logging.getLogger(__name__)
 
 
@@ -33,7 +35,6 @@ def load_input(file_path: Union[str, Path]):
     if suffix in ['.h5ad', '.h5']:
         # Load h5ad file
         try:
-            import anndata as ad
             return ad.read_h5ad(path)
         except ImportError:
             raise ImportError(
@@ -51,7 +52,6 @@ def load_input(file_path: Union[str, Path]):
         # Try to detect file type by content
         try:
             # First try h5ad
-            import anndata as ad
             return ad.read_h5ad(path)
         except Exception:
             # Fall back to pickle
@@ -65,77 +65,225 @@ def load_input(file_path: Union[str, Path]):
                 )
 
 
-def save_outputs(
-    outputs: Dict[str, Any],
-    output_dir: Union[str, Path] = "."
-) -> Dict[str, str]:
+def save_results(
+    results: Dict[str, Any],
+    outputs_config: Dict[str, Dict[str, str]],
+    output_base_dir: Union[str, Path] = ".",
+    params: Optional[Dict[str, Any]] = None
+) -> Dict[str, List[str]]:
     """
-    Save multiple outputs to files and return a dict {filename: absolute_path}.
-    (Always a dict, even if just one file.)
-
+    Save results to appropriate locations based on explicit typed blueprint outputs.
+    
     Parameters
     ----------
-    outputs : dict
-        Dictionary where:
-        - key: filename (with extension)
-        - value: object to save
-    output_dir : str or Path
-        Directory to save files
-
+    results : dict
+        Dictionary of results to save where:
+        - key: result type ("analysis", "dataframes", "figures", "html")
+        - value: object(s) to save (can be single object or list/dict of objects)
+    outputs_config : dict
+        Explicit typed output configuration from blueprint JSON:
+        {
+            "figures": {"type": "directory", "name": "figure_dir"},
+            "DataFrames": {"type": "file", "name": "dataframe.csv"},
+            "DataFrames": {"type": "directory", "name": "dataframe_dir"},
+            "analysis": {"type": "file", "name": "output.pickle"},
+            "html": {"type": "directory", "name": "html_dir"}
+        }
+    output_base_dir : str or Path, optional
+        Base directory for all outputs. Default is current directory.
+    params : dict, optional
+        Optional parameters dict (for backward compatibility)
+        
     Returns
     -------
     dict
-        Dictionary of saved file paths
-
+        Dictionary mapping output types to lists of saved file paths
+        
     Example
     -------
-    >>> outputs = {
-    ...     "adata.pickle": adata,  # Preferred format
-    ...     "results.csv": results_df,
-    ...     "adata.h5ad": adata  # Still supported
+    >>> outputs_config = {
+    ...     "figures": {"type": "directory", "name": "plots"},
+    ...     "analysis": {"type": "file", "name": "results.pickle"}
     ... }
-    >>> saved = save_outputs(outputs, "results/")
+    >>> results = {
+    ...     "figures": {"plot1": fig1, "plot2": fig2},
+    ...     "analysis": adata
+    ... }
+    >>> saved = save_results(results, outputs_config, "outputs/")
     """
-    output_dir = Path(output_dir)
-    output_dir.mkdir(parents=True, exist_ok=True)
-
+    output_base_dir = Path(output_base_dir)
     saved_files = {}
-
-    for filename, obj in outputs.items():
-        filepath = output_dir / filename
-
-        # Save based on file extension
-        if filename.endswith('.csv'):
-            obj.to_csv(filepath, index=False)
[…]
