### 2025-11-05 d703f55
refactor: boxplot_template outputs: call save_results from template_utils
