+++ tests/templates/test_template_utils_refactored.py
@@ -1,425 +1,0 @@
-"""
-Unit tests for refactored template_utils functions.
-Focus on testing the new save_results function and related utilities.
-"""
-import pytest
-import tempfile
-import shutil
-import json
-from pathlib import Path
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-import pickle
-import sys
-
-# Add parent directory for imports
-sys.path.append(str(Path(__file__).parent))
-from template_utils_refactored import (
-    save_results,
-    save_outputs,  # Legacy function
-    load_input,
-    parse_params,
-    parse_list_parameter,
-    text_to_value,
-)
-
-
-class TestSaveResults:
-    """Test the new generalized save_results function."""
-    
-    def setup_method(self):
-        """Setup test fixtures."""
-        self.test_dir = Path(tempfile.mkdtemp(prefix="test_template_utils_"))
-        
-    def teardown_method(self):
-        """Clean up test directory."""
-        if self.test_dir.exists():
-            shutil.rmtree(self.test_dir)
-        plt.close('all')
-    
-    def test_save_results_with_folders(self):
-        """Test saving results to organized folders."""
-        # Create test data
-        df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
-        fig, ax = plt.subplots()
-        ax.plot([1, 2, 3])
-        
-        # Blueprint configuration
-        outputs_config = {
-            "DataFrames": "dataframe_folder",
-            "figures": "figure_folder"
-        }
-        
-        # Results to save
-        results = {
-            "dataframes": {"test_data": df},
-            "figures": {"test_plot": fig}
-        }
-        
-        # Save results
-        saved_files = save_results(
-            results=results,
-            outputs_config=outputs_config,
-            output_base_dir=self.test_dir
-        )
-        
-        # Verify structure
-        assert "DataFrames" in saved_files
-        assert "figures" in saved_files
-        assert len(saved_files["DataFrames"]) == 1
-        assert len(saved_files["figures"]) == 1
-        
-        # Check directories were created
-        assert (self.test_dir / "dataframe_folder").exists()
-        assert (self.test_dir / "figure_folder").exists()
-        
-        # Check files exist
-        assert (self.test_dir / "dataframe_folder" / "test_data.csv").exists()
-        assert (self.test_dir / "figure_folder" / "test_plot.png").exists()
-    
-    def test_save_results_with_specific_file(self):
-        """Test saving results to specific file paths."""
-        # Create mock analysis object
-        class MockAnnData:
-            def __init__(self):
-                self.X = np.array([[1, 2], [3, 4]])
-        
-        adata = MockAnnData()
-        
-        # Blueprint configuration with specific file
-        outputs_config = {
-            "analysis": "transform_output.pickle"
-        }
-        
-        # Results to save
-        results = {
-            "analysis": adata
-        }
-        
-        # Save results
-        saved_files = save_results(
-            results=results,
-            outputs_config=outputs_config,
-            output_base_dir=self.test_dir
-        )
-        
-        # Verify
-        assert "analysis" in saved_files
-        assert len(saved_files["analysis"]) == 1
-        assert (self.test_dir / "transform_output.pickle").exists()
-        
-        # Verify content can be loaded
-        with open(self.test_dir / "transform_output.pickle", 'rb') as f:
-            loaded = pickle.load(f)
-            assert np.array_equal(loaded.X, adata.X)
-    
-    def test_save_results_with_lists(self):
-        """Test saving multiple items as lists."""
-        # Create multiple figures
-        figs = []
-        for i in range(3):
-            fig, ax = plt.subplots()
-            ax.plot([i, i+1, i+2])
[…]
