### 2025-11-05 cdf0102
refactor: modified test_boxplot_template
