+++ tests/templates/test_boxplot_template.py
@@ -8,7 +8,7 @@ import sys
 import tempfile
 import unittest
 import warnings
-from unittest.mock import patch, MagicMock
+from unittest.mock import patch, MagicMock, Mock
 from pathlib import Path
 
 import matplotlib
@@ -116,7 +116,7 @@ class TestBoxplotTemplate(unittest.TestCase):
             # Test with save_to_file=True (default)
             saved_files = run_from_json(self.params)
             
-            # Check that save_results created the expected structure
+            # Check that save_to_file created the expected structure
             self.assertIn("figure", saved_files)
             self.assertIn("summary", saved_files)
             
@@ -144,9 +144,9 @@ class TestBoxplotTemplate(unittest.TestCase):
         })
         mock_boxplot.return_value = (mock_fig, mock_ax, mock_df)
         
-        # Test with save_results=False
+        # Test with save_to_file=False
         fig, summary_df = run_from_json(
-            self.params, save_results=False
+            self.params, save_to_file=False
         )
         
         # Verify we got the figure and summary dataframe back
@@ -157,12 +157,21 @@ class TestBoxplotTemplate(unittest.TestCase):
         self.assertIn('mean', summary_df['index'].values)
         self.assertIn('std', summary_df['index'].values)
 
+    @patch('spac.templates.boxplot_template.save_results')
     @patch('spac.templates.boxplot_template.boxplot')
     @patch('matplotlib.pyplot.show')
-    def test_blueprint_configuration(self, mock_show, mock_boxplot) -> None:
+    def test_blueprint_configuration(self, mock_show, mock_boxplot, mock_save_results) -> None:
         """Test blueprint configuration for outputs."""
         mock_fig, mock_ax, mock_df = self._create_mock_boxplot_return()
         mock_boxplot.return_value = (mock_fig, mock_ax, mock_df)
+        
+        # Add this to mock save_results and create files
+        mock_save_results.return_value = {
+            "figure": [str(Path(self.tmp_dir.name) / "boxplot.png")],
+            "summary": [str(Path(self.tmp_dir.name) / "summary.csv")]
+        }
+        (Path(self.tmp_dir.name) / "boxplot.png").touch()
+        (Path(self.tmp_dir.name) / "summary.csv").touch()
         
         # Test with default blueprint configuration
         saved_files = run_from_json(self.params, save_to_file=True)
@@ -178,9 +187,10 @@ class TestBoxplotTemplate(unittest.TestCase):
         summary_path = Path(self.tmp_dir.name) / "summary.csv"
         self.assertTrue(summary_path.exists())
 
+    @patch('spac.templates.boxplot_template.save_results')
     @patch('spac.templates.boxplot_template.boxplot')
     @patch('matplotlib.pyplot.show')
-    def test_custom_blueprint_in_params(self, mock_show, mock_boxplot) -> None:
+    def test_custom_blueprint_in_params(self, mock_show, mock_boxplot, mock_save_results) -> None:
         """Test custom blueprint configuration passed in params."""
         mock_fig, mock_ax, mock_df = self._create_mock_boxplot_return()
         mock_boxplot.return_value = (mock_fig, mock_ax, mock_df)
@@ -191,6 +201,14 @@ class TestBoxplotTemplate(unittest.TestCase):
             "figure": {"type": "file", "name": "my_plot.pdf"},
             "summary": {"type": "file", "name": "my_stats.csv"}
         }
+
+        # Mock save_results and create files
+        mock_save_results.return_value = {
+            "figure": [str(Path(self.tmp_dir.name) / "my_plot.pdf")],
+            "summary": [str(Path(self.tmp_dir.name) / "my_stats.csv")]
+        }
[…]
