### 2025-11-05 3ba042d
refactor: boxplot mvp tool with format_values script
