### 2025-11-07 a8b6e13
refactor(boxplot_template): use centralized save_results from template_utils
