+++ tests/templates/test_boxplot_template.py
@@ -91,15 +91,20 @@ class TestBoxplotTemplate(unittest.TestCase):
             "Font_Size": 10,
             "Keep_Outliers": True,
             "Output_File": self.out_file,
+            "outputs": {
+                "figures": {"type": "file", "name": "boxplot.png"},
+                "dataframe": {"type": "file", "name": "summary.csv"}
+            }
         }
 
     def tearDown(self) -> None:
         self.tmp_dir.cleanup()
 
+    @patch('spac.templates.boxplot_template.save_results')
     @patch('spac.templates.boxplot_template.boxplot')
     @patch('matplotlib.pyplot.show')
-    def test_run_with_save(self, mock_show, mock_boxplot) -> None:
-        """Test boxplot with file saving."""
+    def test_run_with_save(self, mock_show, mock_boxplot, mock_save_results) -> None:
+        """Test boxplot with saving results to disk."""
         # Mock the boxplot function to return figure, ax, and dataframe
         rng = np.random.default_rng(42)
         mock_fig, mock_ax, mock_df = self._create_mock_boxplot_return({
@@ -109,33 +114,49 @@ class TestBoxplotTemplate(unittest.TestCase):
         })
         mock_boxplot.return_value = (mock_fig, mock_ax, mock_df)
         
+        # Mock save_results to return expected structure
+        mock_save_results.return_value = {
+            "figures": [str(Path(self.tmp_dir.name) / "boxplot.png")],
+            "dataframe": [str(Path(self.tmp_dir.name) / "summary.csv")]
+        }
+        
         # Suppress warnings for cleaner test output
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
 
-            # Test with save_to_file=True (default)
+            # Test with save_to_disk=True (default)
             saved_files = run_from_json(self.params)
             
-            # Check that save_to_file created the expected structure
-            self.assertIn("figure", saved_files)
-            self.assertIn("summary", saved_files)
+            # Check that save_results was called with correct params
+            mock_save_results.assert_called_once()
+            call_args = mock_save_results.call_args
             
+            # Verify params were passed correctly
+            self.assertIn('results', call_args[1])
+            self.assertIn('params', call_args[1])
+            self.assertEqual(call_args[1]['params'], self.params)
+            
+            # Check that results contain figure and summary
+            results = call_args[1]['results']
+            self.assertIn("figures", results)
+            self.assertIn("dataframe", results)
+
             # Verify boxplot was called with correct parameters
             mock_boxplot.assert_called_once()
-            call_args = mock_boxplot.call_args
+            boxplot_args = mock_boxplot.call_args
             
             # Check keyword arguments
-            self.assertEqual(call_args[1]['annotation'], "cell_type")
-            self.assertEqual(call_args[1]['second_annotation'], None)
-            self.assertEqual(call_args[1]['layer'], None)  # Original -> None
-            self.assertEqual(call_args[1]['log_scale'], False)
-            self.assertEqual(call_args[1]['orient'], "v")
-            self.assertEqual(call_args[1]['showfliers'], True)
+            self.assertEqual(boxplot_args[1]['annotation'], "cell_type")
+            self.assertEqual(boxplot_args[1]['second_annotation'], None)
+            self.assertEqual(boxplot_args[1]['layer'], None)  # Original -> None
+            self.assertEqual(boxplot_args[1]['log_scale'], False)
[…]
