### 2025-11-07 80f7e3a
refactor: template_utils and setup_analysis_template for save_results
