+++ tests/templates/test_setup_analysis_template.py
@@ -55,7 +55,7 @@ class TestSetupAnalysisTemplate(unittest.TestCase):
         )
         self.test_df.to_csv(self.csv_file, index=False)
 
-        # Minimal parameters
+        # Minimal parameters with standardized outputs config
         self.params = {
             "Upstream_Dataset": self.csv_file,
             "Features_to_Analyze": ["CD25", "CD3D", "CD45", "CD4", "CD8A"],
@@ -63,15 +63,20 @@ class TestSetupAnalysisTemplate(unittest.TestCase):
             "X_Coordinate_Column": "X_centroid",
             "Y_Coordinate_Column": "Y_centroid",
             "Annotation_s_": ["broad_cell_type", "detailed_cell_type"],
-            "Output_File": "analysis_output.pickle"
+            "Output_File": "analysis_output.pickle",
+            # Standardized outputs config - analysis as file
+            "outputs": {
+                "analysis": {"type": "file", "name": "analysis_output.pickle"}
+            }
         }
 
     def tearDown(self) -> None:
         self.tmp_dir.cleanup()
 
+    @patch('spac.templates.setup_analysis_template.save_results')
     @patch('spac.templates.setup_analysis_template.ingest_cells')
-    def test_run_with_save(self, mock_ingest) -> None:
-        """Test setup analysis with file saving."""
+    def test_run_with_save(self, mock_ingest, mock_save_results) -> None:
+        """Test setup analysis with saving to disk."""
         # Mock the ingest_cells function
         mock_adata = ad.AnnData(
             X=np.random.rand(10, 5),
@@ -80,7 +85,27 @@ class TestSetupAnalysisTemplate(unittest.TestCase):
         mock_adata.var_names = ['CD25', 'CD3D', 'CD45', 'CD4', 'CD8A']
         mock_ingest.return_value = mock_adata
 
+        # Mock save_results to return expected structure
+        mock_save_results.return_value = {
+            "analysis": str(Path(self.tmp_dir.name) / "analysis_output.pickle")
+        }
+
+        # Test with save_to_disk=True (default)
         saved_files = run_from_json(self.params)
+
+        # Check that save_results was called with correct params
+        mock_save_results.assert_called_once()
+        call_args = mock_save_results.call_args
+
+        # Verify params were passed correctly
+        self.assertIn('results', call_args[1])
+        self.assertIn('params', call_args[1])
+        self.assertEqual(call_args[1]['params'], self.params)
+
+        # Check that results contain analysis
+        results = call_args[1]['results']
+        self.assertIn("analysis", results)
+        self.assertIs(results["analysis"], mock_adata)        
 
         # Check that ingest_cells was called with correct parameters
         mock_ingest.assert_called_once()
@@ -104,8 +129,9 @@ class TestSetupAnalysisTemplate(unittest.TestCase):
         for feature in self.params["Features_to_Analyze"]:
             self.assertIn(f"^{feature}$", regex_list)
 
-        # Check that output file was created
-        self.assertIn("analysis_output.pickle", saved_files)
+        # Verify return structure
+        self.assertIsInstance(saved_files, dict)
+        self.assertIn("analysis", saved_files)
 
     @patch('spac.templates.setup_analysis_template.ingest_cells')
     def test_run_without_save(self, mock_ingest) -> None:
@@ -118,12 +144,95 @@ class TestSetupAnalysisTemplate(unittest.TestCase):
         mock_adata.var_names = ['CD25', 'CD3D', 'CD45', 'CD4', 'CD8A']
         mock_ingest.return_value = mock_adata
 
-        result = run_from_json(self.params, save_results=False)
+        # Test with save_to_disk=False
+        result = run_from_json(self.params, save_to_disk=False)
 
         # Check that we got an AnnData object back
[…]

[…]
