### 2025-11-11 e7bc509
feat: add setup_analysis.xml optimize boxplot.xml with help and default, update format_values
