### 2025-11-12 b5dc73b
feat: added galaxy_xml_synthesizer and updated format_values.py
