+++ galaxy_tools/galaxy_xml_synthesizer/galaxy_xml_synthesizer.py
@@ -1,0 +1,474 @@
+#!/usr/bin/env python3
+"""
+Generalized Galaxy XML Synthesizer
+Blueprint JSON drives ALL decisions - no hardcoding, no special cases
+"""
+
+import json
+import xml.etree.ElementTree as ET
+from xml.dom import minidom
+from pathlib import Path
+from typing import Dict, List, Any
+import re
+import argparse
+import sys
+
+
+class GalaxyXMLSynthesizer:
+    """
+    Synthesize Galaxy XML purely from blueprint JSON.
+    NO special cases, NO tool-specific logic.
+    """
+    
+    def __init__(self, blueprint: Dict[str, Any], docker_image: str = "spac:latest"):
+        self.blueprint = blueprint
+        self.docker_image = docker_image
+        
+    def synthesize(self) -> str:
+        """Generate Galaxy tool XML from blueprint."""
+        tool = ET.Element('tool')
+        
+        # Extract basic info from blueprint
+        tool_id = self._make_tool_id(self.blueprint.get('title', 'tool'))
+        tool.set('id', tool_id)
+        tool.set('name', self.blueprint.get('title', 'Tool'))
+        tool.set('version', '2.0.0')
+        tool.set('profile', '24.2')
+        
+        # Description
+        desc = ET.SubElement(tool, 'description')
+        desc.text = self._clean_text(self.blueprint.get('description', ''))[:200]
+        
+        # Requirements
+        reqs = ET.SubElement(tool, 'requirements')
+        container = ET.SubElement(reqs, 'container')
+        container.set('type', 'docker')
+        container.text = self.docker_image
+        
+        # Configfiles - Galaxy serializes inputs to JSON
+        configfiles = ET.SubElement(tool, 'configfiles')
+        inputs_config = ET.SubElement(configfiles, 'inputs')
+        inputs_config.set('name', 'params_json')
+        inputs_config.set('filename', 'galaxy_params.json')
+        inputs_config.set('data_style', 'paths')
+        
+        # Command
+        self._add_command(tool, tool_id)
+        
+        # Inputs
+        self._add_inputs(tool)
+        
+        # Outputs - directly from blueprint, no inference
+        self._add_outputs(tool)
+        
+        # Help
+        help_elem = ET.SubElement(tool, 'help')
+        help_elem.text = f"<![CDATA[\n{self._generate_help()}\n]]>"
+        
+        # Citations
+        citations = ET.SubElement(tool, 'citations')
+        citation = ET.SubElement(citations, 'citation')
+        citation.set('type', 'doi')
+        citation.text = '10.1038/s41586-019-1876-x'
+        
+        return self._format_xml(tool)
+    
+    def _make_tool_id(self, title: str) -> str:
+        """Generate tool ID from title."""
+        # Remove brackets and clean
+        clean_title = re.sub(r'\[.*?\]', '', title).strip()
+        tool_id = clean_title.lower().replace(' ', '_')
+        return f"spac_{tool_id}"
+    
+    def _clean_text(self, text: str) -> str:
+        """Clean text from markdown and escapes."""
+        text = re.sub(r'\[([^\]]+)\]\([^)]+\)', r'\1', text)
+        text = text.replace('\\n', ' ')
+        return text.strip()
+    
+    def _add_command(self, tool: ET.Element, tool_id: str):
+        """Add command section."""
+        # Collect parameter types for format_values.py
+        bool_params = []
+        list_params = []
+        
+        for param in self.blueprint.get('parameters', []):
+            param_type = param.get('paramType')
+            param_key = param.get('key')
+            
+            if param_type == 'BOOLEAN':
+                bool_params.append(param_key)
+            elif param_type == 'LIST':
+                list_params.append(param_key)
+        
[…]
