+++ galaxy_tools/galaxy_xml_synthesizer/format_values.py
@@ -1,0 +1,303 @@
+#!/usr/bin/env python3
+"""
+format_values.py - Complete Galaxy Parameter Processor for SPAC
+Standalone file with all functionality integrated
+"""
+
+import json
+import argparse
+import sys
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+
+
+def normalize_boolean(value: Any) -> bool:
+    """
+    Convert Galaxy boolean strings to Python boolean.
+    
+    Parameters
+    ----------
+    value : Any
+        Value to convert (typically "True"/"False" strings from Galaxy)
+        
+    Returns
+    -------
+    bool
+        Python boolean value
+    """
+    if isinstance(value, bool):
+        return value
+    
+    if isinstance(value, str):
+        value_lower = value.lower().strip()
+        if value_lower in ('true', 'yes', '1', 't'):
+            return True
+        elif value_lower in ('false', 'no', '0', 'f', 'none', ''):
+            return False
+    
+    # Default to False for unexpected values
+    return bool(value) if value else False
+
+
+def extract_list_from_repeat(params: Dict[str, Any], param_name: str) -> List[str]:
+    """
+    Extract list values from Galaxy repeat structure or text area.
+    
+    Galaxy generates repeat parameters with '_repeat' suffix:
+        "Feature_s_to_Plot_repeat": [{"value": "CD3"}, {"value": "CD4"}]
+    
+    Text areas generate newline-separated strings:
+        "Features_to_Analyze": "CD3\\nCD4\\nCD8"
+    
+    Parameters
+    ----------
+    params : dict
+        Full Galaxy parameters dictionary
+    param_name : str
+        Base parameter name (without _repeat suffix)
+        
+    Returns
+    -------
+    list
+        Simple list of string values, or empty list if no values found
+    """
+    # Check for repeat structure first
+    repeat_key = f"{param_name}_repeat"
+    if repeat_key in params:
+        repeat_value = params[repeat_key]
+        if isinstance(repeat_value, list):
+            result = []
+            for item in repeat_value:
+                if isinstance(item, dict) and 'value' in item:
+                    val = str(item['value']).strip()
+                    if val:  # Skip empty values
+                        result.append(val)
+            return result
+    
+    # Check for regular parameter (could be text area or single value)
+    if param_name in params:
+        value = params[param_name]
+        
+        # Handle newline-separated text area input
+        if isinstance(value, str):
+            if '\n' in value:
+                # Text area with multiple lines
+                return [line.strip() for line in value.split('\n') if line.strip()]
+            elif ',' in value:
+                # Comma-separated values
+                return [v.strip() for v in value.split(',') if v.strip()]
+            elif value.strip():
+                # Single value
+                return [value.strip()]
+        elif isinstance(value, list):
+            # Already a list
+            return [str(v).strip() for v in value if v and str(v).strip()]
+    
+    # No parameter found, return empty list
+    return []
+
+
+def inject_output_configuration(
+    cleaned: Dict[str, Any], 
+    outputs_config: Optional[Dict[str, Any]] = None
+) -> None:
+    """
+    Inject output configuration for template_utils.save_results().
+    
+    Parameters
+    ----------
+    cleaned : dict
+        Cleaned parameters dictionary (modified in-place)
+    outputs_config : dict, optional
+        Output configuration to inject. If None, no outputs are configured.
+        Format: {"output_name": {"type": "file|directory", "name": "filename"}}
+    """
+    if outputs_config:
+        cleaned['outputs'] = outputs_config
+    
+    # Enable result saving if outputs are configured
+    if cleaned.get('outputs'):
[…]

[…]
