### 2025-11-14 52be33e
feat: Add Docker development setup with live code mounting
- Add Makefile with production and development commands
- Add docker/Dockerfile.dev for development with volume mounting
- Add docker/docker-compose.dev.yml for easy dev environment setup
- Add docker/README.md with comprehensive Docker documentation
- Update main README.md with Docker quick start guide
- Production Dockerfile remains at root following convention
- Development setup allows instant code changes without rebuild

[ci skip]
