### 2025-11-17 #8787 @kevinyamauchi
Thank you for the updates, @fangliu117 ! I apologize for the delay - I was out on holiday.

I have completed my checklist. @fangliu117 's updates have checked all of the boxes. I have a couple of minor comments below, but in my opinion, they shouldn't block acceptance and don't warrant another round of review. Thank you all for the constructive review process!

- When I installed SPAC using the instructions in the repo, `pip` raised some depedency conflicts. The installation still completed and the notebook ran, so I guess they aren't critical. I made an issue to track here: https://github.com/FNLCR-DMAP/SCSAWorkflow/issues/411
- Giotto Suite has a (recently published) peer-reviewed paper. I think it would be better to cite that one: https://www.nature.com/articles/s41592-025-02817-w
- I noticed that the benchmarking and updates from the review process are on separate branches. I think it would be nice if those were both merged into main upon acceptance
