### 2025-11-19 #8787 @fangliu117
Thank you @kevinyamauchi again for completing checklist and for the helpful comments! We can address all three points you raised and will also update the manuscript to include the most recent, peer‑reviewed reference you suggested.
- pip dependency warnings: Thanks for opening FNLCR-DMAP/SCSAWorkflow#411. We’ll reproduce the conflicts, tighten/adjust our version pins, and update the installation docs so the pip path is conflict‑free (alongside the conda route). We’ll also expand CI to test both conda and pip installs across supported Python versions.
- Giotto Suite citation: We’ll update the paper to cite the newly published Nature Methods article on Giotto Suite and revise the “State of the Field” text accordingly.
- Branch hygiene: it’s cleaner if the benchmarking and review‑update branches are merged into main upon acceptance; we can also prepare PRs earlier if that’s preferred.

Hi @adamltyson, for process guidance: would you like us to push these non‑blocking fixes and the citation update now, or wait until both reviewers have finished before consolidating changes and merging branches into main?

Thank you all again for the constructive review process!
