### 2025-11-20 2ba8ec4
fix(deps): resolve scimap 2.1.3 dependency conflicts
Pin package versions in environment.yml to match scimap requirements:
- scipy=1.12.0 (was unpinned, scimap requires <=1.12.0)
- dask=2023.12.1 (was unpinned, scimap requires <2024.0.0)
- tifffile=2023.9.26 (was unpinned, scimap requires <2024.0.0)
- mpl-scatter-density=0.7 (was unpinned, scimap requires <0.8)
- pytest==7.4.4 (scimap requires <8.0.0)
- combat==0.3.3 (was missing, required by scimap)

Enhanced Dockerfile with conda environment activation for Galaxy tool compatibility.
Updated Docker development setup with live code mounting su
[…]
