### 2025-11-20 ad64d29
feat(env): add jupyter to conda environment
Add jupyter to environment.yml to support local notebook execution.
This ensures users can run Jupyter notebooks when using conda
environments locally, not just in Docker.

- Added jupyter dependency to environment.yml
- Enables local notebook execution with test_notebook_execution.sh
- Improves developer experience for local development workflows
