### 2025-11-21 3ccc650
refactor(template): refactor save resutls for 22 templates.
