+++ src/spac/templates/visualize_ripley_l_template.py
@@ -13,6 +13,7 @@ from pathlib import Path
 from typing import Any, Dict, Union, List, Optional, Tuple
 import pandas as pd
 import matplotlib.pyplot as plt
+import logging
 
 # Add parent directory to path for imports
 sys.path.append(str(Path(__file__).parent.parent.parent))
@@ -20,7 +21,7 @@ sys.path.append(str(Path(__file__).parent.parent.parent))
 from spac.visualization import plot_ripley_l
 from spac.templates.template_utils import (
     load_input,
-    save_outputs,
+    save_results,
     parse_params,
     text_to_value,
 )
@@ -28,8 +29,9 @@ from spac.templates.template_utils import (
 
 def run_from_json(
     json_path: Union[str, Path, Dict[str, Any]],
-    save_results: bool = True,
-    show_plot: bool = True
+    save_to_disk: bool = True,
+    show_plot: bool = True,
+    output_dir: Optional[Union[str, Path]] = None
 ) -> Union[Dict[str, str], Tuple[Any, pd.DataFrame]]:
     """
     Execute Visualize Ripley L analysis with parameters from JSON.
@@ -39,17 +41,19 @@ def run_from_json(
     ----------
     json_path : str, Path, or dict
         Path to JSON file, JSON string, or parameter dictionary
-    save_results : bool, optional
+    save_to_disk : bool, optional
         Whether to save results to file. If False, returns the figure and
         dataframe directly for in-memory workflows. Default is True.
     show_plot : bool, optional
         Whether to display the plot. Default is True.
+    output_dir : str or Path, optional
+        Directory for outputs. If None, uses current directory.
 
     Returns
     -------
     dict or tuple
-        If save_results=True: Dictionary of saved file paths
-        If save_results=False: Tuple of (figure, dataframe)
+        If save_to_disk=True: Dictionary of saved file paths
+        If save_to_disk=False: Tuple of (figure, dataframe)
     """
     # Parse parameters from JSON
     params = parse_params(json_path)
@@ -64,7 +68,7 @@ def run_from_json(
     regions_labels = params.get("Regions_Labels", [])
     plot_simulations = params.get("Plot_Simulations", True)
 
-    print(f"running with center_phenotype: {center_phenotype}, neighbor_phenotype: {neighbor_phenotype}")
+    logging.info(f"Running with center_phenotype: {center_phenotype}, neighbor_phenotype: {neighbor_phenotype}")
 
     # Process regions parameter exactly as in NIDAP template
     if plot_specific_regions:
@@ -89,29 +93,59 @@ def run_from_json(
         plt.show()
 
     # Print the dataframe to console
-    print(plots_df.to_string())
+    logging.info(f"\n{plots_df.to_string()}")
 
-    # Handle results based on save_results flag
-    if save_results:
-        # Save outputs
-        output_file = params.get("Output_File", "plots.csv")
-        saved_files = save_outputs({output_file: plots_df})
+    # Handle results based on save_to_disk flag
+    if save_to_disk:
+        # Prepare results dictionary based on outputs config
+        results_dict = {}
+        
+        # Check for dataframe output in config
+        if "dataframe" in params["outputs"]:
+            results_dict["dataframe"] = plots_df
+        
+        # Add figure if configured (usually not in the original template)
+        # but we can add it as an enhancement
+        if "figures" in params.get("outputs", {}):
+            # Package figure in a dictionary for directory saving
+            results_dict["figures"] = {"ripley_l_plot": fig}
+        
+        # Add analysis output if in config (for compatibility)
+        if "analysis" in params.get("outputs", {}):
+            results_dict["analysis"] = adata
+        
+        # Use centralized save_results function
[…]
