+++ src/spac/templates/interactive_spatial_plot_template.py
@@ -2,6 +2,9 @@
 Platform-agnostic Interactive Spatial Plot template converted from NIDAP.
 Maintains the exact logic from the NIDAP template.
 
+Refactored to use centralized save_results from template_utils.
+Follows standardized output schema where HTML files are saved as a directory.
+
 Usage
 -----
 >>> from spac.templates.interactive_spatial_plot_template import run_from_json
@@ -21,7 +24,7 @@ sys.path.append(str(Path(__file__).parent.parent.parent))
 from spac.visualization import interactive_spatial_plot
 from spac.templates.template_utils import (
     load_input,
-    save_outputs,
+    save_results,
     parse_params,
     text_to_value,
 )
@@ -29,8 +32,9 @@ from spac.templates.template_utils import (
 
 def run_from_json(
     json_path: Union[str, Path, Dict[str, Any]],
-    save_results: bool = True
-) -> Optional[Dict[str, str]]:
+    save_to_disk: bool = True,
+    output_dir: str = None,
+) -> Union[Dict[str, Union[str, List[str]]], None]:
     """
     Execute Interactive Spatial Plot analysis with parameters from JSON.
     Replicates the NIDAP template functionality exactly.
@@ -38,22 +42,67 @@ def run_from_json(
     Parameters
     ----------
     json_path : str, Path, or dict
-        Path to JSON file, JSON string, or parameter dictionary
-    save_results : bool, optional
-        Whether to save results to file. Default is True.
+        Path to JSON file, JSON string, or parameter dictionary.
+        Expected JSON structure:
+        {
+            "Upstream_Analysis": "path/to/data.pickle",
+            "Color_By": "Annotation",
+            "Annotation_s_to_Highlight": ["renamed_phenotypes"],
+            "outputs": {
+                "html": {"type": "directory", "name": "html_dir"}
+            }
+        }
+    save_to_disk : bool, optional
+        Whether to save results to disk. If False, returns None as plots are 
+        shown interactively. Default is True.
+    output_dir : str, optional
+        Base directory for outputs. If None, uses params['Output_Directory']
+        or current directory. All outputs will be saved relative to this directory.
 
     Returns
     -------
     dict or None
-        If save_results=True: Dictionary of saved file paths
-        If save_results=False: None (plots are shown interactively)
+        If save_to_disk=True: Dictionary of saved file paths with structure:
+            {"html": ["path/to/html_dir/plot1.html", ...]}
+        If save_to_disk=False: None (plots are shown interactively)
+
+    Notes
+    -----
+    Output Structure:
+    - HTML files are saved in a directory (standardized for HTML outputs)
+    - When save_to_disk=False, plots are shown interactively
+    
+    Examples
+    --------
+    >>> # Save results to disk
+    >>> saved_files = run_from_json("params.json")
+    >>> print(saved_files["html"])  # List of HTML file paths
+    >>> # ['./html_dir/plot_1.html', './html_dir/plot_2.html']
+    
+    >>> # Display plots interactively without saving
+    >>> run_from_json("params.json", save_to_disk=False)
+    
+    >>> # Custom output directory
+    >>> saved = run_from_json("params.json", output_dir="/custom/path")
     """
     # Parse parameters from JSON
     params = parse_params(json_path)
 
+    # Set output directory
+    if output_dir is None:
+        output_dir = params.get("Output_Directory", ".")
+    
+    # Ensure outputs configuration exists with standardized defaults
+    # HTML uses directory type per standardized schema
+    if "outputs" not in params:
+        params["outputs"] = {
+            "html": {"type": "directory", "name": "html_dir"}
+        }
+
     # Load the upstream analysis data
     adata = load_input(params["Upstream_Analysis"])
 
+    # Extract parameters
     color_by = params["Color_By"]
[…]
