### 2025-11-21 d5be7c3
feat: add 21 xmls, update format_values, galaxy_xml_synthesizer, blueprint_json
