+++ galaxy_tools/MVP_tools/format_values.py
@@ -2,24 +2,33 @@
 """
 format_values.py - Utility to normalize Galaxy parameters JSON for template consumption.
 
-Version 2.0 - No Cheetah needed! Processes Galaxy repeat structures directly.
+Version 3.0 - Clean version per supervisor guidance
+              NO parameter-specific hardcoding
+              Generic delimiter support via CLI flags
 
 Handles three main transformations:
 1. Converts Galaxy repeat structures to simple lists
 2. Converts boolean string values to actual Python booleans
-3. Injects output configuration for template_utils (single files for boxplot MVP)
+3. Supports delimited text fields (e.g., text areas with semicolons or newlines)
+4. Injects output configuration for template_utils
 
 Usage:
+    # For repeat structures:
     python format_values.py galaxy_params.json cleaned_params.json \
-        --bool-values Horizontal_Plot Keep_Outliers Value_Axis_Log_Scale \
+        --bool-values Horizontal_Plot Keep_Outliers \
         --list-values Feature_s_to_Plot
+    
+    # For text area with newline separators:
+    python format_values.py galaxy_params.json cleaned_params.json \
+        --list-sep '\\n' --list-fields Anchor_Neighbor_List
 
-This processes Galaxy's raw parameter output directly without needing Cheetah preprocessing.
+This version is template-agnostic with no parameter name hardcoding.
 """
 
 import json
 import argparse
 import sys
+import re
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -99,6 +108,37 @@ def extract_list_from_repeat(params: Dict[str, Any], param_name: str) -> List[st
     return []
 
 
+def parse_delimited_text(text: str, separator: str = ';') -> List[str]:
+    """
+    Parse a delimited text string into a list of values.
+    
+    Parameters
+    ----------
+    text : str
+        Delimited text string (may contain newlines for multi-line input)
+    separator : str
+        Separator character (default: semicolon)
+        
+    Returns
+    -------
+    list
+        List of trimmed, non-empty values
+    """
+    if not text:
+        return []
+    
+    # Handle newline separator for text areas
+    if separator == '\n':
+        # Split by newlines
+        lines = text.strip().split('\n')
+        # Clean up each line
+        values = [line.strip() for line in lines if line.strip()]
+    else:
+        # Split by specified separator
+        values = [s.strip() for s in text.split(separator) if s.strip()]
+    
+    return values
+
 def inject_output_configuration(cleaned: Dict[str, Any], outputs_config: Optional[Dict[str, Any]] = None) -> None:
     """
     Inject output configuration for template_utils.save_results().
@@ -112,7 +152,7 @@ def inject_output_configuration(cleaned: Dict[str, Any], outputs_config: Optiona
         Cleaned parameters dictionary (modified in-place)
     outputs_config : dict, optional
         Output configuration to inject. If None, no outputs are configured.
-        Format: {"output_name": {"type": "file|directory", "name": "filename"}}
+        Format: {"output_name": {"type": "file", "name": "filename"}}
     """
     if outputs_config:
         cleaned['outputs'] = outputs_config
@@ -126,6 +166,7 @@ def process_galaxy_params(
     params: Dict[str, Any],
     bool_params: List[str],
     list_params: List[str],
+    delimited_params: Dict[str, str] = None,
     outputs_config: Optional[Dict[str, Any]] = None,
     inject_outputs: bool = False
 ) -> Dict[str, Any]:
@@ -140,6 +181,9 @@ def process_galaxy_params(
         List of parameter names that should be booleans
     list_params : list
         List of parameter names that should be extracted from repeat structures
+    delimited_params : dict, optional
+        Parameters with delimited text {param_name: separator}
+        For text areas or fields that use delimiters instead of repeat structures
[…]
