### 2025-11-24 08bef3b
refactor: synthesizer with ordering and grouping param, and output xmls accordingly
