### 2025-11-25 8145438
feat: add README to MVP_tools directory
