+++ galaxy_tools/MVP_tools/README.md
@@ -1,0 +1,472 @@
+# MVP_tools: Blueprint to Galaxy Tool Pipeline
+
+This guide walks you through converting NIDAP blueprint JSON files into fully functional Galaxy tools. It covers:
+
+1. **What changes are needed** in the blueprint to make it Galaxy-ready
+2. **How to run the synthesizer** to generate Galaxy XML
+3. **What `format_values.py` does** to normalize parameters at runtime
+
+This documentation is intended for developers who want to generate Galaxy tools from blueprints, including adapting this pipeline for other projects (e.g., CCBR).
+
+---
+
+## Table of Contents
+
+- [Overview: The Pipeline](#overview-the-pipeline)
+- [Step 1: Preparing Your Blueprint](#step-1-preparing-your-blueprint)
+  - [The Critical `outputs` Section](#the-critical-outputs-section)
+  - [Output Types Explained](#output-types-explained)
+  - [Complete Blueprint Example](#complete-blueprint-example)
+- [Step 2: Running the Synthesizer](#step-2-running-the-synthesizer)
+  - [Basic Usage](#basic-usage)
+  - [Batch Processing](#batch-processing)
+  - [Command Line Options](#command-line-options)
+- [Step 3: Understanding format_values.py](#step-3-understanding-format_valuespy)
+  - [Why format_values.py Is Needed](#why-format_valuespy-is-needed)
+  - [Raw vs Cleaned JSON](#raw-vs-cleaned-json)
+  - [What format_values.py Expects](#what-format_valuespy-expects)
+  - [What format_values.py Outputs](#what-format_valuespy-outputs)
+  - [Adapting for Your Project](#adapting-for-your-project)
+- [How It All Fits Together](#how-it-all-fits-together)
+- [Directory Structure](#directory-structure)
+- [Parameter Type Mapping](#parameter-type-mapping)
+- [Available Tools](#available-tools)
+- [Troubleshooting](#troubleshooting)
+
+---
+
+## Overview: The Pipeline
+
+The pipeline converts a NIDAP blueprint into a Galaxy tool through these steps:
+
+```
+┌─────────────────┐     ┌──────────────────────┐     ┌─────────────────┐
+│  Blueprint JSON │ ──▶ │ galaxy_xml_synthesizer│ ──▶│  Galaxy XML     │
+│  (with outputs) │     │        .py            │    │  (tool def)     │
+└─────────────────┘     └──────────────────────┘     └─────────────────┘
+                                                              │
+                                                              ▼
+                        ┌──────────────────────────────────────────────────┐
+                        │              At Runtime in Galaxy:               │
+                        │                                                  │
+                        │  galaxy_params.json ──▶ format_values.py ──▶     │
+                        │  cleaned_params.json ──▶ template.py ──▶ outputs │
+                        └──────────────────────────────────────────────────┘
+```
+
+**Key components:**
+
+| Component | Purpose |
+|-----------|---------|
+| `blueprint JSON` | Defines tool parameters, inputs, and **outputs** |
+| `galaxy_xml_synthesizer.py` | Converts blueprint → Galaxy XML |
+| `format_values.py` | Normalizes Galaxy's raw JSON → clean JSON for templates |
+| `*_template.py` | Python script that performs the actual analysis |
+
+---
+
+## Step 1: Preparing Your Blueprint
+
+The main change required to make a NIDAP blueprint Galaxy-ready is **adding an `outputs` section**. This tells the synthesizer what files/directories the tool will produce.
+
[…]
