### 2025-11-30 2232474
Merge reviewer comments: paper examples, paper.bib, paper.md, test scripts, Dockerfile, Makefile, README, environment.yml, docker setup
