### 2025-12-05 8cb77f3
refactor: updated the rest 16 templates and all 38 templates are the latest except posit-it-python
