+++ src/spac/templates/analysis_to_csv_template.py
@@ -2,6 +2,9 @@
 Platform-agnostic Analysis to CSV template converted from NIDAP.
 Maintains the exact logic from the NIDAP template.
 
+Refactored to use centralized save_results from template_utils.
+Reads outputs configuration from blueprint JSON file.
+
 Usage
 -----
 >>> from spac.templates.analysis_to_csv_template import run_from_json
@@ -9,6 +12,7 @@ Usage
 """
 import json
 import sys
+import logging
 from pathlib import Path
 from typing import Any, Dict, Union
 import pandas as pd
@@ -19,15 +23,18 @@ sys.path.append(str(Path(__file__).parent.parent.parent))
 from spac.utils import check_table
 from spac.templates.template_utils import (
     load_input,
-    save_outputs,
+    save_results,
     parse_params,
     text_to_value,
 )
 
+logger = logging.getLogger(__name__)
+
 
 def run_from_json(
     json_path: Union[str, Path, Dict[str, Any]],
-    save_results: bool = True
+    save_to_disk: bool = True,
+    output_dir: str = None,
 ) -> Union[Dict[str, str], pd.DataFrame]:
     """
     Execute Analysis to CSV analysis with parameters from JSON.
@@ -36,33 +43,59 @@ def run_from_json(
     Parameters
     ----------
     json_path : str, Path, or dict
-        Path to JSON file, JSON string, or parameter dictionary
-    save_results : bool, optional
-        Whether to save results to file. If False, returns the dataframe
+        Path to JSON file, JSON string, or parameter dictionary.
+        Expected JSON structure:
+        {
+            "Upstream_Analysis": "path/to/data.pickle",
+            "Table_to_Export": "Original",
+            "Save_as_CSV_File": false,
+            "outputs": {
+                "dataframe": {"type": "file", "name": "dataframe.csv"}
+            }
+        }
+    save_to_disk : bool, optional
+        Whether to save results to disk. If False, returns the dataframe
         directly for in-memory workflows. Default is True.
+    output_dir : str, optional
+        Base directory for outputs. If None, uses params['Output_Directory']
+        or current directory.
 
     Returns
     -------
     dict or DataFrame
-        If save_results=True: Dictionary of saved file paths
-        If save_results=False: The processed DataFrame
+        If save_to_disk=True: Dictionary of saved file paths with structure:
+            {"dataframe": "path/to/dataframe.csv"}
+        If save_to_disk=False: The processed DataFrame
+
+    Notes
+    -----
+    Output Structure:
+    - DataFrame is saved as a CSV file when save_to_disk is True 
+    - Otherwise, the DataFrame is returned for programmatic use
     """
     # Parse parameters from JSON
     params = parse_params(json_path)
+
+    # Set output directory
+    if output_dir is None:
+        output_dir = params.get("Output_Directory", ".")
+
+    # Ensure outputs configuration exists with standardized defaults
+    if "outputs" not in params:
+        params["outputs"] = {
+            "dataframe": {"type": "file", "name": "dataframe.csv"}
+        }
 
     # Load the upstream analysis data
     adata = load_input(params["Upstream_Analysis"])
 
     # Extract parameters
     input_layer = params.get("Table_to_Export", "Original")
-    save_file = params.get("Save_as_CSV_File", False)
 
     if input_layer == "Original":
         input_layer = None
 
-    def export_layer_to_csv(
-        adata,
-        layer=None):
+    def export_layer_to_csv(adata, layer=None):
         """
         Exports the specified layer or the default .X data matrix of an 
         AnnData object to a CSV file.
@@ -86,7 +119,6 @@ def run_from_json(
         full_data_df = data_to_export.join(adata.obs)
 
         # Join the spatial coordinates
-
[…]
