+++ src/spac/templates/spatial_plot_template.py
@@ -2,6 +2,9 @@
 Platform-agnostic Spatial Plot template converted from NIDAP.
 Maintains the exact logic from the NIDAP template.
 
+Refactored to use centralized save_results from template_utils.
+Reads outputs configuration from blueprint JSON file.
+
 Usage
 -----
 >>> from spac.templates.spatial_plot_template import run_from_json
@@ -13,6 +16,7 @@ from pathlib import Path
 from typing import Any, Dict, Union, List, Optional
 import matplotlib.pyplot as plt
 from functools import partial
+import logging
 
 # Add parent directory to path for imports
 sys.path.append(str(Path(__file__).parent.parent.parent))
@@ -22,6 +26,7 @@ from spac.data_utils import select_values
 from spac.utils import check_annotation
 from spac.templates.template_utils import (
     load_input,
+    save_results,
     parse_params,
     text_to_value,
 )
@@ -29,9 +34,10 @@ from spac.templates.template_utils import (
 
 def run_from_json(
     json_path: Union[str, Path, Dict[str, Any]],
-    save_results: bool = True,
-    show_plots: bool = True
-) -> Union[Dict[str, str], None]:
+    save_to_disk: bool = True,
+    show_plots: bool = True,
+    output_dir: str = None,
+) -> Union[Dict[str, Union[str, List[str]]], List[plt.Figure]]:
     """
     Execute Spatial Plot analysis with parameters from JSON.
     Replicates the NIDAP template functionality exactly.
@@ -39,51 +45,81 @@ def run_from_json(
     Parameters
     ----------
     json_path : str, Path, or dict
-        Path to JSON file, JSON string, or parameter dictionary
-    save_results : bool, optional
-        Whether to save results to file. If False, returns None.
-        Default is True.
+        Path to JSON file, JSON string, or parameter dictionary.
+        Expected JSON structure:
+        {
+            "Upstream_Analysis": "path/to/data.pickle",
+            "Stratify": true,
+            "Stratify_By": ["slide_id"],
+            "Color_By": "Annotation",
+            ...
+            "outputs": {
+                "figures": {"type": "directory", "name": "figures_dir"}
+            }
+        }
+    save_to_disk : bool, optional
+        Whether to save results to disk. If False, returns the figures
+        directly for in-memory workflows. Default is True.
     show_plots : bool, optional
         Whether to display the plots. Default is True.
+    output_dir : str, optional
+        Base directory for outputs. If None, uses params['Output_Directory']
+        or current directory.
 
     Returns
     -------
-    dict or None
-        If save_results=True: Dictionary of saved file paths
-        If save_results=False: None
+    dict or list
+        If save_to_disk=True: Dictionary of saved file paths
+        If save_to_disk=False: List of matplotlib figures
     """
+    # Set up logging
+    logging.basicConfig(level=logging.INFO)
+    logger = logging.getLogger(__name__)
+
     # Parse parameters from JSON
     params = parse_params(json_path)
+
+    # Set output directory
+    if output_dir is None:
+        output_dir = params.get("Output_Directory", ".")
+
+    # Ensure outputs configuration exists with standardized defaults
+    if "outputs" not in params:
+        params["outputs"] = {
+            "figures": {"type": "directory", "name": "figures_dir"}
+        }
 
     # Load the upstream analysis data
     adata = load_input(params["Upstream_Analysis"])
 
     # Extract parameters exactly as in NIDAP template
-    annotation = params["Annotation_to_Highlight"]
-    feature = params["Feature_to_Highlight"]
-    layer = params["Table"]
+    annotation = params.get("Annotation_to_Highlight", "None")
+    feature = params.get("Feature_to_Highlight", "")
[…]

[…]
