### 2025-12-05 e5b817b
refactor: Update configs and fix template packaging for pip install
