### 2025-12-08 805f9fc
refactor: Update paper.bib and paper.md based on reviewer comments
