+++ paper/paper.bib
@@ -212,14 +212,15 @@
 }
 
 
-@article{Chen:2023,
-  title = {Giotto Suite: a multi-scale and technology-agnostic spatial multi-omics analysis ecosystem},
-  author = {Chen, Jiaji George and Ch{\'a}vez-Fuentes, Joselyn Cristina and O'Brien, Matthew and Xu, Junxiang and Ruiz, Edward and Wang, Wen and Amin, Iqra and Sheridan, Jeffrey and Shin, Sujung Crystal and Hasyagar, Sanjana Varada and Sarfraz, Irzam and Guckhool, Pratishtha and Sistig, Adriana and Jarzabek, Veronica and Yuan, Guo-Cheng and Dries, Ruben},
-  journal = {bioRxiv},
-  year = {2023},
-  doi = {10.1101/2023.11.26.568752},
-  url = {https://www.biorxiv.org/content/10.1101/2023.11.26.568752v1},
-  note = {Preprint}
+@article{Chen:2025,
+  author  = {Chen, Jiaji George and Ch{\'a}vez-Fuentes, Joselyn Cristina and O'Brien, Matthew and Xu, Junxiang and Ruiz, Edward and Wang, Wen and Amin, Iqra and Sheridan, Jeffrey and Shin, Sujung Crystal and Hasyagar, Sanjana Varada and Sarfraz, Irzam and Guckhool, Pratishtha and Sistig, Adriana and Jarzabek, Veronica and Yuan, Guo-Cheng and Dries, Ruben},
+  title   = {Giotto Suite: a multiscale and technology-agnostic spatial multiomics analysis ecosystem},
+  journal = {Nature Methods},
+  year    = {2025},
+  volume  = {22},
+  pages   = {2052--2064},
+  doi     = {10.1038/s41592-025-02817-w},
+  url     = {https://doi.org/10.1038/s41592-025-02817-w}
 }
 
 
