+++ paper/paper.md
@@ -48,7 +48,7 @@ Multiplexed immunofluorescence microscopy captures detailed measurements of spat
 
 Multiplexed protein imaging (e.g., CODEX [@Goltsev:2018], CyCIF [@Lin:2018], MxIF [@Gerdes:2013]) yields gigabyte-scale, single-cell datasets with millions of cells per study. Extracting biological signal from these data requires reproducible workflows that span preprocessing, phenotyping, spatial statistics, and interactive visualization.
 
-The spatial omics ecosystem includes general frameworks for single-cell data and spatial analysis (AnnData [@Virshup:2024], Scanpy [@Wolf:2018], Squidpy [@Palla:2022]) and toolkits focused on multiplex imaging and spatial proteomics (SCIMAP [@Nirmal:2024]), alongside methods aimed primarily at spatial transcriptomics (Seurat [@Hao:2021], Giotto Suite [@Chen:2023, [GiottoSuite](https://giottosuite.com/)], GraphST [@Long:2023], Bento [@Mah:2024]). These packages collectively provide data structures, clustering, dimensionality reduction, neighborhood graphs, and spatial interaction metrics, and they are widely adopted for programmatic analyses.
+The spatial omics ecosystem includes general frameworks for single-cell data and spatial analysis (AnnData [@Virshup:2024], Scanpy [@Wolf:2018], Squidpy [@Palla:2022]) and toolkits focused on multiplex imaging and spatial proteomics (SCIMAP [@Nirmal:2024]), alongside methods aimed primarily at spatial transcriptomics (Seurat [@Hao:2021], Giotto Suite [@Chen:2025], GraphST [@Long:2023], Bento [@Mah:2024]). These packages collectively provide data structures, clustering, dimensionality reduction, neighborhood graphs, and spatial interaction metrics, and they are widely adopted for programmatic analyses.
 
 Despite this progress, biologists without coding expertise face steep barriers to conduct end-to-end analyses independently and iteratively. Projects often require stitching together multiple libraries and ad hoc scripts, which slows hypothesis testing, complicates reproducibility, and makes figure generation brittle and hard to reuse at scale.
 
