### 2025-12-08 9ae3ef3
refactor: Merge paper.bib and paper.md updates from address-reviewer-comments branch
