### 2025-12-08 25b27dc
refactor: Dockerfile.v2 to build nciccbr/spac:v2-dev docker image from refactor_template_outputs branch
