### 2025-12-09 6a65dad
refactor: fix umap-learn/scikit-learn compatibility in Dockerfile.v2 and environment.yml
Issue:
TypeError: check_array() got an unexpected keyword argument 'ensure_all_finite'
- UMAP calls scikit-learn's check_array() with 'ensure_all_finite' parameter
- scikit-learn 1.5+ deprecated this parameter in favor of 'force_all_finite'
- The conda-installed umap-learn version was incompatible with scikit-learn 1.5.2

Solution 1 (Dockerfile workaround - used in v1, not chosen):
- Uninstall conda umap-learn and reinstall via pip: umap-learn==0.5.9.post2
- As implemented in Dockerfile.v1
- Hacky workarou
[…]
