### 2025-12-09 2b4c119
refactor: updated paper_edit.md according to the gpt grammar check for JOSS proof
