### 2025-12-09 74efbb2
refactor: synthesizer with new docker image nciccbr/spac:v2-dev, blueprints, xmls in galaxy_tools, as well as optimization of template functions on refactor_template_outputs branch
