+++ galaxy_tools/MVP_tools/boxplot_template.py
@@ -1,269 +1,0 @@
-"""
-Platform-agnostic Boxplot template converted from NIDAP.
-Maintains the exact logic from the NIDAP template.
-
-Refactored to use centralized save_results from template_utils.
-Follows standardized output schema where figures are saved as directories.
-
-Usage
------
->>> from spac.templates.boxplot_template import run_from_json
->>> run_from_json("examples/boxplot_params.json")
-"""
-import json
-import sys
-from pathlib import Path
-from typing import Any, Dict, Union, List, Optional, Tuple
-import logging
-import pandas as pd
-import matplotlib.pyplot as plt
-import seaborn as sns
-
-# Add parent directory to path for imports
-sys.path.append(str(Path(__file__).parent.parent.parent))
-
-from spac.visualization import boxplot
-from spac.templates.template_utils import (
-    load_input,
-    save_results,
-    parse_params,
-    text_to_value,
-)
-
-
-def run_from_json(
-    json_path: Union[str, Path, Dict[str, Any]],
-    save_to_disk: bool = True,
-    show_plot: bool = True,
-    output_dir: str = None,
-) -> Union[Dict[str, Union[str, List[str]]], Tuple[Any, pd.DataFrame]]:
-    """
-    Execute Boxplot analysis with parameters from JSON.
-    Replicates the NIDAP template functionality exactly.
-
-    Parameters
-    ----------
-    json_path : str, Path, or dict
-        Path to JSON file, JSON string, or parameter dictionary.
-        Expected JSON structure:
-        {
-            "Upstream_Analysis": "path/to/data.pickle",
-            "Primary_Annotation": "cell_type",
-            "Feature_s_to_Plot": ["CD4", "CD8"],
-            "outputs": {
-                "figures": {"type": "directory", "name": "figures"},
-                "dataframe": {"type": "file", "name": "output.csv"}
-            }
-        }
-    save_to_disk : bool, optional
-        Whether to save results to disk. If True, saves figures to a directory
-        and summary statistics to a CSV file. If False, returns the figure and
-        summary dataframe directly for in-memory workflows. Default is True.
-    show_plot : bool, optional
-        Whether to display the plot. Default is True.
-    output_dir : str, optional
-        Base directory for outputs. If None, uses params['Output_Directory']
-        or current directory. All outputs will be saved relative to this directory.
-
-    Returns
-    -------
-    dict or tuple
-        If save_to_disk=True: Dictionary of saved file paths with structure:
-            {
-                "figures": ["path/to/figures/boxplot.png"],  # List of figure paths
-                "DataFrame": "path/to/output.csv"           # Single file path
-            }
-        If save_to_disk=False: Tuple of (matplotlib.figure.Figure, pd.DataFrame)
-            containing the figure object and summary statistics dataframe
-
-    Notes
-    -----
-    Output Structure:
-    - Figures are saved in a directory (standardized for all figure outputs)
-    - Summary statistics are saved as a single CSV file
-    - When save_to_disk=False, objects are returned for programmatic use
-    
-    Examples
-    --------
-    >>> # Save results to disk
-    >>> saved_files = run_from_json("params.json")
-    >>> print(saved_files["figure"])  # List of paths to saved plots
-    >>> # ['./figures/boxplot.png']
-    
-    >>> # Get results in memory
-    >>> fig, summary_df = run_from_json("params.json", save_to_disk=False)
-
-    >>> # Custom output directory
-    >>> saved = run_from_json("params.json", output_dir="/custom/path")
-    """
-    # Parse parameters from JSON
-    params = parse_params(json_path)
-
-    # Set output directory
-    if output_dir is None:
-        output_dir = params.get("Output_Directory", ".")
-    
[…]
