+++ galaxy_tools/refactor_tools/spac_galaxy_runner.py
@@ -1,515 +1,0 @@
-#!/usr/bin/env python3
-"""
-spac_galaxy_runner.py - Hybrid version combining refactored structure with robust parameter handling
-Incorporates critical fixes from original wrapper for parameter processing
-"""
-
-import json
-import os
-import sys
-import subprocess
-import shutil
-from pathlib import Path
-import re
-
-def main():
-    """Main entry point for SPAC Galaxy runner"""
-    if len(sys.argv) != 3:
-        print("Usage: spac_galaxy_runner.py <params.json> <template_name>")
-        sys.exit(1)
-    
-    params_path = sys.argv[1]
-    template_name = sys.argv[2]
-    
-    print(f"=== SPAC Galaxy Runner v2.0 (Hybrid) ===")
-    print(f"Template: {template_name}")
-    print(f"Parameters: {params_path}")
-    
-    # Load parameters
-    with open(params_path) as f:
-        params = json.load(f)
-    
-    # Extract outputs specification from environment variable
-    outputs_spec_env = os.environ.get('GALAXY_OUTPUTS_SPEC', '')
-    if outputs_spec_env:
-        try:
-            outputs = json.loads(outputs_spec_env)
-        except json.JSONDecodeError:
-            print(f"WARNING: Could not parse GALAXY_OUTPUTS_SPEC: {outputs_spec_env}")
-            outputs = determine_default_outputs(template_name)
-    else:
-        # Fallback: try to get from params
-        outputs = params.pop('outputs', {})
-        if isinstance(outputs, str):
-            try:
-                outputs = json.loads(unsanitize_galaxy_params(outputs))
-            except json.JSONDecodeError:
-                print(f"WARNING: Could not parse outputs: {outputs}")
-                outputs = determine_default_outputs(template_name)
-    
-    print(f"Outputs specification: {outputs}")
-    
-    # CRITICAL: Unsanitize and normalize parameters (from original)
-    params = process_galaxy_parameters(params, template_name)
-    
-    # Handle multiple file inputs that were copied to directories by Galaxy
-    handle_multiple_file_inputs(params)
-    
-    # Create output directories
-    create_output_directories(outputs)
-    
-    # Add output paths to params - critical for templates that save results
-    params['save_results'] = True
-    
-    if 'analysis' in outputs:
-        params['output_path'] = outputs['analysis']
-        params['Output_Path'] = outputs['analysis']
-        params['Output_File'] = outputs['analysis']
-    
-    if 'DataFrames' in outputs:
-        df_path = outputs['DataFrames']
-        # Check if it's a single file or a directory
-        if df_path.endswith('.csv') or df_path.endswith('.tsv'):
-            # Single file output (like Load CSV Files)
-            params['output_file'] = df_path
-            params['Output_File'] = df_path
-            print(f"  Set output_file to: {df_path}")
-        else:
-            # Directory for multiple files (like boxplot)
-            params['output_dir'] = df_path
-            params['Export_Dir'] = df_path
-            params['Output_File'] = os.path.join(df_path, f'{template_name}_output.csv')
-            print(f"  Set output_dir to: {df_path}")
-    
-    if 'figures' in outputs:
-        fig_dir = outputs['figures']
-        params['figure_dir'] = fig_dir
-        params['Figure_Dir'] = fig_dir
-        params['Figure_File'] = os.path.join(fig_dir, f'{template_name}.png')
-        print(f"  Set figure_dir to: {fig_dir}")
-    
-    if 'html' in outputs:
-        html_dir = outputs['html']
-        params['html_dir'] = html_dir
-        params['Output_File'] = os.path.join(html_dir, f'{template_name}.html')
-        print(f"  Set html_dir to: {html_dir}")
-    
-    # Save config for debugging (without outputs key)
-    with open('config_used.json', 'w') as f:
[…]

[…]
