### 2025-12-10 5287382
Update README with SPAC ecosystem details
Added repository information and installation instructions.
