### 2025-12-10 d704626
Enhance README with SPAC ecosystem table
Formatted the SPAC ecosystem section into a table for better readability.
