### 2025-12-12 a1e9c93
Update paper.md and replace figure.tif with figure.pptx
