### 2025-12-12 a2b4a9d
Update setup.py and environment.yml from refactor_template_outputs
