### 2025-12-12 #8787 @fangliu117
> Hi [@fangliu117](https://github.com/fangliu117) that all sounds good. I've updated the review issue so that everything points to `main` now. Let me know when you're finished with the edits, and I will take a look.
>
Hi @adamltyson , sorry for the delay. We have carefully proofread the manuscript, and here is a summary of the revisions:

1. **Grammar corrections** in [paper.md](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/main/paper/paper.md): Fixed grammatical errors throughout the manuscript, including sentence structure improvements, subject-verb agreement, proper article usage, serial comma consistency, and punctuation refinements.
2. **Updated Figure 1 (figure.pptx)**: Switched title case to sentence case for the small image captions (e.g., "Single Cell Data" → "Single cell data"). Note: When downloading the raw [figure.pptx](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/main/paper/figure.pptx) from the GitHub repository, the figure may appear transformed with misaligned numbers and lower resolution. We have also added [figure.pdf](https://github.com/FNLCR-DMAP/SCSAWorkflow/blob/main/paper/figure.pdf) as an alternative to ensure proper rendering.
5. **Figure sizing note**: For your awareness, in the original proof, the figure extends slightly beyond the page edge (~72pt). This appears to be a LaTeX float issue during PDF generation.

All changes have been merged to the main branch:
[https://github.com/FNLCR-DMAP/SCSAWorkflow/tree/main/paper](https://github.com/FNLCR-DMAP/SCSAWorkflow/tree/main/paper)

Thank you again for your time and effort in processing the SPAC manuscript.
