### 2025-12-16 a2f25ff
Fix capitalization of 'SPAC' in README.md
Corrected capitalization of 'SPAC' in the README.
