### 2025-12-17 310e0a3
Update paper.md (add ORCIDs, fix EDA text) and replace figure files with figure.tif
