### 2025-12-17 #8787 @fangliu117
> Hi [@fangliu117](https://github.com/fangliu117), I don't think figures as pptx are supported. Could you revert to the tif image as before?
>
> Also, you define EDA as "Exploratory Data Analysis". I don't think this should be capitalised, and the abbreviation isn't used again so doesn't need to be defined.

Hi @adamltyson , Thanks for the detailed comments.

- I have reverted the figure file back to the original **.tif** format and removed the **.pptx** version from the repository.
- In `paper.md` I changed “Exploratory Data Analysis (EDA)” to “exploratory data analysis” (lower case, without introducing the EDA abbreviation, since it is not used again).
