### 2025-12-17 6ce255a
Update README.md: templates in src/spac/templates, synthesizer at root, expand to 38 tools
