+++ galaxy_tools/README.md
@@ -50,7 +50,8 @@ The pipeline converts a NIDAP blueprint into a Galaxy tool through these steps:
                         │              At Runtime in Galaxy:               │
                         │                                                  │
                         │  galaxy_params.json ──▶ format_values.py ──▶     │
-                        │  cleaned_params.json ──▶ template.py ──▶ outputs │
+                        │  cleaned_params.json ──▶ spac.templates ──▶      │
+                        │  outputs                                         │
                         └──────────────────────────────────────────────────┘
 ```
 
@@ -61,7 +62,7 @@ The pipeline converts a NIDAP blueprint into a Galaxy tool through these steps:
 | `blueprint JSON` | Defines tool parameters, inputs, and **outputs** |
 | `galaxy_xml_synthesizer.py` | Converts blueprint → Galaxy XML |
 | `format_values.py` | Normalizes Galaxy's raw JSON → clean JSON for templates |
-| `*_template.py` | Python script that performs the actual analysis |
+| `spac.templates.*_template` | Python modules in `src/spac/templates/` that perform the actual analysis |
 
 ---
 
@@ -174,23 +175,23 @@ Once your blueprint has the `outputs` section, generate the Galaxy XML.
 ### Basic Usage
 
 ```bash
-# Navigate to MVP_tools directory
-cd galaxy_tools/MVP_tools
+# Navigate to galaxy_tools directory
+cd galaxy_tools
 
 # Generate XML from a single blueprint
-python galaxy_xml_synthesizer.py blueprint_json/template_json_boxplot.json -o .
+python galaxy_xml_synthesizer.py blueprint_jsons/template_json_boxplot.json -o MVP_tools
 ```
 
-This creates `boxplot.xml` in the current directory.
+This creates `boxplot.xml` in the MVP_tools directory.
 
 ### Batch Processing
 
 ```bash
 # Process all blueprints in a directory
-python galaxy_xml_synthesizer.py blueprint_json/ -o .
+python galaxy_xml_synthesizer.py blueprint_jsons/ -o MVP_tools
 
 # Or use a glob pattern
-python galaxy_xml_synthesizer.py "blueprint_json/template_json_*.json" -o .
+python galaxy_xml_synthesizer.py "blueprint_jsons/template_json_*.json" -o MVP_tools
 ```
 
 ### Command Line Options
@@ -203,7 +204,7 @@ positional arguments:
 
 optional arguments:
   -h, --help            Show help message
-  -o, --output OUTPUT   Output directory for XML files (default: galaxy_tools)
+  -o, --output OUTPUT   Output directory for XML files (default: MVP_tools)
   --docker DOCKER       Docker image name (default: spac:mvp)
   --debug               Enable debug output with feature summary
 ```
@@ -227,7 +228,7 @@ python '$__tool_directory__/format_values.py' 'galaxy_params.json' 'cleaned_para
     --bool-values Horizontal_Plot Keep_Outliers Value_Axis_Log_Scale \
     --list-values Feature_s_to_Plot \
     --inject-outputs --outputs-config outputs_config.json && \
-python '$__tool_directory__/boxplot_template.py' 'cleaned_params.json'
+python -c "from spac.templates.boxplot_template import run_from_json; run_from_json('cleaned_params.json')"
 
 ]]></command>
 ```
@@ -350,7 +351,8 @@ Here's the complete flow from blueprint to Galaxy execution:
    │   ├── Fixes:  booleans, lists, injects outputs
    │   └── Output: cleaned_params.json
    │
-   └── template.py runs:
+   └── spac.templates.*_template.run_from_json() runs:
+       ├── Imported from: src/spac/templates/
        ├── Input:  cleaned_params.json
        ├── Executes analysis logic
[…]
