### 2025-12-18 09303c5
Update figure to PNG format for better compatibility
