### 2025-12-19 #8787 @fangliu117
Hi @adamltyson , thanks for triggering the new proof.

We have carefully proofread the latest generated PDF and confirmed that:
- the updated figure renders correctly, and
- all text, references, author information are correct.

From our side, the manuscript and software are ready for the next step.
Please let us know if there is anything else you’d like us to adjust.
