### 2026-01-06 #8787 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
