### 2026-01-06 #8787 @kevinyamauchi
> this paper is ready to be accepted and published.

Exciting! Congrats @fangliu117  et al. 🚀
