### 2026-01-06 #8787 @fangliu117
Thanks everyone, we’re thrilled that the paper has been accepted for publication in JOSS!
Thank you to @adamltyson for the excellent editorial handling and guidance throughout the process, and to the reviewers @kevinyamauchi @jluethi for the thoughtful, constructive feedback. Your suggestions substantially improved both the manuscript and the software.
We also really appreciate the work from the JOSS team in moving this through to publication. Thanks again all!
