### 2026-01-13 5933bb1
Update paper: add citations and standardize affiliations and bibliography formatting
