+++ paper/paper.bib
@@ -91,7 +91,7 @@
     pages = {968--981.e15},
     pmid = {30078711},
     publisher = {Cell Press},
-    title = {{Deep Profiling of Mouse Splenic Architecture with CODEX Multiplexed Imaging}},
+    title = {Deep profiling of mouse splenic architecture with {CODEX} multiplexed imaging},
     volume = {174},
     year = {2018}
 }
@@ -182,7 +182,7 @@
     number = {97},
     pages = {6604},
     publisher = {The Open Journal},
-    title = {{SCIMAP: A Python Toolkit for Integrated Spatial Analysis of Multiplexed Imaging Data}},
+    title = {{SCIMAP}: a {Python} toolkit for integrated spatial analysis of multiplexed imaging data},
     volume = {9},
     year = {2024}
 }
@@ -205,7 +205,7 @@
 @misc{PalantirTechnologies,
     author = {{Palantir Technologies}},
     booktitle = {Palantir Technologies},
-    title = {{Palantir Foundry Documentation}},
+    title = {{Palantir Foundry} documentation},
     url = {https://palantir.com/docs/foundry/},
     urldate = {2025-04-15},
     year = {2003}
@@ -226,7 +226,7 @@
 
 @article{Virshup:2024,
   author  = {Virshup, Isaac and Rybakov, Sergey and Theis, Fabian J. and Angerer, Philipp and Wolf, Falko A.},
-  title   = {anndata: Access and store annotated data matrices},
+  title   = {{anndata}: access and store annotated data matrices},
   journal = {Journal of Open Source Software},
   year    = {2024},
   volume  = {9},
@@ -238,10 +238,69 @@
 
 @article{Wolf:2018,
   author  = {Wolf, Falko A. and Angerer, Philipp and Theis, Fabian J.},
-  title   = {SCANPY: Large-scale single-cell gene expression data analysis},
+  title   = {{SCANPY}: large-scale single-cell gene expression data analysis},
   journal = {Genome Biology},
   year    = {2018},
   volume  = {19},
   doi     = {10.1186/s13059-017-1382-0},
   url     = {https://doi.org/10.1186/s13059-017-1382-0}
+}
+
+@article{Harris:2020,
+    doi = {10.1038/s41586-020-2649-2},
+    url = {https://doi.org/10.1038/s41586-020-2649-2},
+    year = {2020},
+    month = sep,
+    publisher = {Springer Science and Business Media {LLC}},
+    volume = {585},
+    number = {7825},
+    pages = {357--362},
+    author = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+              van der Walt and Ralf Gommers and Pauli Virtanen and David
+              Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+              Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+              and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+              Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+              R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+              G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+              Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+              Travis E. Oliphant},
+    title = {Array programming with {NumPy}},
+    journal = {Nature}
+}
+
+@article{Waskom:2021,
+    doi = {10.21105/joss.03021},
+    url = {https://doi.org/10.21105/joss.03021},
+    year = {2021},
+    publisher = {The Open Journal},
+    volume = {6},
+    number = {60},
+    pages = {3021},
[…]
