### 2026-01-16 078d518
Refactor: rename galaxy_tools to tools, MVP_tools to galaxy_tools, add code_ocean_tools
