+++ tools/code_ocean_tools/code_ocean_synthesizer.py
@@ -1,0 +1,15 @@
+"""
+Code Ocean Synthesizer
+
+This module provides functionality for generating Code Ocean capsule configurations
+and tools from SPAC workflow templates.
+"""
+
+
+def main():
+    """Main entry point for Code Ocean synthesizer."""
+    pass
+
+
+if __name__ == "__main__":
+    main()
