### 2026-01-16 #8787 @rkurchin
Yup, looks good now, great!

Congrats to @fangliu117 and thanks to editor @adamltyson and reviewers @jluethi and @kevinyamauchi, JOSS couldn't run without your efforts!
